Logging object for objective function evaluations

Container around a data.table::data.table which stores all performed function calls of the Objective and the associated mlr3::BenchmarkResult.

$benchmark_result stores a mlr3::BenchmarkResult which contains the mlr3::ResampleResult of all performed function calls. The mlr3::BenchmarkResult is connected to the data.table::data.table via the uhash column.

Technical details

The data is stored in a private .data field that contains a data.table::data.table which logs all performed function calls of the ObjectiveFSelect. This data.table::data.table is accessed with the public $data() method. New values can be added with the $add_evals() method. This however is usually done through the evaluation of the FSelectInstanceSingleCrit or FSelectInstanceMultiCrit by the FSelector.

Super class

bbotk::Archive -> ArchiveFSelect

Public fields

benchmark_result

(mlr3::BenchmarkResult)
Stores benchmark result.

Methods

Public methods

ArchiveFSelect$clone()

Inherited methods

bbotk::Archive$add_evals()

bbotk::Archive$best()

bbotk::Archive$clear()

bbotk::Archive$data()

bbotk::Archive$format()

bbotk::Archive$initialize()

bbotk::Archive$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

ArchiveFSelect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.