AutoFeselect

The AutoFSelect is a mlr3::Learner which automatically selects the optimal feature set and fits the model on the training data with the optimal feature set.

Super class

mlr3::Learner -> AutoFSelect

Public fields

instance_args

list()

fselect

FSelect

store_fselect_instance

logical(1)

Active bindings

learner

mlr3::Learner

fselect_instance

FSelectInstance

fselect_result

list() Result of the feature selection i.e. the optimal feature set and its estimated performances.

Methods

Public methods

AutoFSelect$new()

AutoFSelect$train_internal()

AutoFSelect$predict_internal()

AutoFSelect$archive()

AutoFSelect$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Create new AutoFSelect object.

Usage

AutoFSelect$new(
  learner,
  resampling,
  measures,
  terminator,
  fselect,
  bm_args = list()
)

Arguments

learner

mlr3::Learner

resampling

mlr3::Resampling

measures

list of mlr3::Measure

terminator

Terminator

fselect

FSelect

bm_args

named list() Further arguments for mlr3::benchmark().

Returns

AutoFSelect

Method train_internal()

Selects the optimal feature set, applies the feature set to the task and fits the model on the training data. If store_fselect_instance is TRUE, the mlr3::Learner with the FSelectInstance is returned, otherwise just the mlr3::Learner. For internal use.

Usage

AutoFSelect$train_internal(task)

Arguments

task

mlr3::Task

Returns

list(mlr3::Learner) or list(mlr3::Learner, FSelectInstance)

Method predict_internal()

Creates a new mlr3::Prediction based on the learner fitted on the training data with the optimal feature subset. For internal use.

Usage

AutoFSelect$predict_internal(task)

Arguments

task

mlr3::Task

Returns

mlr3::Prediction

Method archive()

Returns a table of contained resample results with corresponding feature sets.

Usage

AutoFSelect$archive()

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

AutoFSelect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measures = msr("classif.ce")

terminator = term("evals", n_evals = 15)
fs = fs("exhaustive")
afs = AutoFSelect$new(learner, resampling, measures, terminator, fs)
afs$store_fselect_instance = TRUE

afs$train(task)
afs$model

#> $learner
#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $fselect_instance
#> <FSelectInstance>
#> * Task: <TaskClassif:iris>
#> * Learner: <LearnerClassifRpart:classif.rpart>
#> * Measures: classif.ce
#> * Resampling: <ResamplingHoldout>
#> * Terminator: <TerminatorEvals>
#> * bm_args: list()
#> * Result feat: Petal.Length, Sepal.Length
#> * Result perf: classif.ce=0.06
#> 

afs$learner

#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights