AutoFSelector

The AutoFSelector is a mlr3::Learner which wraps another mlr3::Learner and performs the following steps during $train():

The wrapped (inner) learner is trained on the feature subsets via resampling. The feature selection can be specified by providing a FSelector, a bbotk::Terminator, a mlr3::Resampling and a mlr3::Measure.

A final model is fit on the complete training data with the best found feature subset.

During $predict() the AutoFSelector just calls the predict method of the wrapped (inner) learner.

Note that this approach allows to perform nested resampling by passing an AutoFSelector object to mlr3::resample() or mlr3::benchmark(). To access the inner resampling results, set store_fselect_instance = TRUE and execute mlr3::resample() or mlr3::benchmark() with store_models = TRUE (see examples).

Super class

mlr3::Learner -> AutoFSelector

Public fields

instance_args

(list())
All arguments from construction to create the FSelectInstanceSingleCrit.

fselector

(FSelector)
Stores the feature selection algorithm.

Active bindings

archive

([ArchiveFSelect)
Returns FSelectInstanceSingleCrit archive.

learner

(mlr3::Learner)
Trained learner.

fselect_instance

(FSelectInstanceSingleCrit)
Internally created feature selection instance with all intermediate results.

fselect_result

(named list())
Short-cut to $result from FSelectInstanceSingleCrit.

Methods

Public methods

AutoFSelector$new()

AutoFSelector$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

AutoFSelector$new(
  learner,
  resampling,
  measure,
  terminator,
  fselector,
  store_fselect_instance = TRUE,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE
)

Arguments

learner

(mlr3::Learner)
Learner to optimize the feature subset for, see FSelectInstanceSingleCrit.

resampling

(mlr3::Resampling)
Resampling strategy during feature selection, see FSelectInstanceSingleCrit. This mlr3::Resampling is meant to be the inner resampling, operating on the training set of an arbitrary outer resampling. For this reason it is not feasible to pass an instantiated mlr3::Resampling here.

measure

(mlr3::Measure)
Performance measure to optimize.

terminator

(bbotk::Terminator)
When to stop feature selection, see FSelectInstanceSingleCrit.

fselector

(FSelector)
Feature selection algorithm to run.

store_fselect_instance

(logical(1))
If TRUE (default), stores the internally created FSelectInstanceSingleCrit with all intermediate results in slot $fselect_instance.

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1)). Store models in benchmark result?

check_values

(logical(1))
Check the parameters before the evaluation and the results for validity?

Method clone()

The objects of this class are cloneable with this method.

Usage

AutoFSelector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measure = msr("classif.ce")

terminator = trm("evals", n_evals = 15)
fselector = fs("exhaustive_search")
afs = AutoFSelector$new(learner, resampling, measure, terminator, fselector,
  store_fselect_instance = TRUE)

afs$train(task)
afs$model

#> $learner
#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $fselect_instance
#> <FSelectInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveFSelect:classif.rpart_on_iris>
#> * Search Space:
#> <ParamSet>
#>              id    class lower upper      levels        default value
#> 1: Petal.Length ParamLgl    NA    NA  TRUE,FALSE <NoDefault[3]>      
#> 2:  Petal.Width ParamLgl    NA    NA  TRUE,FALSE <NoDefault[3]>      
#> 3: Sepal.Length ParamLgl    NA    NA  TRUE,FALSE <NoDefault[3]>      
#> 4:  Sepal.Width ParamLgl    NA    NA  TRUE,FALSE <NoDefault[3]>      
#> * Terminator: <TerminatorEvals>
#> * Terminated: FALSE
#> * Result:
#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features  x_domain
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length <list[4]>
#>    classif.ce
#> 1:       0.06
#> * Archive:
#> <ArchiveFSelect>
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:         TRUE       FALSE        FALSE       FALSE       0.06
#>  2:        FALSE        TRUE        FALSE       FALSE       0.06
#>  3:        FALSE       FALSE         TRUE       FALSE       0.50
#>  4:        FALSE       FALSE        FALSE        TRUE       0.44
#>  5:         TRUE        TRUE        FALSE       FALSE       0.06
#>  6:         TRUE       FALSE         TRUE       FALSE       0.06
#>  7:         TRUE       FALSE        FALSE        TRUE       0.06
#>  8:        FALSE        TRUE         TRUE       FALSE       0.06
#>  9:        FALSE        TRUE        FALSE        TRUE       0.06
#> 10:        FALSE       FALSE         TRUE        TRUE       0.50
#> 11:         TRUE        TRUE         TRUE       FALSE       0.06
#> 12:         TRUE        TRUE        FALSE        TRUE       0.06
#> 13:         TRUE       FALSE         TRUE        TRUE       0.06
#> 14:        FALSE        TRUE         TRUE        TRUE       0.06
#> 15:         TRUE        TRUE         TRUE        TRUE       0.06
#>                                    uhash  x_domain           timestamp batch_nr
#>  1: abe3cf29-86e3-41bb-9ef2-c178f2f8b3f7 <list[4]> 2020-09-30 04:26:59        1
#>  2: e4d0adfa-0f45-4450-acb8-95e04c5f30ca <list[4]> 2020-09-30 04:26:59        1
#>  3: 4d156fb1-b525-4d87-a47d-4e791221b442 <list[4]> 2020-09-30 04:26:59        1
#>  4: ad97db74-257d-4ff8-894a-52e0c3cc895f <list[4]> 2020-09-30 04:26:59        1
#>  5: 3966b472-6701-44c0-af9b-d5101b2efd3a <list[4]> 2020-09-30 04:26:59        2
#>  6: 1d9bcfc1-6867-45e3-b186-70e4a7df5b05 <list[4]> 2020-09-30 04:26:59        2
#>  7: 1dbf5ef6-44fc-40ec-b3dd-34daeb710b55 <list[4]> 2020-09-30 04:26:59        2
#>  8: c649708c-bc11-4682-abad-9e5edefe0b9e <list[4]> 2020-09-30 04:26:59        2
#>  9: 518cbd74-51b8-4543-8c42-e665d1920b22 <list[4]> 2020-09-30 04:26:59        2
#> 10: f54c0675-c54a-4c0c-94b8-b67f9a28bf2c <list[4]> 2020-09-30 04:26:59        2
#> 11: 0c6f2963-0f17-477c-82c5-ff3fafe1df6f <list[4]> 2020-09-30 04:26:59        3
#> 12: 9dd55652-8909-4d69-8350-1710e06f269b <list[4]> 2020-09-30 04:26:59        3
#> 13: 8fd320ef-a9af-427c-98a8-d2555deee7e2 <list[4]> 2020-09-30 04:26:59        3
#> 14: c95b0c6f-cc7b-468d-8114-432910cbaa3e <list[4]> 2020-09-30 04:26:59        3
#> 15: de820bf8-5e29-4870-ada6-cb72ff1f9106 <list[4]> 2020-09-30 04:26:59        4
#> 

afs$learner

#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

#  Nested resampling
afs = AutoFSelector$new(learner, resampling, measure, terminator, fselector,
  store_fselect_instance = TRUE)

resampling_outer = rsmp("cv", folds = 2)
rr = resample(task, afs, resampling_outer, store_models = TRUE)

# Aggregate performance of outer results
rr$aggregate()

#> classif.ce 
#> 0.04666667 

# Retrieve inner feature selection results
as.data.table(rr)$learner[[1]]$fselect_result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features  x_domain
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length <list[4]>
#>    classif.ce
#> 1:          0