FSelect

Abstract FSelect class that implements the base functionality each FSelect* class must provide. A FSelect object describes the feature selection strategy, i.e. how to optimize the black-box function and its feasible set defined by the FSelectInstance object.

A list of measures can be passed to the instance, and they will always be all evaluated. However, single-criteria algorithms optimize only the first measure

A FSelect object must write its result to the $assign_result() method of the FSelectInstance at the end in order to store the best selected feature subset and its estimated performance vector.

Public fields

param_set

paradox::ParamSet

packages

character()

Methods

Public methods

FSelect$new()

FSelect$format()

FSelect$print()

FSelect$select()

FSelect$clone()

Method new()

Creates a new instance of this R6 class.

Usage

FSelect$new(param_set, packages = character())

Arguments

param_set

paradox::ParamSet

packages

character() Set of control parameter for the feature selection.

Method format()

Format method.

Usage

FSelect$format()

Returns

character()

Method print()

Print method.

Usage

FSelect$print()

Returns

character()

Method select()

Performs the feature selection on a FSelectInstance until termination.

Usage

FSelect$select(instance)

Arguments

instance

FSelectInstance

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = term("evals", n_evals = 15)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("random") # swap this line to use a different feature selection algorithm
fs$select(instance) # modifies the instance by reference
instance$result # returns best configuration and best performance

#> $feat
#> [1] "Petal.Width"  "Sepal.Length"
#> 
#> $perf
#> classif.ce 
#>       0.02 
#> 

instance$archive() # allows access of data.table / benchmark result of full path of all evaluation

#>     nr  resample_result task_id    learner_id resampling_id iters classif.ce
#>  1:  1 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  2:  2 <ResampleResult>    iris classif.rpart       holdout     1       0.30
#>  3:  3 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  4:  4 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#>  5:  5 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  6:  6 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#>  7:  7 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  8:  8 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#>  9:  9 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 10: 10 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 11: 11 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 12: 12 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 13: 13 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#> 14: 14 <ResampleResult>    iris classif.rpart       holdout     1       0.22
#> 15: 15 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#> 16: 16 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 17: 17 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#> 18: 18 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 19: 19 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#> 20: 20 <ResampleResult>    iris classif.rpart       holdout     1       0.06
#>     batch_nr                                              feat
#>  1:        1                          Petal.Width,Sepal.Length
#>  2:        1                                       Sepal.Width
#>  3:        1              Petal.Width,Sepal.Length,Sepal.Width
#>  4:        1                          Petal.Length,Sepal.Width
#>  5:        1                                       Petal.Width
#>  6:        1             Petal.Length,Petal.Width,Sepal.Length
#>  7:        1                          Petal.Width,Sepal.Length
#>  8:        1                                      Petal.Length
#>  9:        1                          Petal.Length,Petal.Width
#> 10:        1             Petal.Length,Petal.Width,Sepal.Length
#> 11:        2             Petal.Length,Sepal.Length,Sepal.Width
#> 12:        2 Petal.Length,Petal.Width,Sepal.Length,Sepal.Width
#> 13:        2                          Petal.Width,Sepal.Length
#> 14:        2                                      Sepal.Length
#> 15:        2                          Petal.Width,Sepal.Length
#> 16:        2                         Petal.Length,Sepal.Length
#> 17:        2              Petal.Width,Sepal.Length,Sepal.Width
#> 18:        2                          Petal.Length,Petal.Width
#> 19:        2              Petal.Length,Petal.Width,Sepal.Width
#> 20:        2 Petal.Length,Petal.Width,Sepal.Length,Sepal.Width