FSelectEvolutionary Class

Class for evolutionary feature selection calling ecr::ecr() from package ecr.

Parameters

mu

integer(1)

lambda

integer(1)

p.recomb

double(1)

p.mut

double(1)

survival.strategy

character(1)

n.elite

integer(1)

initial.solutions

list of integer()

parent.selector

object of class ecr::selector

survival.selector

object of class ecr::selector

For the meaning of the control parameter, see ecr::ecr(). Note that mu and lambda must be set by the user. The terminators parameter is replaced by the Terminator subclasses.

Super class

mlr3fselect::FSelect -> FSelectEvolutionary

Methods

Public methods

FSelectEvolutionary$new()

FSelectEvolutionary$clone()

Inherited methods

mlr3fselect::FSelect$format()

mlr3fselect::FSelect$print()

mlr3fselect::FSelect$select()

Method new()

Create new FSelectEvolutionary object.

Usage

FSelectEvolutionary$new()

Returns

FSelectEvolutionary

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectEvolutionary$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = term("evals", n_evals = 20)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("evolutionary", mu = 4, lambda = 8)
fs$select(instance) # modifies the instance by reference
instance$result # returns best configuration and best performance

#> $feat
#> [1] "Petal.Width"  "Sepal.Length"
#> 
#> $perf
#> classif.ce 
#>       0.04 
#> 

instance$archive() # allows access of data.table / benchmark result of full path of all evaluation

#>     nr  resample_result task_id    learner_id resampling_id iters classif.ce
#>  1:  1 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  2:  2 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  3:  3 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  4:  4 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  5:  5 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  6:  6 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  7:  7 <ResampleResult>    iris classif.rpart       holdout     1       0.04
#>  8:  8 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#>  9:  9 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 10: 10 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 11: 11 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 12: 12 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 13: 13 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 14: 14 <ResampleResult>    iris classif.rpart       holdout     1       0.04
#> 15: 15 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 16: 16 <ResampleResult>    iris classif.rpart       holdout     1       0.04
#> 17: 17 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 18: 18 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 19: 19 <ResampleResult>    iris classif.rpart       holdout     1       0.10
#> 20: 20 <ResampleResult>    iris classif.rpart       holdout     1       0.04
#>     batch_nr                                  feat
#>  1:        1              Petal.Length,Sepal.Width
#>  2:        2              Petal.Length,Sepal.Width
#>  3:        3 Petal.Length,Petal.Width,Sepal.Length
#>  4:        4                          Petal.Length
#>  5:        5              Petal.Length,Sepal.Width
#>  6:        6              Petal.Length,Sepal.Width
#>  7:        7              Petal.Width,Sepal.Length
#>  8:        8              Petal.Length,Sepal.Width
#>  9:        9 Petal.Length,Petal.Width,Sepal.Length
#> 10:       10              Petal.Length,Sepal.Width
#> 11:       11              Petal.Length,Sepal.Width
#> 12:       12 Petal.Length,Petal.Width,Sepal.Length
#> 13:       13              Petal.Length,Sepal.Width
#> 14:       14              Petal.Width,Sepal.Length
#> 15:       15 Petal.Length,Petal.Width,Sepal.Length
#> 16:       16              Petal.Width,Sepal.Length
#> 17:       17              Petal.Length,Sepal.Width
#> 18:       18              Petal.Length,Sepal.Width
#> 19:       19              Petal.Length,Sepal.Width
#> 20:       20              Petal.Width,Sepal.Length