FSelectInstance Class

Specifies a general feature selection scenario including performance evaluator and archive for FSelect objects to act upon. This class encodes the black box objective function that a FSelect object has to optimize. It allows the basis operations of querying the objective with feature sets ($eval_batch()), storing the evaluations in an internal archive and querying the archive ($archive()).

Evaluations of feature sets are performed in batches by calling mlr3::benchmark() internally. Before and after a batch is evaluated, the Terminator is queried for the remaining budget. If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on.

A list of measures can be passed to the instance, and they will always be all evaluated. However, single-criteria tuners optimize only the first measure.

The FSelect object is also supposed to store its final result, consisting of a selected feature set and associated estimated performance values, by calling the method instance$assign_result().

This class allows to display the optimization path of the feature selection in varying levels of detail ($optimization_path()).

Super class

bbotk::Instance -> FSelectInstance

Active bindings

n_batch

Number of batches.

result

Result of the feature selection i.e. the optimal feature set and its estimated performances.

Methods

Public methods

FSelectInstance$new()

FSelectInstance$print()

FSelectInstance$eval_batch()

FSelectInstance$fselect_objective()

FSelectInstance$archive()

FSelectInstance$optimization_path()

FSelectInstance$best()

FSelectInstance$assign_result()

FSelectInstance$clone()

Inherited methods

bbotk::Instance$format()

bbotk::Instance$objective()

Method new()

Create new FSelectInstance object.

Usage

FSelectInstance$new(
  task,
  learner,
  resampling,
  measures,
  terminator,
  bm_args = list()
)

Arguments

task

mlr3::Task

learner

mlr3::Learner

resampling

mlr3::Resampling Note that uninstantiated resamplings are instantiated during construction so that all configurations.

measures

list of mlr3::Measure

terminator

Terminator

bm_args

named list() Further arguments for mlr3::benchmark(). Stores all evaluated mlr3::ResampleResults when evaluating feature combinations.

bmr

mlr3::BenchmarkResult

Returns

FSelectInstance

Method print()

Print method.

Usage

FSelectInstance$print()

Returns

character()

Method eval_batch()

Evaluates all feature sets in states through resampling. Updates the internal BenchmarkResult $bmr by reference. Before and after each batch-evaluation, the Terminator is checked, and if it is positive, an exception of class terminated_error is raised. This function should be internally called by the FSelect object.

Usage

FSelectInstance$eval_batch(states)

Arguments

states

matrix Each row represents a 0/1 encoded feature set.

Returns

named list()

Method fselect_objective()

Evaluates a feature set x and returns a scalar objective value, where the return value is negated if the measure is maximized. Internally, $eval_batch() is called with a single row. This method is useful for feature selection algorithms that take a objective function.

Usage

FSelectInstance$fselect_objective(x)

Arguments

x

numeric 0/1 encoded feature set

Returns

numeric(1)

Method archive()

Returns a table of contained resample results with corresponding feature sets.

Usage

FSelectInstance$archive()

Returns

data.table::data.table

Method optimization_path()

Queries the mlr3::BenchmarkResult for the n best feature sets (default is 1) of the batches specified in m (default is the last batch).

Usage

FSelectInstance$optimization_path(n = 1, m = NULL)

Arguments

n

integer(1) Number of feature sets per batch.

m

integer Batch numbers.

Returns

data.table::data.table

Method best()

Queries the mlr3::BenchmarkResult for the best mlr3::ResampleResult according to measure (default is the first measure in $measures of the batches specified in m (default is the all batches). In case of ties, one of the tied values is selected randomly.

Usage

FSelectInstance$best(measure = NULL, m = NULL)

Arguments

measure

mlr3::Measure

m

ìnteger

Returns

mlr3::ResampleResult

Method assign_result()

The FSelect object writes the best found feature set and estimated performance values here. For internal use.

Usage

FSelectInstance$assign_result(feat, perf)

Arguments

feat

character Must be character vector of feature names existing in task

perf

numeric Must be named numeric of performance measures, named with performance IDs, regarding all elements in measures.

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectInstance$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

# Objects required to define the performance evaluator
task = tsk("iris")
measures = msrs(c("classif.ce"))
learner = lrn("classif.rpart")
resampling = rsmp("cv")
terminator = term("evals", n_evals = 8)

instance = FSelectInstance$new(
  task = task,
  learner = learner,
  resampling = resampling,
  measures = measures,
  terminator = terminator
)

# Evaluate 4 feature sets
m = diag(4)
instance$eval_batch(m)

#> $batch_nr
#> [1] 1
#> 
#> $uhashes
#> [1] "7f12c37a-e231-4c83-9a9c-22e11763db6e"
#> [2] "dc90554d-95d3-4558-a674-f3917d2d8222"
#> [3] "d271c904-3b82-41d9-be8b-1bb5bb949248"
#> [4] "75fdbda2-1912-46e9-b5f4-51163f652266"
#> 
#> $perf
#>    classif.ce
#> 1: 0.06000000
#> 2: 0.04666667
#> 3: 0.30000000
#> 4: 0.50000000
#> 

instance$archive()

#>    nr  resample_result task_id    learner_id resampling_id iters classif.ce
#> 1:  1 <ResampleResult>    iris classif.rpart            cv    10 0.06000000
#> 2:  2 <ResampleResult>    iris classif.rpart            cv    10 0.04666667
#> 3:  3 <ResampleResult>    iris classif.rpart            cv    10 0.30000000
#> 4:  4 <ResampleResult>    iris classif.rpart            cv    10 0.50000000
#>    batch_nr         feat
#> 1:        1 Petal.Length
#> 2:        1  Petal.Width
#> 3:        1 Sepal.Length
#> 4:        1  Sepal.Width