Multi-Criteria Feature Selection with Rush

The FSelectInstanceAsyncMultiCrit specifies a feature selection problem for a FSelectorAsync. The function fsi_async() creates a FSelectInstanceAsyncMultiCrit and the function fselect() creates an instance internally.

Default Measures

If no measure is passed, the default measure is used. The default measure depends on the task type.

Task	Default Measure	Package

"classif"	"classif.ce"	mlr3

"regr"	"regr.mse"	mlr3

"surv"	"surv.cindex"	mlr3proba

"dens"	"dens.logloss"	mlr3proba

"classif_st"	"classif.ce"	mlr3spatial

"regr_st"	"regr.mse"	mlr3spatial

"clust"	"clust.dunn"	mlr3cluster

Analysis

For analyzing the feature selection results, it is recommended to pass the ArchiveAsyncFSelect to as.data.table(). The returned data table contains the mlr3::ResampleResult for each feature subset evaluation.

Resources

There are several sections about feature selection in the mlr3book.

Getting started with wrapper feature selection.

Do a sequential forward selection Palmer Penguins data set.

The gallery features a collection of case studies and demos about optimization.

Utilize the built-in feature importance of models with Recursive Feature Elimination.

Run a feature selection with Shadow Variable Search.

Super classes

bbotk::OptimInstance -> bbotk::OptimInstanceAsync -> bbotk::OptimInstanceAsyncMultiCrit -> FSelectInstanceAsyncMultiCrit

Methods

Public methods

FSelectInstanceAsyncMultiCrit$new()

FSelectInstanceAsyncMultiCrit$assign_result()

FSelectInstanceAsyncMultiCrit$clone()

Inherited methods

bbotk::OptimInstance$format()

bbotk::OptimInstanceAsync$clear()

bbotk::OptimInstanceAsync$print()

bbotk::OptimInstanceAsync$reconnect()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectInstanceAsyncMultiCrit$new(
  task,
  learner,
  resampling,
  measures,
  terminator,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE,
  callbacks = NULL,
  rush = NULL
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner)
Learner to optimize the feature subset for.

resampling

(mlr3::Resampling)
Resampling that is used to evaluated the performance of the feature subsets. Uninstantiated resamplings are instantiated during construction so that all feature subsets are evaluated on the same data splits. Already instantiated resamplings are kept unchanged.

measures

(list of mlr3::Measure)
Measures to optimize. If NULL, mlr3's default measure is used.

terminator

(bbotk::Terminator)
Stop criterion of the feature selection.

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1)). Store models in benchmark result?

check_values

(logical(1))
Check the parameters before the evaluation and the results for validity?

callbacks

(list of CallbackBatchFSelect)
List of callbacks.

rush

(Rush)
If a rush instance is supplied, the optimization runs without batches.

Method assign_result()

The FSelectorAsync object writes the best found points and estimated performance values here (probably the Pareto set / front). For internal use.

Usage

FSelectInstanceAsyncMultiCrit$assign_result(xdt, ydt, extra = NULL, ...)

Arguments

xdt

(data.table::data.table())
x values as data.table. Each row is one point. Contains the value in the search space of the FSelectInstanceBatchMultiCrit object. Can contain additional columns for extra information.

ydt

(numeric())
Optimal outcomes, e.g. the Pareto front.

extra

(data.table::data.table())
Additional information.

...

(any)
ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectInstanceAsyncMultiCrit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.