FSelectRFE Class

Class for feature selection by recursive feature elimination (RFE). The recursive algorithm (recursive = TRUE) recomputes the feature importances on the reduced feature set in every iteration. The non-recursive algorithm (recursive = TRUE) only uses the feature importances of the model fitted with all features to eliminate the next most unimportant feature in every iteration.

Parameters

min_features

integer(1) Minimum number of features. By default, 1.

recursive

logical(1)

Super class

mlr3fselect::FSelect -> FSelectRFE

Public fields

importance

Stores the feature importance of the model with all variables if recrusive is set to FALSE

Methods

Public methods

FSelectRFE$new()

FSelectRFE$clone()

Inherited methods

mlr3fselect::FSelect$format()

mlr3fselect::FSelect$print()

mlr3fselect::FSelect$select()

Method new()

Create new FSelectRFE object.

Usage

FSelectRFE$new()

Returns

FSelectRFE

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectRFE$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.