FSelectRandom Class

Subclass for random feature selection. Feature sets are randomly drawn.

Dictionary

This FSelect can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random")
fs("random")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

prob

double(1)
Probability of choosing a feature.

In order to support general termination criteria and parallelization, feature sets are evaluated in a batch-fashion of size batch_size. Larger batches mean more is parallelized, smaller batches imply a more fine-grained checking of termination criteria.

Super classes

bbotk::Optimizer -> mlr3fselect::FSelect -> FSelectRandom

Methods

Public methods

FSelectRandom$new()

FSelectRandom$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$print()

mlr3fselect::FSelect$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectRandom$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectRandom$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = term("evals", n_evals = 10)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("random")
fs$optimize(instance)
instance$result

#> $feat
#> [1] "Petal.Width"
#> 
#> $perf
#> classif.ce 
#>       0.02 
#> 

instance$archive$data

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:         TRUE       FALSE         TRUE        TRUE       0.04
#>  2:         TRUE        TRUE        FALSE       FALSE       0.04
#>  3:        FALSE       FALSE         TRUE       FALSE       0.26
#>  4:        FALSE        TRUE        FALSE       FALSE       0.02
#>  5:        FALSE        TRUE        FALSE       FALSE       0.02
#>  6:        FALSE       FALSE         TRUE       FALSE       0.26
#>  7:         TRUE        TRUE        FALSE        TRUE       0.04
#>  8:         TRUE       FALSE         TRUE       FALSE       0.04
#>  9:         TRUE       FALSE        FALSE       FALSE       0.04
#> 10:        FALSE        TRUE        FALSE        TRUE       0.02
#>      resample_result  opt_x  timestamp batch_nr
#>  1: <ResampleResult> <list> 1586579409        1
#>  2: <ResampleResult> <list> 1586579409        1
#>  3: <ResampleResult> <list> 1586579409        1
#>  4: <ResampleResult> <list> 1586579409        1
#>  5: <ResampleResult> <list> 1586579409        1
#>  6: <ResampleResult> <list> 1586579409        1
#>  7: <ResampleResult> <list> 1586579409        1
#>  8: <ResampleResult> <list> 1586579409        1
#>  9: <ResampleResult> <list> 1586579409        1
#> 10: <ResampleResult> <list> 1586579409        1