FSelectSequential Class

Class for sequential feature selection. The sequential forward selection (strategy = fsf) extends the feature set in each step with the feature that increases the models performance the most. The sequential backward selection (strategy = fsb) starts with the complete future set and removes in each step the feature that decreases the models performance the least.

Parameters

max_features

integer(1) Maximum number of features. By default, number of features in mlr3::Task.

strategy

character(1) For forward feature selection fsf, for backward feature selection fsb.

The feature sets are evaluated in batches. Each batch is one step in the sequential feature selection.

Super class

mlr3fselect::FSelect -> FSelectSequential

Methods

Public methods

FSelectSequential$new()

FSelectSequential$clone()

Inherited methods

mlr3fselect::FSelect$format()

mlr3fselect::FSelect$print()

mlr3fselect::FSelect$select()

Method new()

Create new FSelectSequential object.

Usage

FSelectSequential$new()

Returns

FSelectSequential

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectSequential$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = term("evals", n_evals = 15)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("sequential")
fs$select(instance) # modifies the instance by reference
instance$result # returns best configuration and best performance

#> $feat
#> [1] "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> $perf
#> classif.ce 
#>       0.02 
#> 

instance$archive() # allows access of data.table / benchmark result of full path of all evaluation

#>     nr  resample_result task_id    learner_id resampling_id iters classif.ce
#>  1:  1 <ResampleResult>    iris classif.rpart       holdout     1       0.12
#>  2:  2 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  3:  3 <ResampleResult>    iris classif.rpart       holdout     1       0.28
#>  4:  4 <ResampleResult>    iris classif.rpart       holdout     1       0.44
#>  5:  5 <ResampleResult>    iris classif.rpart       holdout     1       0.12
#>  6:  6 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  7:  7 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#>  8:  8 <ResampleResult>    iris classif.rpart       holdout     1       0.12
#>  9:  9 <ResampleResult>    iris classif.rpart       holdout     1       0.02
#> 10: 10 <ResampleResult>    iris classif.rpart       holdout     1       0.12
#>     batch_nr                                              feat
#>  1:        1                                      Petal.Length
#>  2:        1                                       Petal.Width
#>  3:        1                                      Sepal.Length
#>  4:        1                                       Sepal.Width
#>  5:        2                          Petal.Length,Petal.Width
#>  6:        2                          Petal.Width,Sepal.Length
#>  7:        2                           Petal.Width,Sepal.Width
#>  8:        3             Petal.Length,Petal.Width,Sepal.Length
#>  9:        3              Petal.Width,Sepal.Length,Sepal.Width
#> 10:        4 Petal.Length,Petal.Width,Sepal.Length,Sepal.Width

instance$optimization_path(n = 1, m = 1:4) # returns selected feature of each step

#>    batch_nr Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#> 1:        1            0           1            0           0       0.02
#> 2:        2            0           1            1           0       0.02
#> 3:        3            0           1            1           1       0.02
#> 4:        4            1           1            1           1       0.12