Feature Selection via Genetic Search

FSelectorGeneticSearch class that implements an Genetic Search. Calls genalg::rbga.bin() from package genalg.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("genetic_search")
fs("genetic_search")

Parameters

suggestions

list()

popSize

integer(1)

mutationChance

numeric(1)

elitism

integer(1)

zeroToOneRatio

integer(1)

iters

integer(1)

For the meaning of the control parameters, see genalg::rbga.bin(). genalg::rbga.bin() internally terminates after iters iteration. We set ìters = 100000 to allow the termination via our terminators. If more iterations are needed, set ìters to a higher value in the parameter set.

Super class

mlr3fselect::FSelector -> FSelectorGeneticSearch

Methods

Public methods

FSelectorGeneticSearch$new()

FSelectorGeneticSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorGeneticSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorGeneticSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = trm("evals", n_evals = 5)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

fselector = fs("genetic_search", popSize = 10L)
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features  x_domain
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length <list[4]>
#>    classif.ce
#> 1:       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features  x_domain
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length <list[4]>
#>    classif.ce
#> 1:       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE       0.06
#> 2:        FALSE       FALSE        FALSE        TRUE       0.48
#> 3:        FALSE       FALSE        FALSE        TRUE       0.48
#> 4:        FALSE       FALSE         TRUE       FALSE       0.26
#> 5:        FALSE       FALSE         TRUE       FALSE       0.26
#>                                   uhash           timestamp batch_nr
#> 1: 13a65d48-fd7f-43dd-866c-e16af4b407d6 2020-12-16 16:21:37        1
#> 2: ccbaa48d-d8b9-457a-b0c5-fd31db4986ea 2020-12-16 16:21:38        2
#> 3: b797bf81-b634-4621-95e9-1eecdc64c837 2020-12-16 16:21:38        3
#> 4: e69bf83d-b2af-4e2e-9b34-7c9cf0304f34 2020-12-16 16:21:38        4
#> 5: 5fa62b97-ba74-4dd1-b693-0b3369d824ed 2020-12-16 16:21:39        5
#>    x_domain_Petal.Length x_domain_Petal.Width x_domain_Sepal.Length
#> 1:                  TRUE                FALSE                 FALSE
#> 2:                 FALSE                FALSE                 FALSE
#> 3:                 FALSE                FALSE                 FALSE
#> 4:                 FALSE                FALSE                  TRUE
#> 5:                 FALSE                FALSE                  TRUE
#>    x_domain_Sepal.Width
#> 1:                FALSE
#> 2:                 TRUE
#> 3:                 TRUE
#> 4:                FALSE
#> 5:                FALSE