Feature Selection via Recursive Feature Elimination

FSelectorRFE class that implements Recursive Feature Elimination (RFE). The recursive algorithm (recursive = TRUE) recomputes the feature importance on the reduced feature set in every iteration. The non-recursive algorithm (recursive = FALSE) only uses the feature importance of the model fitted with all features to eliminate the next most unimportant features in every iteration.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("rfe")
fs("rfe")

Parameters

min_features

integer(1)
The minimum number of features to select, default is 1L.

feature_fraction

double(1)
Fraction of features to retain in each iteration, default is 0.5.

feature_number

integer(1)
Number of features to remove in each iteration.

subset_sizes

integer()
Vector of number of features to retain in each iteration. Must be sorted in decreasing order.

recursive

logical(1)
Use the recursive version? Default is FALSE.

The parameter feature_fraction, feature_number and subset_sizes are mutually exclusive.

Super class

mlr3fselect::FSelector -> FSelectorRFE

Public fields

importance

numeric()
Stores the feature importance of the model with all variables if recrusive is set to FALSE

Methods

Public methods

FSelectorRFE$new()

FSelectorRFE$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRFE$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRFE$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = trm("evals", n_evals = 10)
instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator,
  store_models = TRUE
)

fselector = fs("rfe")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width
#> 1:         TRUE        TRUE         TRUE        TRUE
#>                                             features classif.ce
#> 1: Petal.Length,Petal.Width,Sepal.Length,Sepal.Width        0.1

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width
#> 1:         TRUE        TRUE         TRUE        TRUE
#>                                             features classif.ce
#> 1: Petal.Length,Petal.Width,Sepal.Length,Sepal.Width        0.1

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#> 1:         TRUE        TRUE         TRUE        TRUE        0.1
#> 2:         TRUE        TRUE        FALSE       FALSE        0.1
#> 3:        FALSE        TRUE        FALSE       FALSE        0.1
#>    runtime_learners           timestamp batch_nr
#> 1:            0.053 2021-07-12 04:47:21        1
#> 2:            0.050 2021-07-12 04:47:21        2
#> 3:            0.057 2021-07-12 04:47:21        3
#>                             importance      resample_result
#> 1: 64.07556,61.31460,37.69358,20.01587 <ResampleResult[21]>
#> 2:                   64.07556,61.31460 <ResampleResult[21]>
#> 3:                            64.07556 <ResampleResult[21]>