Feature Selection via Sequential Search with Shadow Variables

FSelectorShadowVariableSearch class that implements a sequential search with shadow variables.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi: 10.1155/2017/1421409 . Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi: 10.1198/016214506000000843 .

Note

FSelectorShadowVariableSearch terminates itself and should be used with TerminatorNone.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("none")
)

fselector = fs("shadow_variable_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>  1:         TRUE       FALSE        FALSE       FALSE                  FALSE
#>  2:        FALSE        TRUE        FALSE       FALSE                  FALSE
#>  3:        FALSE       FALSE         TRUE       FALSE                  FALSE
#>  4:        FALSE       FALSE        FALSE        TRUE                  FALSE
#>  5:        FALSE       FALSE        FALSE       FALSE                   TRUE
#>  6:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  7:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  8:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  9:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 10:         TRUE       FALSE         TRUE       FALSE                  FALSE
#> 11:         TRUE       FALSE        FALSE        TRUE                  FALSE
#> 12:         TRUE       FALSE        FALSE       FALSE                   TRUE
#> 13:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 14:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 15:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 16:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 17:         TRUE        TRUE        FALSE        TRUE                  FALSE
#> 18:         TRUE        TRUE        FALSE       FALSE                   TRUE
#> 19:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 20:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 21:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 22:         TRUE        TRUE         TRUE        TRUE                  FALSE
#> 23:         TRUE        TRUE         TRUE       FALSE                   TRUE
#> 24:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 25:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 26:         TRUE        TRUE         TRUE       FALSE                  FALSE
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>  1:                 FALSE                  FALSE                 FALSE
#>  2:                 FALSE                  FALSE                 FALSE
#>  3:                 FALSE                  FALSE                 FALSE
#>  4:                 FALSE                  FALSE                 FALSE
#>  5:                 FALSE                  FALSE                 FALSE
#>  6:                  TRUE                  FALSE                 FALSE
#>  7:                 FALSE                   TRUE                 FALSE
#>  8:                 FALSE                  FALSE                  TRUE
#>  9:                 FALSE                  FALSE                 FALSE
#> 10:                 FALSE                  FALSE                 FALSE
#> 11:                 FALSE                  FALSE                 FALSE
#> 12:                 FALSE                  FALSE                 FALSE
#> 13:                  TRUE                  FALSE                 FALSE
#> 14:                 FALSE                   TRUE                 FALSE
#> 15:                 FALSE                  FALSE                  TRUE
#> 16:                 FALSE                  FALSE                 FALSE
#> 17:                 FALSE                  FALSE                 FALSE
#> 18:                 FALSE                  FALSE                 FALSE
#> 19:                  TRUE                  FALSE                 FALSE
#> 20:                 FALSE                   TRUE                 FALSE
#> 21:                 FALSE                  FALSE                  TRUE
#> 22:                 FALSE                  FALSE                 FALSE
#> 23:                 FALSE                  FALSE                 FALSE
#> 24:                  TRUE                  FALSE                 FALSE
#> 25:                 FALSE                   TRUE                 FALSE
#> 26:                 FALSE                  FALSE                  TRUE
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>     classif.ce                                uhash           timestamp
#>  1:       0.06 8ea95558-e940-4cd4-85cf-11884bee6f5e 2021-06-08 05:05:16
#>  2:       0.06 97dbde1a-8211-4575-9c4f-a03d115a2acc 2021-06-08 05:05:16
#>  3:       0.34 fb801fb4-5391-4cba-b1f4-857959f1b9af 2021-06-08 05:05:16
#>  4:       0.50 63734b38-9b40-4713-96df-7ffcf1c49766 2021-06-08 05:05:16
#>  5:       0.70 268fa2af-ec3c-450e-ab30-d0f8bb41e226 2021-06-08 05:05:16
#>  6:       0.62 d64411d9-b427-442c-bc7b-c710bdba039d 2021-06-08 05:05:16
#>  7:       0.76 f480e058-590c-4708-80f2-1aa2a483b75f 2021-06-08 05:05:16
#>  8:       0.64 dd64adf2-f1a0-44d7-bd0f-e59fe90bb4d3 2021-06-08 05:05:16
#>  9:       0.06 3442af94-c5f1-404e-9386-4e81473d740b 2021-06-08 05:05:17
#> 10:       0.06 80b400a1-bd1e-4991-942a-986ec59c3029 2021-06-08 05:05:17
#> 11:       0.06 774bc346-158b-4a6d-b10f-213c2352821c 2021-06-08 05:05:17
#> 12:       0.06 de709d76-c97e-455f-a7ad-14e6ec77107d 2021-06-08 05:05:17
#> 13:       0.06 d2bc126b-741e-415b-896d-bcd544b98acd 2021-06-08 05:05:17
#> 14:       0.06 a2fae3e1-fffe-4bb9-8452-67333b0907b7 2021-06-08 05:05:17
#> 15:       0.06 89d10574-32d3-4d08-99ac-cf236fbc3283 2021-06-08 05:05:17
#> 16:       0.06 3be5cd7a-99e4-4eba-a881-2435992c3804 2021-06-08 05:05:18
#> 17:       0.06 104d5e3d-d095-4d6e-8f1a-7d3e2c8df09b 2021-06-08 05:05:18
#> 18:       0.06 51673b63-3b44-4cba-9175-9448fc1f0de7 2021-06-08 05:05:18
#> 19:       0.06 41798817-c573-4997-bbc1-7ad1fb2ac383 2021-06-08 05:05:18
#> 20:       0.06 ff6a9be1-0249-48e9-a397-2215e6abc522 2021-06-08 05:05:18
#> 21:       0.06 0ce3c43f-7efb-4ab3-b40c-ef67d9a846ea 2021-06-08 05:05:18
#> 22:       0.06 74103ce7-f563-4952-9d6e-445afcc1a920 2021-06-08 05:05:20
#> 23:       0.06 64e7dcc0-c488-48fa-a469-14e1e471b579 2021-06-08 05:05:20
#> 24:       0.06 2f0f5d56-476b-4614-b51c-2532401f2ddb 2021-06-08 05:05:20
#> 25:       0.06 febf1112-4e7e-4884-975d-ddc69b177b3a 2021-06-08 05:05:20
#> 26:       0.06 7463228f-7604-44b2-821e-6798bcacb0e7 2021-06-08 05:05:20
#>     classif.ce                                uhash           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        2
#> 10:        2
#> 11:        2
#> 12:        2
#> 13:        2
#> 14:        2
#> 15:        2
#> 16:        3
#> 17:        3
#> 18:        3
#> 19:        3
#> 20:        3
#> 21:        3
#> 22:        4
#> 23:        4
#> 24:        4
#> 25:        4
#> 26:        4
#>     batch_nr