Function for Automatic Feature Selection

The AutoFSelector wraps a mlr3::Learner and augments it with an automatic feature selection. The auto_fselector() function creates an AutoFSelector object.

Usage

auto_fselector(
  fselector,
  learner,
  resampling,
  measure = NULL,
  term_evals = NULL,
  term_time = NULL,
  terminator = NULL,
  store_fselect_instance = TRUE,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE,
  callbacks = NULL,
  ties_method = "least_features"
)

Arguments

fselector

(FSelector)
Optimization algorithm.

learner

(mlr3::Learner)
Learner to optimize the feature subset for.

resampling

(mlr3::Resampling)
Resampling that is used to evaluated the performance of the feature subsets. Uninstantiated resamplings are instantiated during construction so that all feature subsets are evaluated on the same data splits. Already instantiated resamplings are kept unchanged.

measure

(mlr3::Measure)
Measure to optimize. If NULL, default measure is used.

term_evals

(integer(1))
Number of allowed evaluations. Ignored if terminator is passed.

term_time

(integer(1))
Maximum allowed time in seconds. Ignored if terminator is passed.

terminator

(bbotk::Terminator)
Stop criterion of the feature selection.

store_fselect_instance

(logical(1))
If TRUE (default), stores the internally created FSelectInstanceBatchSingleCrit with all intermediate results in slot $fselect_instance. Is set to TRUE, if store_models = TRUE

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1)). Store models in benchmark result?

check_values

(logical(1))
Check the parameters before the evaluation and the results for validity?

callbacks

(list of CallbackBatchFSelect)
List of callbacks.

ties_method

(character(1))
The method to break ties when selecting sets while optimizing and when selecting the best set. Can be "least_features" or "random". The option "least_features" (default) selects the feature set with the least features. If there are multiple best feature sets with the same number of features, one is selected randomly. The random method returns a random feature set from the best feature sets. Ignored if multiple measures are used.

Value

AutoFSelector.

Details

The AutoFSelector is a mlr3::Learner which wraps another mlr3::Learner and performs the following steps during $train():

The wrapped (inner) learner is trained on the feature subsets via resampling. The feature selection can be specified by providing a FSelector, a bbotk::Terminator, a mlr3::Resampling and a mlr3::Measure.

A final model is fit on the complete training data with the best-found feature subset.

During $predict() the AutoFSelector just calls the predict method of the wrapped (inner) learner.

Resources

There are several sections about feature selection in the mlr3book.

Estimate Model Performance with nested resampling.

The gallery features a collection of case studies and demos about optimization.

Nested Resampling

Nested resampling can be performed by passing an AutoFSelector object to mlr3::resample() or mlr3::benchmark(). To access the inner resampling results, set store_fselect_instance = TRUE and execute mlr3::resample() or mlr3::benchmark() with store_models = TRUE (see examples). The mlr3::Resampling passed to the AutoFSelector is meant to be the inner resampling, operating on the training set of an arbitrary outer resampling. For this reason it is not feasible to pass an instantiated mlr3::Resampling here.

Examples

# Automatic Feature Selection
# \donttest{

# split to train and external set
task = tsk("penguins")
split = partition(task, ratio = 0.8)

# create auto fselector
afs = auto_fselector(
  fselector = fs("random_search"),
  learner = lrn("classif.rpart"),
  resampling = rsmp ("holdout"),
  measure = msr("classif.ce"),
  term_evals = 4)

# optimize feature subset and fit final model
afs$train(task, row_ids = split$train)

# predict with final model
afs$predict(task, row_ids = split$test)
#> <PredictionClassif> for 69 observations:
#>     row_ids     truth  response
#>           5    Adelie    Adelie
#>          11    Adelie    Adelie
#>          12    Adelie    Adelie
#> ---                            
#>         338 Chinstrap Chinstrap
#>         339 Chinstrap Chinstrap
#>         340 Chinstrap Chinstrap

# show result
afs$fselect_result
#>    bill_depth bill_length body_mass flipper_length island    sex   year
#>        <lgcl>      <lgcl>    <lgcl>         <lgcl> <lgcl> <lgcl> <lgcl>
#> 1:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE  FALSE
#>                                               features n_features classif.ce
#>                                                 <list>      <int>      <num>
#> 1: bill_depth,bill_length,body_mass,flipper_length,sex          5 0.06521739

# model slot contains trained learner and fselect instance
afs$model
#> $learner
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: mlr3, rpart
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $features
#> [1] "bill_depth"     "bill_length"    "body_mass"      "flipper_length"
#> [5] "sex"           
#> 
#> $fselect_instance
#> <FSelectInstanceBatchSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveFSelectBatch:classif.rpart_on_penguins>
#> * Terminator: <TerminatorEvals>
#> * Result:
#>    bill_depth bill_length body_mass flipper_length island    sex   year
#>        <lgcl>      <lgcl>    <lgcl>         <lgcl> <lgcl> <lgcl> <lgcl>
#> 1:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE  FALSE
#>    classif.ce
#>         <num>
#> 1: 0.06521739
#> * Archive:
#>     bill_depth bill_length body_mass flipper_length island    sex   year
#>         <lgcl>      <lgcl>    <lgcl>         <lgcl> <lgcl> <lgcl> <lgcl>
#>  1:       TRUE       FALSE     FALSE           TRUE   TRUE   TRUE   TRUE
#>  2:       TRUE        TRUE      TRUE          FALSE   TRUE   TRUE   TRUE
#>  3:      FALSE        TRUE      TRUE          FALSE  FALSE  FALSE  FALSE
#>  4:       TRUE        TRUE      TRUE          FALSE   TRUE   TRUE   TRUE
#>  5:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE   TRUE
#>  6:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE  FALSE
#>  7:      FALSE       FALSE     FALSE          FALSE   TRUE  FALSE  FALSE
#>  8:       TRUE        TRUE      TRUE           TRUE   TRUE   TRUE   TRUE
#>  9:      FALSE       FALSE     FALSE           TRUE   TRUE  FALSE   TRUE
#> 10:       TRUE        TRUE      TRUE          FALSE   TRUE   TRUE   TRUE
#>     classif.ce
#>          <num>
#>  1: 0.16304348
#>  2: 0.07608696
#>  3: 0.13043478
#>  4: 0.07608696
#>  5: 0.06521739
#>  6: 0.06521739
#>  7: 0.31521739
#>  8: 0.06521739
#>  9: 0.16304348
#> 10: 0.07608696
#> 

# shortcut trained learner
afs$learner
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: mlr3, rpart
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

# shortcut fselect instance
afs$fselect_instance
#> <FSelectInstanceBatchSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveFSelectBatch:classif.rpart_on_penguins>
#> * Terminator: <TerminatorEvals>
#> * Result:
#>    bill_depth bill_length body_mass flipper_length island    sex   year
#>        <lgcl>      <lgcl>    <lgcl>         <lgcl> <lgcl> <lgcl> <lgcl>
#> 1:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE  FALSE
#>    classif.ce
#>         <num>
#> 1: 0.06521739
#> * Archive:
#>     bill_depth bill_length body_mass flipper_length island    sex   year
#>         <lgcl>      <lgcl>    <lgcl>         <lgcl> <lgcl> <lgcl> <lgcl>
#>  1:       TRUE       FALSE     FALSE           TRUE   TRUE   TRUE   TRUE
#>  2:       TRUE        TRUE      TRUE          FALSE   TRUE   TRUE   TRUE
#>  3:      FALSE        TRUE      TRUE          FALSE  FALSE  FALSE  FALSE
#>  4:       TRUE        TRUE      TRUE          FALSE   TRUE   TRUE   TRUE
#>  5:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE   TRUE
#>  6:       TRUE        TRUE      TRUE           TRUE  FALSE   TRUE  FALSE
#>  7:      FALSE       FALSE     FALSE          FALSE   TRUE  FALSE  FALSE
#>  8:       TRUE        TRUE      TRUE           TRUE   TRUE   TRUE   TRUE
#>  9:      FALSE       FALSE     FALSE           TRUE   TRUE  FALSE   TRUE
#> 10:       TRUE        TRUE      TRUE          FALSE   TRUE   TRUE   TRUE
#>     classif.ce
#>          <num>
#>  1: 0.16304348
#>  2: 0.07608696
#>  3: 0.13043478
#>  4: 0.07608696
#>  5: 0.06521739
#>  6: 0.06521739
#>  7: 0.31521739
#>  8: 0.06521739
#>  9: 0.16304348
#> 10: 0.07608696

# Nested Resampling

afs = auto_fselector(
  fselector = fs("random_search"),
  learner = lrn("classif.rpart"),
  resampling = rsmp ("holdout"),
  measure = msr("classif.ce"),
  term_evals = 4)

resampling_outer = rsmp("cv", folds = 3)
rr = resample(task, afs, resampling_outer, store_models = TRUE)

# retrieve inner feature selection results.
extract_inner_fselect_results(rr)
#>    iteration bill_depth bill_length body_mass flipper_length island    sex
#>        <int>     <lgcl>      <lgcl>    <lgcl>         <lgcl> <lgcl> <lgcl>
#> 1:         1       TRUE        TRUE      TRUE           TRUE   TRUE  FALSE
#> 2:         2       TRUE        TRUE      TRUE          FALSE   TRUE  FALSE
#> 3:         3       TRUE        TRUE      TRUE           TRUE   TRUE   TRUE
#>      year classif.ce
#>    <lgcl>      <num>
#> 1:  FALSE 0.06578947
#> 2:   TRUE 0.03947368
#> 3:   TRUE 0.06493506
#>                                                          features n_features
#>                                                            <list>      <int>
#> 1:         bill_depth,bill_length,body_mass,flipper_length,island          5
#> 2:                   bill_depth,bill_length,body_mass,island,year          5
#> 3: bill_depth,bill_length,body_mass,flipper_length,island,sex,...          7
#>     task_id              learner_id resampling_id
#>      <char>                  <char>        <char>
#> 1: penguins classif.rpart.fselector            cv
#> 2: penguins classif.rpart.fselector            cv
#> 3: penguins classif.rpart.fselector            cv

# performance scores estimated on the outer resampling
rr$score()
#>     task_id              learner_id resampling_id iteration classif.ce
#>      <char>                  <char>        <char>     <int>      <num>
#> 1: penguins classif.rpart.fselector            cv         1 0.05217391
#> 2: penguins classif.rpart.fselector            cv         2 0.05217391
#> 3: penguins classif.rpart.fselector            cv         3 0.07017544
#> Hidden columns: task, learner, resampling, prediction

# unbiased performance of the final model trained on the full data set
rr$aggregate()
#> classif.ce 
#> 0.05817442 
# }