Syntactic Sugar for FSelect Construction

This function complements mlr_fselectors with functions in the spirit of mlr3::mlr_sugar.

fs(.key, ...)

Arguments

.key
(character(1))
Key passed to the respective dictionary to retrieve the object.

...
(named list())
Named arguments passed to the constructor, to be set as parameters in the paradox::ParamSet, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.

Value

FSelector.

Examples

fs("sequential", max_features = 4)

#> <FSelectorSequential>
#> * Parameters: strategy=sfs, max_features=4
#> * Parameter classes: ParamLgl
#> * Properties: single-crit
#> * Packages: -