Dictionary of FSelectors

A mlr3misc::Dictionary storing objects of class FSelector. Each fselector has an associated help page, see mlr_fselectors_[id].

For a more convenient way to retrieve and construct fselectors, see fs()/fss().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "properties" and "packages" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

See also

Examples

as.data.table(mlr_fselectors)
#>                       key                         label
#> 1:          design_points                 Design Points
#> 2:      exhaustive_search             Exhaustive Search
#> 3:         genetic_search                Genetic Search
#> 4:          random_search                 Random Search
#> 5:                    rfe Recursive Feature Elimination
#> 6:             sequential             Sequential Search
#> 7: shadow_variable_search        Shadow Variable Search
#>                             properties          packages
#> 1: dependencies,single-crit,multi-crit mlr3fselect,bbotk
#> 2:              single-crit,multi-crit       mlr3fselect
#> 3:                         single-crit       mlr3fselect
#> 4:              single-crit,multi-crit       mlr3fselect
#> 5:                         single-crit       mlr3fselect
#> 6:                         single-crit       mlr3fselect
#> 7:                         single-crit       mlr3fselect
mlr_fselectors$get("random_search")
#> <FSelectorRandomSearch>: Random Search
#> * Parameters: batch_size=1
#> * Properties: single-crit, multi-crit
#> * Packages: mlr3fselect
fs("random_search")
#> <FSelectorRandomSearch>: Random Search
#> * Parameters: batch_size=1
#> * Properties: single-crit, multi-crit
#> * Packages: mlr3fselect