Feature Selection via Genetic Search

Genetic search imitates the process of natural selection to generate feature sets.

Calls genalg::rbga.bin() from package genalg.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("genetic_search")
fs("genetic_search")

Parameters

suggestions

list()

popSize

integer(1)

mutationChance

numeric(1)

elitism

integer(1)

zeroToOneRatio

integer(1)

iters

integer(1)

For the meaning of the control parameters, see genalg::rbga.bin(). genalg::rbga.bin() internally terminates after iters iteration. We set ìters = 100000 to allow the termination via our terminators. If more iterations are needed, set ìters to a higher value in the parameter set.

Super class

mlr3fselect::FSelector -> FSelectorGeneticSearch

Methods

Public methods

FSelectorGeneticSearch$new()

FSelectorGeneticSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorGeneticSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorGeneticSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "genetic_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10
)

# best performing feature subset
instance$result
#>      age glucose insulin  mass pedigree pregnant pressure triceps features
#> 1: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  glucose
#>    classif.ce
#> 1:  0.2460938

# all evaluated feature subsets
as.data.table(instance$archive)
#>       age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>  1: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3320312
#>  2: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3242188
#>  3: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2460938
#>  4: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2460938
#>  5: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2578125
#>  6: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3242188
#>  7: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3476562
#>  8: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.4804688
#>  9: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3437500
#> 10: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3242188
#>     runtime_learners           timestamp batch_nr      resample_result
#>  1:            0.082 2022-03-25 04:17:09        1 <ResampleResult[22]>
#>  2:            0.081 2022-03-25 04:17:09        2 <ResampleResult[22]>
#>  3:            0.075 2022-03-25 04:17:09        3 <ResampleResult[22]>
#>  4:            0.076 2022-03-25 04:17:09        4 <ResampleResult[22]>
#>  5:            0.077 2022-03-25 04:17:10        5 <ResampleResult[22]>
#>  6:            0.075 2022-03-25 04:17:10        6 <ResampleResult[22]>
#>  7:            0.077 2022-03-25 04:17:10        7 <ResampleResult[22]>
#>  8:            0.077 2022-03-25 04:17:10        8 <ResampleResult[22]>
#>  9:            0.082 2022-03-25 04:17:11        9 <ResampleResult[22]>
#> 10:            0.086 2022-03-25 04:17:11       10 <ResampleResult[22]>

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }