Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>      age glucose insulin  mass pedigree pregnant pressure triceps features
#> 1: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  glucose
#>    classif.ce
#> 1:  0.2421875

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3984375
#>   2:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>   3:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>   4: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.3046875
#>   5:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>   6:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3632812
#>   7:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.3242188
#>   8: FALSE   FALSE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.3320312
#>   9:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>  10: FALSE   FALSE    TRUE FALSE     TRUE    FALSE    FALSE   FALSE  0.3828125
#>  11: FALSE   FALSE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.3007812
#>  12:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3359375
#>  13:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>  14:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3437500
#>  15: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  16:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  17: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2734375
#>  18:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE    TRUE  0.3359375
#>  19:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.2617188
#>  20: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  21: FALSE   FALSE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.3164062
#>  22: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3671875
#>  23: FALSE   FALSE    TRUE FALSE    FALSE    FALSE     TRUE   FALSE  0.4023438
#>  24:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3984375
#>  25: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.3593750
#>  26:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE     TRUE    TRUE  0.3437500
#>  27:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2617188
#>  28: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3476562
#>  29: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  30:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2617188
#>  31: FALSE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.2695312
#>  32: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3671875
#>  33:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2695312
#>  34:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.3398438
#>  35: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3671875
#>  36:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE   FALSE  0.2500000
#>  37:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  38: FALSE    TRUE   FALSE FALSE    FALSE     TRUE     TRUE    TRUE  0.2500000
#>  39:  TRUE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3320312
#>  40:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2968750
#>  41:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3359375
#>  42:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2968750
#>  43: FALSE    TRUE   FALSE FALSE     TRUE    FALSE     TRUE   FALSE  0.2578125
#>  44:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2617188
#>  45:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3437500
#>  46:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3242188
#>  47:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2500000
#>  48:  TRUE    TRUE    TRUE FALSE    FALSE    FALSE     TRUE    TRUE  0.2617188
#>  49:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2617188
#>  50:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2617188
#>  51: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  52: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2695312
#>  53:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  54: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2890625
#>  55: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2421875
#>  56:  TRUE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3320312
#>  57: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  58: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  59:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>  60:  TRUE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2617188
#>  61: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2851562
#>  62:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2500000
#>  63:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.2460938
#>  64:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE   FALSE  0.2500000
#>  65: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  66: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.3515625
#>  67:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2500000
#>  68:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  69:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2695312
#>  70: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3671875
#>  71:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2968750
#>  72: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  73:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.2617188
#>  74:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>  75:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2578125
#>  76:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.2617188
#>  77:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  78:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3945312
#>  79:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.2460938
#>  80:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2617188
#>  81: FALSE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2851562
#>  82: FALSE    TRUE   FALSE FALSE     TRUE    FALSE     TRUE   FALSE  0.2578125
#>  83:  TRUE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2968750
#>  84:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2578125
#>  85:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2500000
#>  86: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3750000
#>  87: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE    TRUE  0.3710938
#>  88:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2695312
#>  89: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2421875
#>  90:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2695312
#>  91:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2695312
#>  92:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.3398438
#>  93: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  94:  TRUE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2460938
#>  95: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2656250
#>  96:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>  97:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2617188
#>  98: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.3320312
#>  99: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3984375
#> 100:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2578125
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.062 2022-07-28 04:28:41        1 <ResampleResult[21]>
#>   2:            0.063 2022-07-28 04:28:41        2 <ResampleResult[21]>
#>   3:            0.060 2022-07-28 04:28:41        3 <ResampleResult[21]>
#>   4:            0.062 2022-07-28 04:28:42        4 <ResampleResult[21]>
#>   5:            0.063 2022-07-28 04:28:42        5 <ResampleResult[21]>
#>   6:            0.062 2022-07-28 04:28:42        6 <ResampleResult[21]>
#>   7:            0.063 2022-07-28 04:28:42        7 <ResampleResult[21]>
#>   8:            0.070 2022-07-28 04:28:42        8 <ResampleResult[21]>
#>   9:            0.075 2022-07-28 04:28:43        9 <ResampleResult[21]>
#>  10:            0.077 2022-07-28 04:28:43       10 <ResampleResult[21]>
#>  11:            0.085 2022-07-28 04:28:43       11 <ResampleResult[21]>
#>  12:            0.078 2022-07-28 04:28:43       12 <ResampleResult[21]>
#>  13:            0.073 2022-07-28 04:28:44       13 <ResampleResult[21]>
#>  14:            0.069 2022-07-28 04:28:44       14 <ResampleResult[21]>
#>  15:            0.062 2022-07-28 04:28:44       15 <ResampleResult[21]>
#>  16:            0.065 2022-07-28 04:28:44       16 <ResampleResult[21]>
#>  17:            0.074 2022-07-28 04:28:44       17 <ResampleResult[21]>
#>  18:            0.082 2022-07-28 04:28:45       18 <ResampleResult[21]>
#>  19:            0.077 2022-07-28 04:28:45       19 <ResampleResult[21]>
#>  20:            0.057 2022-07-28 04:28:45       20 <ResampleResult[21]>
#>  21:            0.059 2022-07-28 04:28:45       21 <ResampleResult[21]>
#>  22:            0.063 2022-07-28 04:28:45       22 <ResampleResult[21]>
#>  23:            0.064 2022-07-28 04:28:46       23 <ResampleResult[21]>
#>  24:            0.061 2022-07-28 04:28:46       24 <ResampleResult[21]>
#>  25:            0.060 2022-07-28 04:28:46       25 <ResampleResult[21]>
#>  26:            0.059 2022-07-28 04:28:46       26 <ResampleResult[21]>
#>  27:            0.059 2022-07-28 04:28:46       27 <ResampleResult[21]>
#>  28:            0.059 2022-07-28 04:28:47       28 <ResampleResult[21]>
#>  29:            0.060 2022-07-28 04:28:47       29 <ResampleResult[21]>
#>  30:            0.059 2022-07-28 04:28:47       30 <ResampleResult[21]>
#>  31:            0.064 2022-07-28 04:28:47       31 <ResampleResult[21]>
#>  32:            0.060 2022-07-28 04:28:47       32 <ResampleResult[21]>
#>  33:            0.065 2022-07-28 04:28:47       33 <ResampleResult[21]>
#>  34:            0.066 2022-07-28 04:28:48       34 <ResampleResult[21]>
#>  35:            0.060 2022-07-28 04:28:48       35 <ResampleResult[21]>
#>  36:            0.060 2022-07-28 04:28:48       36 <ResampleResult[21]>
#>  37:            0.061 2022-07-28 04:28:48       37 <ResampleResult[21]>
#>  38:            0.061 2022-07-28 04:28:48       38 <ResampleResult[21]>
#>  39:            0.060 2022-07-28 04:28:49       39 <ResampleResult[21]>
#>  40:            0.061 2022-07-28 04:28:49       40 <ResampleResult[21]>
#>  41:            0.060 2022-07-28 04:28:49       41 <ResampleResult[21]>
#>  42:            0.060 2022-07-28 04:28:49       42 <ResampleResult[21]>
#>  43:            0.060 2022-07-28 04:28:49       43 <ResampleResult[21]>
#>  44:            0.059 2022-07-28 04:28:49       44 <ResampleResult[21]>
#>  45:            0.064 2022-07-28 04:28:50       45 <ResampleResult[21]>
#>  46:            0.071 2022-07-28 04:28:50       46 <ResampleResult[21]>
#>  47:            0.072 2022-07-28 04:28:50       47 <ResampleResult[21]>
#>  48:            0.061 2022-07-28 04:28:50       48 <ResampleResult[21]>
#>  49:            0.061 2022-07-28 04:28:50       49 <ResampleResult[21]>
#>  50:            0.061 2022-07-28 04:28:51       50 <ResampleResult[21]>
#>  51:            0.059 2022-07-28 04:28:51       51 <ResampleResult[21]>
#>  52:            0.060 2022-07-28 04:28:51       52 <ResampleResult[21]>
#>  53:            0.060 2022-07-28 04:28:51       53 <ResampleResult[21]>
#>  54:            0.061 2022-07-28 04:28:51       54 <ResampleResult[21]>
#>  55:            0.066 2022-07-28 04:28:51       55 <ResampleResult[21]>
#>  56:            0.070 2022-07-28 04:28:52       56 <ResampleResult[21]>
#>  57:            0.071 2022-07-28 04:28:52       57 <ResampleResult[21]>
#>  58:            0.075 2022-07-28 04:28:52       58 <ResampleResult[21]>
#>  59:            0.072 2022-07-28 04:28:52       59 <ResampleResult[21]>
#>  60:            0.067 2022-07-28 04:28:52       60 <ResampleResult[21]>
#>  61:            0.061 2022-07-28 04:28:53       61 <ResampleResult[21]>
#>  62:            0.063 2022-07-28 04:28:53       62 <ResampleResult[21]>
#>  63:            0.063 2022-07-28 04:28:53       63 <ResampleResult[21]>
#>  64:            0.062 2022-07-28 04:28:53       64 <ResampleResult[21]>
#>  65:            0.061 2022-07-28 04:28:53       65 <ResampleResult[21]>
#>  66:            0.061 2022-07-28 04:28:54       66 <ResampleResult[21]>
#>  67:            0.061 2022-07-28 04:28:54       67 <ResampleResult[21]>
#>  68:            0.061 2022-07-28 04:28:54       68 <ResampleResult[21]>
#>  69:            0.062 2022-07-28 04:28:54       69 <ResampleResult[21]>
#>  70:            0.067 2022-07-28 04:28:54       70 <ResampleResult[21]>
#>  71:            0.074 2022-07-28 04:28:55       71 <ResampleResult[21]>
#>  72:            0.070 2022-07-28 04:28:55       72 <ResampleResult[21]>
#>  73:            0.063 2022-07-28 04:28:55       73 <ResampleResult[21]>
#>  74:            0.062 2022-07-28 04:28:55       74 <ResampleResult[21]>
#>  75:            0.062 2022-07-28 04:28:55       75 <ResampleResult[21]>
#>  76:            0.062 2022-07-28 04:28:56       76 <ResampleResult[21]>
#>  77:            0.063 2022-07-28 04:28:56       77 <ResampleResult[21]>
#>  78:            0.060 2022-07-28 04:28:56       78 <ResampleResult[21]>
#>  79:            0.061 2022-07-28 04:28:56       79 <ResampleResult[21]>
#>  80:            0.063 2022-07-28 04:28:56       80 <ResampleResult[21]>
#>  81:            0.066 2022-07-28 04:28:56       81 <ResampleResult[21]>
#>  82:            0.070 2022-07-28 04:28:57       82 <ResampleResult[21]>
#>  83:            0.076 2022-07-28 04:28:57       83 <ResampleResult[21]>
#>  84:            0.076 2022-07-28 04:28:57       84 <ResampleResult[21]>
#>  85:            0.076 2022-07-28 04:28:57       85 <ResampleResult[21]>
#>  86:            0.076 2022-07-28 04:28:57       86 <ResampleResult[21]>
#>  87:            0.075 2022-07-28 04:28:58       87 <ResampleResult[21]>
#>  88:            0.096 2022-07-28 04:28:58       88 <ResampleResult[21]>
#>  89:            0.068 2022-07-28 04:28:58       89 <ResampleResult[21]>
#>  90:            0.061 2022-07-28 04:28:58       90 <ResampleResult[21]>
#>  91:            0.063 2022-07-28 04:28:58       91 <ResampleResult[21]>
#>  92:            0.061 2022-07-28 04:28:59       92 <ResampleResult[21]>
#>  93:            0.060 2022-07-28 04:28:59       93 <ResampleResult[21]>
#>  94:            0.064 2022-07-28 04:28:59       94 <ResampleResult[21]>
#>  95:            0.072 2022-07-28 04:28:59       95 <ResampleResult[21]>
#>  96:            0.081 2022-07-28 04:28:59       96 <ResampleResult[21]>
#>  97:            0.078 2022-07-28 04:29:00       97 <ResampleResult[21]>
#>  98:            0.073 2022-07-28 04:29:00       98 <ResampleResult[21]>
#>  99:            0.069 2022-07-28 04:29:00       99 <ResampleResult[21]>
#> 100:            0.062 2022-07-28 04:29:00      100 <ResampleResult[21]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }