Feature Selection via Sequential Search

Sequential search iteratively adds features to the set.

Sequential forward selection (strategy = fsf) extends the feature set in each iteration with the feature that increases the models performance the most. Sequential backward selection (strategy = fsb) follows the same idea but starts with all features and removes features from the set.

The feature selection terminates itself when min_features or max_features is reached. It is not necessary to set a termination criterion.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("sequential")
fs("sequential")

Parameters

min_features

integer(1)
Minimum number of features. By default, 1.

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

strategy

character(1)
Search method sfs (forward search) or sbs (backward search).

Super class

mlr3fselect::FSelector -> FSelectorSequential

Methods

Public methods

FSelectorSequential$new()

FSelectorSequential$optimization_path()

FSelectorSequential$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorSequential$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorSequential$optimization_path(inst, include_uhash = FALSE)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorSequential.

include_uhash

(logical(1))
Include uhash column?

Returns

data.table::data.table()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorSequential$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "sequential",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10
)

# best performing feature subset
instance$result
#>     age glucose insulin  mass pedigree pregnant pressure triceps    features
#> 1: TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE age,glucose
#>    classif.ce
#> 1:   0.265625

# all evaluated feature subsets
as.data.table(instance$archive)
#>       age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>  1:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3398438
#>  2: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2695312
#>  3: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.4804688
#>  4: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3750000
#>  5: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3281250
#>  6: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3359375
#>  7: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3671875
#>  8: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3437500
#>  9:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2656250
#> 10: FALSE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2695312
#> 11: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2773438
#> 12: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2734375
#> 13: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2773438
#> 14: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2695312
#> 15: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3125000
#>     runtime_learners           timestamp batch_nr      resample_result
#>  1:            0.056 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  2:            0.058 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  3:            0.057 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  4:            0.057 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  5:            0.063 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  6:            0.067 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  7:            0.058 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  8:            0.056 2022-03-04 04:14:44        1 <ResampleResult[22]>
#>  9:            0.057 2022-03-04 04:14:45        2 <ResampleResult[22]>
#> 10:            0.058 2022-03-04 04:14:45        2 <ResampleResult[22]>
#> 11:            0.064 2022-03-04 04:14:45        2 <ResampleResult[22]>
#> 12:            0.063 2022-03-04 04:14:45        2 <ResampleResult[22]>
#> 13:            0.057 2022-03-04 04:14:45        2 <ResampleResult[22]>
#> 14:            0.062 2022-03-04 04:14:45        2 <ResampleResult[22]>
#> 15:            0.058 2022-03-04 04:14:45        2 <ResampleResult[22]>

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }