Feature Selection via Shadow Variable Search

Shadow variable search creates for each feature a permutated copy and stops when one of them is selected.

The feature selection terminates itself when the first shadow variable is selected. It is not necessary to set a termination criterion.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi:10.1155/2017/1421409 .

Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi:10.1198/016214506000000843 .

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "shadow_variable_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
)

# best performing feature subset
instance$result
#>     age glucose insulin mass pedigree pregnant pressure triceps
#> 1: TRUE    TRUE   FALSE TRUE    FALSE    FALSE    FALSE   FALSE
#>            features classif.ce
#> 1: age,glucose,mass   0.234375

# all evaluated feature subsets
as.data.table(instance$archive)
#>       age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>  1:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#>  2: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2773438
#>  3: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3671875
#>  4: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3710938
#>  5: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3867188
#>  6: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3554688
#>  7: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3945312
#>  8: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3945312
#>  9: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.4101562
#> 10: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#> 11: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#> 12: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.4375000
#> 13: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.4062500
#> 14: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#> 15: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#> 16: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#> 17:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2695312
#> 18: FALSE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2890625
#> 19: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 20: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2812500
#> 21: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2617188
#> 22: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2617188
#> 23: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2500000
#> 24: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2929688
#> 25: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2812500
#> 26: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2773438
#> 27: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2929688
#> 28: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2734375
#> 29: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2929688
#> 30: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2773438
#> 31: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2851562
#> 32:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 33: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 34: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2695312
#> 35: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2421875
#> 36: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2500000
#> 37: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2460938
#> 38: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 39: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2695312
#> 40: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 41: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 42: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2539062
#> 43: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 44: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 45: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#> 46:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 47:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2578125
#> 48:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2343750
#> 49:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2460938
#> 50:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2500000
#> 51:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 52:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2578125
#> 53:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2539062
#> 54:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 55:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2421875
#> 56:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 57:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 58:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 59:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2578125
#> 60:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2343750
#> 61:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2460938
#> 62:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2500000
#> 63:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 64:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2578125
#> 65:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2539062
#> 66:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 67:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2421875
#> 68:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 69:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#> 70:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2343750
#>       age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>     runtime_learners           timestamp batch_nr permuted__age
#>  1:            0.087 2022-06-27 04:35:01        1         FALSE
#>  2:            0.086 2022-06-27 04:35:01        1         FALSE
#>  3:            0.087 2022-06-27 04:35:01        1         FALSE
#>  4:            0.088 2022-06-27 04:35:01        1         FALSE
#>  5:            0.086 2022-06-27 04:35:01        1         FALSE
#>  6:            0.086 2022-06-27 04:35:01        1         FALSE
#>  7:            0.086 2022-06-27 04:35:01        1         FALSE
#>  8:            0.094 2022-06-27 04:35:01        1         FALSE
#>  9:            0.084 2022-06-27 04:35:01        1          TRUE
#> 10:            0.085 2022-06-27 04:35:01        1         FALSE
#> 11:            0.092 2022-06-27 04:35:01        1         FALSE
#> 12:            0.095 2022-06-27 04:35:01        1         FALSE
#> 13:            0.094 2022-06-27 04:35:01        1         FALSE
#> 14:            0.091 2022-06-27 04:35:01        1         FALSE
#> 15:            0.083 2022-06-27 04:35:01        1         FALSE
#> 16:            0.085 2022-06-27 04:35:01        1         FALSE
#> 17:            0.095 2022-06-27 04:35:04        2         FALSE
#> 18:            0.089 2022-06-27 04:35:04        2         FALSE
#> 19:            0.097 2022-06-27 04:35:04        2         FALSE
#> 20:            0.089 2022-06-27 04:35:04        2         FALSE
#> 21:            0.098 2022-06-27 04:35:04        2         FALSE
#> 22:            0.087 2022-06-27 04:35:04        2         FALSE
#> 23:            0.089 2022-06-27 04:35:04        2         FALSE
#> 24:            0.105 2022-06-27 04:35:04        2          TRUE
#> 25:            0.098 2022-06-27 04:35:04        2         FALSE
#> 26:            0.088 2022-06-27 04:35:04        2         FALSE
#> 27:            0.088 2022-06-27 04:35:04        2         FALSE
#> 28:            0.089 2022-06-27 04:35:04        2         FALSE
#> 29:            0.099 2022-06-27 04:35:04        2         FALSE
#> 30:            0.087 2022-06-27 04:35:04        2         FALSE
#> 31:            0.138 2022-06-27 04:35:04        2         FALSE
#> 32:            0.109 2022-06-27 04:35:07        3         FALSE
#> 33:            0.087 2022-06-27 04:35:07        3         FALSE
#> 34:            0.091 2022-06-27 04:35:07        3         FALSE
#> 35:            0.088 2022-06-27 04:35:07        3         FALSE
#> 36:            0.105 2022-06-27 04:35:07        3         FALSE
#> 37:            0.106 2022-06-27 04:35:07        3         FALSE
#> 38:            0.088 2022-06-27 04:35:07        3          TRUE
#> 39:            0.090 2022-06-27 04:35:07        3         FALSE
#> 40:            0.089 2022-06-27 04:35:07        3         FALSE
#> 41:            0.106 2022-06-27 04:35:07        3         FALSE
#> 42:            0.089 2022-06-27 04:35:07        3         FALSE
#> 43:            0.089 2022-06-27 04:35:07        3         FALSE
#> 44:            0.102 2022-06-27 04:35:07        3         FALSE
#> 45:            0.089 2022-06-27 04:35:07        3         FALSE
#> 46:            0.092 2022-06-27 04:35:10        4         FALSE
#> 47:            0.117 2022-06-27 04:35:10        4         FALSE
#> 48:            0.091 2022-06-27 04:35:10        4         FALSE
#> 49:            0.092 2022-06-27 04:35:10        4         FALSE
#> 50:            0.094 2022-06-27 04:35:10        4         FALSE
#> 51:            0.094 2022-06-27 04:35:10        4          TRUE
#> 52:            0.118 2022-06-27 04:35:10        4         FALSE
#> 53:            0.092 2022-06-27 04:35:10        4         FALSE
#> 54:            0.089 2022-06-27 04:35:10        4         FALSE
#> 55:            0.091 2022-06-27 04:35:10        4         FALSE
#> 56:            0.106 2022-06-27 04:35:10        4         FALSE
#> 57:            0.150 2022-06-27 04:35:10        4         FALSE
#> 58:            0.088 2022-06-27 04:35:10        4         FALSE
#> 59:            0.092 2022-06-27 04:35:12        5         FALSE
#> 60:            0.105 2022-06-27 04:35:12        5         FALSE
#> 61:            0.106 2022-06-27 04:35:12        5         FALSE
#> 62:            0.118 2022-06-27 04:35:12        5         FALSE
#> 63:            0.092 2022-06-27 04:35:12        5          TRUE
#> 64:            0.090 2022-06-27 04:35:12        5         FALSE
#> 65:            0.092 2022-06-27 04:35:12        5         FALSE
#> 66:            0.118 2022-06-27 04:35:12        5         FALSE
#> 67:            0.111 2022-06-27 04:35:12        5         FALSE
#> 68:            0.092 2022-06-27 04:35:12        5         FALSE
#> 69:            0.088 2022-06-27 04:35:12        5         FALSE
#> 70:            0.091 2022-06-27 04:35:12        5         FALSE
#>     runtime_learners           timestamp batch_nr permuted__age
#>     permuted__glucose permuted__insulin permuted__mass permuted__pedigree
#>  1:             FALSE             FALSE          FALSE              FALSE
#>  2:             FALSE             FALSE          FALSE              FALSE
#>  3:             FALSE             FALSE          FALSE              FALSE
#>  4:             FALSE             FALSE          FALSE              FALSE
#>  5:             FALSE             FALSE          FALSE              FALSE
#>  6:             FALSE             FALSE          FALSE              FALSE
#>  7:             FALSE             FALSE          FALSE              FALSE
#>  8:             FALSE             FALSE          FALSE              FALSE
#>  9:             FALSE             FALSE          FALSE              FALSE
#> 10:              TRUE             FALSE          FALSE              FALSE
#> 11:             FALSE              TRUE          FALSE              FALSE
#> 12:             FALSE             FALSE           TRUE              FALSE
#> 13:             FALSE             FALSE          FALSE               TRUE
#> 14:             FALSE             FALSE          FALSE              FALSE
#> 15:             FALSE             FALSE          FALSE              FALSE
#> 16:             FALSE             FALSE          FALSE              FALSE
#> 17:             FALSE             FALSE          FALSE              FALSE
#> 18:             FALSE             FALSE          FALSE              FALSE
#> 19:             FALSE             FALSE          FALSE              FALSE
#> 20:             FALSE             FALSE          FALSE              FALSE
#> 21:             FALSE             FALSE          FALSE              FALSE
#> 22:             FALSE             FALSE          FALSE              FALSE
#> 23:             FALSE             FALSE          FALSE              FALSE
#> 24:             FALSE             FALSE          FALSE              FALSE
#> 25:              TRUE             FALSE          FALSE              FALSE
#> 26:             FALSE              TRUE          FALSE              FALSE
#> 27:             FALSE             FALSE           TRUE              FALSE
#> 28:             FALSE             FALSE          FALSE               TRUE
#> 29:             FALSE             FALSE          FALSE              FALSE
#> 30:             FALSE             FALSE          FALSE              FALSE
#> 31:             FALSE             FALSE          FALSE              FALSE
#> 32:             FALSE             FALSE          FALSE              FALSE
#> 33:             FALSE             FALSE          FALSE              FALSE
#> 34:             FALSE             FALSE          FALSE              FALSE
#> 35:             FALSE             FALSE          FALSE              FALSE
#> 36:             FALSE             FALSE          FALSE              FALSE
#> 37:             FALSE             FALSE          FALSE              FALSE
#> 38:             FALSE             FALSE          FALSE              FALSE
#> 39:              TRUE             FALSE          FALSE              FALSE
#> 40:             FALSE              TRUE          FALSE              FALSE
#> 41:             FALSE             FALSE           TRUE              FALSE
#> 42:             FALSE             FALSE          FALSE               TRUE
#> 43:             FALSE             FALSE          FALSE              FALSE
#> 44:             FALSE             FALSE          FALSE              FALSE
#> 45:             FALSE             FALSE          FALSE              FALSE
#> 46:             FALSE             FALSE          FALSE              FALSE
#> 47:             FALSE             FALSE          FALSE              FALSE
#> 48:             FALSE             FALSE          FALSE              FALSE
#> 49:             FALSE             FALSE          FALSE              FALSE
#> 50:             FALSE             FALSE          FALSE              FALSE
#> 51:             FALSE             FALSE          FALSE              FALSE
#> 52:              TRUE             FALSE          FALSE              FALSE
#> 53:             FALSE              TRUE          FALSE              FALSE
#> 54:             FALSE             FALSE           TRUE              FALSE
#> 55:             FALSE             FALSE          FALSE               TRUE
#> 56:             FALSE             FALSE          FALSE              FALSE
#> 57:             FALSE             FALSE          FALSE              FALSE
#> 58:             FALSE             FALSE          FALSE              FALSE
#> 59:             FALSE             FALSE          FALSE              FALSE
#> 60:             FALSE             FALSE          FALSE              FALSE
#> 61:             FALSE             FALSE          FALSE              FALSE
#> 62:             FALSE             FALSE          FALSE              FALSE
#> 63:             FALSE             FALSE          FALSE              FALSE
#> 64:              TRUE             FALSE          FALSE              FALSE
#> 65:             FALSE              TRUE          FALSE              FALSE
#> 66:             FALSE             FALSE           TRUE              FALSE
#> 67:             FALSE             FALSE          FALSE               TRUE
#> 68:             FALSE             FALSE          FALSE              FALSE
#> 69:             FALSE             FALSE          FALSE              FALSE
#> 70:             FALSE             FALSE          FALSE              FALSE
#>     permuted__glucose permuted__insulin permuted__mass permuted__pedigree
#>     permuted__pregnant permuted__pressure permuted__triceps
#>  1:              FALSE              FALSE             FALSE
#>  2:              FALSE              FALSE             FALSE
#>  3:              FALSE              FALSE             FALSE
#>  4:              FALSE              FALSE             FALSE
#>  5:              FALSE              FALSE             FALSE
#>  6:              FALSE              FALSE             FALSE
#>  7:              FALSE              FALSE             FALSE
#>  8:              FALSE              FALSE             FALSE
#>  9:              FALSE              FALSE             FALSE
#> 10:              FALSE              FALSE             FALSE
#> 11:              FALSE              FALSE             FALSE
#> 12:              FALSE              FALSE             FALSE
#> 13:              FALSE              FALSE             FALSE
#> 14:               TRUE              FALSE             FALSE
#> 15:              FALSE               TRUE             FALSE
#> 16:              FALSE              FALSE              TRUE
#> 17:              FALSE              FALSE             FALSE
#> 18:              FALSE              FALSE             FALSE
#> 19:              FALSE              FALSE             FALSE
#> 20:              FALSE              FALSE             FALSE
#> 21:              FALSE              FALSE             FALSE
#> 22:              FALSE              FALSE             FALSE
#> 23:              FALSE              FALSE             FALSE
#> 24:              FALSE              FALSE             FALSE
#> 25:              FALSE              FALSE             FALSE
#> 26:              FALSE              FALSE             FALSE
#> 27:              FALSE              FALSE             FALSE
#> 28:              FALSE              FALSE             FALSE
#> 29:               TRUE              FALSE             FALSE
#> 30:              FALSE               TRUE             FALSE
#> 31:              FALSE              FALSE              TRUE
#> 32:              FALSE              FALSE             FALSE
#> 33:              FALSE              FALSE             FALSE
#> 34:              FALSE              FALSE             FALSE
#> 35:              FALSE              FALSE             FALSE
#> 36:              FALSE              FALSE             FALSE
#> 37:              FALSE              FALSE             FALSE
#> 38:              FALSE              FALSE             FALSE
#> 39:              FALSE              FALSE             FALSE
#> 40:              FALSE              FALSE             FALSE
#> 41:              FALSE              FALSE             FALSE
#> 42:              FALSE              FALSE             FALSE
#> 43:               TRUE              FALSE             FALSE
#> 44:              FALSE               TRUE             FALSE
#> 45:              FALSE              FALSE              TRUE
#> 46:              FALSE              FALSE             FALSE
#> 47:              FALSE              FALSE             FALSE
#> 48:              FALSE              FALSE             FALSE
#> 49:              FALSE              FALSE             FALSE
#> 50:              FALSE              FALSE             FALSE
#> 51:              FALSE              FALSE             FALSE
#> 52:              FALSE              FALSE             FALSE
#> 53:              FALSE              FALSE             FALSE
#> 54:              FALSE              FALSE             FALSE
#> 55:              FALSE              FALSE             FALSE
#> 56:               TRUE              FALSE             FALSE
#> 57:              FALSE               TRUE             FALSE
#> 58:              FALSE              FALSE              TRUE
#> 59:              FALSE              FALSE             FALSE
#> 60:              FALSE              FALSE             FALSE
#> 61:              FALSE              FALSE             FALSE
#> 62:              FALSE              FALSE             FALSE
#> 63:              FALSE              FALSE             FALSE
#> 64:              FALSE              FALSE             FALSE
#> 65:              FALSE              FALSE             FALSE
#> 66:              FALSE              FALSE             FALSE
#> 67:              FALSE              FALSE             FALSE
#> 68:               TRUE              FALSE             FALSE
#> 69:              FALSE               TRUE             FALSE
#> 70:              FALSE              FALSE              TRUE
#>     permuted__pregnant permuted__pressure permuted__triceps
#>          resample_result
#>  1: <ResampleResult[22]>
#>  2: <ResampleResult[22]>
#>  3: <ResampleResult[22]>
#>  4: <ResampleResult[22]>
#>  5: <ResampleResult[22]>
#>  6: <ResampleResult[22]>
#>  7: <ResampleResult[22]>
#>  8: <ResampleResult[22]>
#>  9: <ResampleResult[22]>
#> 10: <ResampleResult[22]>
#> 11: <ResampleResult[22]>
#> 12: <ResampleResult[22]>
#> 13: <ResampleResult[22]>
#> 14: <ResampleResult[22]>
#> 15: <ResampleResult[22]>
#> 16: <ResampleResult[22]>
#> 17: <ResampleResult[22]>
#> 18: <ResampleResult[22]>
#> 19: <ResampleResult[22]>
#> 20: <ResampleResult[22]>
#> 21: <ResampleResult[22]>
#> 22: <ResampleResult[22]>
#> 23: <ResampleResult[22]>
#> 24: <ResampleResult[22]>
#> 25: <ResampleResult[22]>
#> 26: <ResampleResult[22]>
#> 27: <ResampleResult[22]>
#> 28: <ResampleResult[22]>
#> 29: <ResampleResult[22]>
#> 30: <ResampleResult[22]>
#> 31: <ResampleResult[22]>
#> 32: <ResampleResult[22]>
#> 33: <ResampleResult[22]>
#> 34: <ResampleResult[22]>
#> 35: <ResampleResult[22]>
#> 36: <ResampleResult[22]>
#> 37: <ResampleResult[22]>
#> 38: <ResampleResult[22]>
#> 39: <ResampleResult[22]>
#> 40: <ResampleResult[22]>
#> 41: <ResampleResult[22]>
#> 42: <ResampleResult[22]>
#> 43: <ResampleResult[22]>
#> 44: <ResampleResult[22]>
#> 45: <ResampleResult[22]>
#> 46: <ResampleResult[22]>
#> 47: <ResampleResult[22]>
#> 48: <ResampleResult[22]>
#> 49: <ResampleResult[22]>
#> 50: <ResampleResult[22]>
#> 51: <ResampleResult[22]>
#> 52: <ResampleResult[22]>
#> 53: <ResampleResult[22]>
#> 54: <ResampleResult[22]>
#> 55: <ResampleResult[22]>
#> 56: <ResampleResult[22]>
#> 57: <ResampleResult[22]>
#> 58: <ResampleResult[22]>
#> 59: <ResampleResult[22]>
#> 60: <ResampleResult[22]>
#> 61: <ResampleResult[22]>
#> 62: <ResampleResult[22]>
#> 63: <ResampleResult[22]>
#> 64: <ResampleResult[22]>
#> 65: <ResampleResult[22]>
#> 66: <ResampleResult[22]>
#> 67: <ResampleResult[22]>
#> 68: <ResampleResult[22]>
#> 69: <ResampleResult[22]>
#> 70: <ResampleResult[22]>
#>          resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }