AutoFeselect

The AutoFSelect is a mlr3::Learner which automatically selects the optimal feature set and fits the model on the training data with the optimal feature set.

Super class

mlr3::Learner -> AutoFSelect

Public fields

instance_args

list()

fselect

FSelect

store_fselect_instance

logical(1)

Active bindings

archive

FSelectInstance archive

learner

mlr3::Learner

fselect_instance

FSelectInstance

fselect_result

list()
Result of the feature selection i.e. the optimal feature set and its estimated performances.

Methods

Public methods

AutoFSelect$new()

AutoFSelect$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

AutoFSelect$new(learner, resampling, measures, terminator, fselect)

Arguments

learner

mlr3::Learner

resampling

mlr3::Resampling

measures

list of mlr3::Measure

terminator

Terminator

fselect

FSelect

Method clone()

The objects of this class are cloneable with this method.

Usage

AutoFSelect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measures = msr("classif.ce")

terminator = term("evals", n_evals = 15)
fs = fs("exhaustive")
afs = AutoFSelect$new(learner, resampling, measures, terminator, fs)
afs$store_fselect_instance = TRUE

afs$train(task)
afs$model

#> $learner
#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $fselect_instance
#> <FSelectInstance>
#> * Objective: <ObjectiveFSelect:feature_selection>
#> * Search Space: <ParamSet> Public: add: function (p) add_dep: function
#>   (id, on, cond) assert: function (xs, .var.name = vname(xs)) check:
#>   function (xs) class: active binding clone: function (deep = FALSE)
#>   default: active binding deps: active binding get_values: function
#>   (class = NULL, is_bounded = NULL, tags = NULL) has_deps: active
#>   binding has_trafo: active binding ids: function (class = NULL,
#>   is_bounded = NULL, tags = NULL) initialize: function (params =
#>   named_list()) is_bounded: active binding is_categ: active binding
#>   is_empty: active binding is_number: active binding length: active
#>   binding levels: active binding lower: active binding nlevels: active
#>   binding params: active binding print: function (..., hide_cols =
#>   c("nlevels", "is_bounded", "special_vals", set_id: active binding
#>   special_vals: active binding storage_type: active binding subset:
#>   function (ids) tags: active binding test: function (xs) trafo: active
#>   binding upper: active binding values: active binding Private: .deps:
#>   data.table, data.frame .params: list .set_id: .trafo: NULL .values:
#>   list deep_clone: function (name, value) get_member_with_idnames:
#>   function (member, astype)
#> * Terminator: <TerminatorEvals>
#> * Terminated: FALSE
#> <Archive>
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:         TRUE       FALSE        FALSE       FALSE       0.06
#>  2:        FALSE        TRUE        FALSE       FALSE       0.06
#>  3:        FALSE       FALSE         TRUE       FALSE       0.50
#>  4:        FALSE       FALSE        FALSE        TRUE       0.44
#>  5:         TRUE        TRUE        FALSE       FALSE       0.06
#>  6:         TRUE       FALSE         TRUE       FALSE       0.06
#>  7:         TRUE       FALSE        FALSE        TRUE       0.06
#>  8:        FALSE        TRUE         TRUE       FALSE       0.06
#>  9:        FALSE        TRUE        FALSE        TRUE       0.06
#> 10:        FALSE       FALSE         TRUE        TRUE       0.50
#> 11:         TRUE        TRUE         TRUE       FALSE       0.06
#> 12:         TRUE        TRUE        FALSE        TRUE       0.06
#> 13:         TRUE       FALSE         TRUE        TRUE       0.06
#> 14:        FALSE        TRUE         TRUE        TRUE       0.06
#> 15:         TRUE        TRUE         TRUE        TRUE       0.06
#>      resample_result  opt_x  timestamp batch_nr
#>  1: <ResampleResult> <list> 1590729624        1
#>  2: <ResampleResult> <list> 1590729624        1
#>  3: <ResampleResult> <list> 1590729624        1
#>  4: <ResampleResult> <list> 1590729624        1
#>  5: <ResampleResult> <list> 1590729625        2
#>  6: <ResampleResult> <list> 1590729625        2
#>  7: <ResampleResult> <list> 1590729625        2
#>  8: <ResampleResult> <list> 1590729625        2
#>  9: <ResampleResult> <list> 1590729625        2
#> 10: <ResampleResult> <list> 1590729625        2
#> 11: <ResampleResult> <list> 1590729625        3
#> 12: <ResampleResult> <list> 1590729625        3
#> 13: <ResampleResult> <list> 1590729625        3
#> 14: <ResampleResult> <list> 1590729625        3
#> 15: <ResampleResult> <list> 1590729625        4
#> 

afs$learner

#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights