AutoFSelector

The AutoFSelector wraps a mlr3::Learner and augments it with an automatic feature selection. The auto_fselector() function creates an AutoFSelector object.

Details

The AutoFSelector is a mlr3::Learner which wraps another mlr3::Learner and performs the following steps during $train():

The wrapped (inner) learner is trained on the feature subsets via resampling. The feature selection can be specified by providing a FSelector, a bbotk::Terminator, a mlr3::Resampling and a mlr3::Measure.

A final model is fit on the complete training data with the best found feature subset.

During $predict() the AutoFSelector just calls the predict method of the wrapped (inner) learner.

Resources

book chapter on automatic feature selection.

Nested Resampling

Nested resampling can be performed by passing an AutoFSelector object to mlr3::resample() or mlr3::benchmark(). To access the inner resampling results, set store_fselect_instance = TRUE and execute mlr3::resample() or mlr3::benchmark() with store_models = TRUE (see examples). The mlr3::Resampling passed to the AutoFSelector is meant to be the inner resampling, operating on the training set of an arbitrary outer resampling. For this reason it is not feasible to pass an instantiated mlr3::Resampling here.

Super class

mlr3::Learner -> AutoFSelector

Public fields

instance_args

(list())
All arguments from construction to create the FSelectInstanceSingleCrit.

fselector

(FSelector)
Optimization algorithm.

Active bindings

archive

([ArchiveFSelect)
Returns FSelectInstanceSingleCrit archive.

learner

(mlr3::Learner)
Trained learner.

fselect_instance

(FSelectInstanceSingleCrit)
Internally created feature selection instance with all intermediate results.

fselect_result

(data.table::data.table)
Short-cut to $result from FSelectInstanceSingleCrit.

predict_type

(character(1))
Stores the currently active predict type, e.g. "response". Must be an element of $predict_types.

hash

(character(1))
Hash (unique identifier) for this object.

Methods

Public methods

AutoFSelector$new()

AutoFSelector$base_learner()

AutoFSelector$importance()

AutoFSelector$selected_features()

AutoFSelector$oob_error()

AutoFSelector$loglik()

AutoFSelector$print()

AutoFSelector$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

AutoFSelector$new(
  learner,
  resampling,
  measure = NULL,
  terminator,
  fselector,
  store_fselect_instance = TRUE,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE
)

Arguments

learner

(mlr3::Learner)
Learner to optimize the feature subset for.

resampling

(mlr3::Resampling)
Resampling that is used to evaluated the performance of the feature subsets. Uninstantiated resamplings are instantiated during construction so that all feature subsets are evaluated on the same data splits. Already instantiated resamplings are kept unchanged.

measure

(mlr3::Measure)
Measure to optimize. If NULL, default measure is used.

terminator

(Terminator)
Stop criterion of the feature selection.

fselector

(FSelector)
Optimization algorithm.

store_fselect_instance

(logical(1))
If TRUE (default), stores the internally created FSelectInstanceSingleCrit with all intermediate results in slot $fselect_instance. Is set to TRUE, if store_models = TRUE

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1)). Store models in benchmark result?

check_values

(logical(1))
Check the parameters before the evaluation and the results for validity?

Method base_learner()

Extracts the base learner from nested learner objects like GraphLearner in mlr3pipelines. If recursive = 0, the (tuned) learner is returned.

Usage

AutoFSelector$base_learner(recursive = Inf)

Arguments

recursive

(integer(1))
Depth of recursion for multiple nested objects.

Returns

Learner.

Method importance()

The importance scores of the final model.

Usage

AutoFSelector$importance()

Returns

Named numeric().

Method selected_features()

The selected features of the final model. These features are selected internally by the learner.

Usage

AutoFSelector$selected_features()

Returns

character().

Method oob_error()

The out-of-bag error of the final model.

Usage

AutoFSelector$oob_error()

Returns

numeric(1).

Method loglik()

The log-likelihood of the final model.

Usage

AutoFSelector$loglik()

Returns

logLik. Printer.

Method print()

Usage

AutoFSelector$print()

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

AutoFSelector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Automafsic Feafsure Selection

task = tsk("penguins")
train_set = sample(task$nrow, 0.8 * task$nrow)
test_set = setdiff(seq_len(task$nrow), train_set)

afs = auto_fselector(
  method = fs("random_search"),
  learner = lrn("classif.rpart"),
  resampling = rsmp ("holdout"),
  measure = msr("classif.ce"),
  term_evals = 4)

# optimize feafsure subset and fit final model
afs$train(task, row_ids = train_set)

# predict with final model
afs$predict(task, row_ids = test_set)
#> <PredictionClassif> for 69 observations:
#>     row_ids     truth  response
#>           1    Adelie    Adelie
#>          11    Adelie    Adelie
#>          12    Adelie    Adelie
#> ---                            
#>         324 Chinstrap Chinstrap
#>         336 Chinstrap Chinstrap
#>         337 Chinstrap Chinstrap

# show fselect result
afs$fselect_result
#>    bill_depth bill_length body_mass flipper_length island  sex year
#> 1:       TRUE        TRUE      TRUE           TRUE   TRUE TRUE TRUE
#>                                                          features classif.ce
#> 1: bill_depth,bill_length,body_mass,flipper_length,island,sex,... 0.06521739

# model slot contains trained learner and fselect instance
afs$model
#> $learner
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: mlr3, rpart
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $features
#> [1] "bill_depth"     "bill_length"    "body_mass"      "flipper_length"
#> [5] "island"         "sex"            "year"          
#> 
#> $fselect_instance
#> <FSelectInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveFSelect:classif.rpart_on_penguins>
#> * Search Space:
#>                id    class lower upper nlevels
#> 1:     bill_depth ParamLgl    NA    NA       2
#> 2:    bill_length ParamLgl    NA    NA       2
#> 3:      body_mass ParamLgl    NA    NA       2
#> 4: flipper_length ParamLgl    NA    NA       2
#> 5:         island ParamLgl    NA    NA       2
#> 6:            sex ParamLgl    NA    NA       2
#> 7:           year ParamLgl    NA    NA       2
#> * Terminator: <TerminatorEvals>
#> * Result:
#>    bill_depth bill_length body_mass flipper_length island  sex year classif.ce
#> 1:       TRUE        TRUE      TRUE           TRUE   TRUE TRUE TRUE 0.06521739
#> * Archive:
#>    bill_depth bill_length body_mass flipper_length island   sex  year
#> 1:       TRUE        TRUE      TRUE           TRUE   TRUE  TRUE  TRUE
#> 2:      FALSE        TRUE      TRUE          FALSE   TRUE FALSE  TRUE
#> 3:       TRUE        TRUE      TRUE           TRUE   TRUE  TRUE  TRUE
#> 4:      FALSE       FALSE     FALSE           TRUE  FALSE  TRUE FALSE
#>    classif.ce
#> 1: 0.06521739
#> 2: 0.06521739
#> 3: 0.06521739
#> 4: 0.18478261
#> 

# shortcut trained learner
afs$learner
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: rpart
#> * Parameters: xval=0
#> * Packages: mlr3, rpart
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

# shortcut fselect instance
afs$tuning_instance
#> NULL

# Nested Resampling

afs = auto_fselector(
  method = fs("random_search"),
  learner = lrn("classif.rpart"),
  resampling = rsmp ("holdout"),
  measure = msr("classif.ce"),
  term_evals = 4)

resampling_outer = rsmp("cv", folds = 3)
rr = resample(task, afs, resampling_outer, store_models = TRUE)

# retrieve inner tuning results.
extract_inner_fselect_results(rr)
#>    iteration bill_depth bill_length body_mass flipper_length island   sex year
#> 1:         1       TRUE       FALSE     FALSE           TRUE   TRUE  TRUE TRUE
#> 2:         2       TRUE        TRUE      TRUE           TRUE   TRUE FALSE TRUE
#> 3:         3       TRUE        TRUE      TRUE           TRUE   TRUE  TRUE TRUE
#>    classif.ce                                                       features
#> 1: 0.15789474                      bill_depth,flipper_length,island,sex,year
#> 2: 0.05263158    bill_depth,bill_length,body_mass,flipper_length,island,year
#> 3: 0.09090909 bill_depth,bill_length,body_mass,flipper_length,island,sex,...
#>     task_id              learner_id resampling_id
#> 1: penguins classif.rpart.fselector            cv
#> 2: penguins classif.rpart.fselector            cv
#> 3: penguins classif.rpart.fselector            cv

# performance scores estimated on the outer resampling
rr$score()
#>                 task  task_id             learner              learner_id
#> 1: <TaskClassif[50]> penguins <AutoFSelector[46]> classif.rpart.fselector
#> 2: <TaskClassif[50]> penguins <AutoFSelector[46]> classif.rpart.fselector
#> 3: <TaskClassif[50]> penguins <AutoFSelector[46]> classif.rpart.fselector
#>            resampling resampling_id iteration              prediction
#> 1: <ResamplingCV[20]>            cv         1 <PredictionClassif[20]>
#> 2: <ResamplingCV[20]>            cv         2 <PredictionClassif[20]>
#> 3: <ResamplingCV[20]>            cv         3 <PredictionClassif[20]>
#>    classif.ce
#> 1: 0.13043478
#> 2: 0.06086957
#> 3: 0.07894737

# unbiased performance of the final model trained on the full dafsa set
rr$aggregate()
#> classif.ce 
#> 0.09008391