FSelect

Abstract FSelect class that implements the base functionality each FSelect subclass must provide. A FSelect object describes the feature selection strategy, i.e. how to optimize the black-box function and its feasible set defined by the ObjectiveFSelect object.

A list of measures can be passed to the instance, and they will always be all evaluated. However, single-criteria algorithms optimize only the first measure

A FSelect object must provide a private method $.assign_result(inst) that writes the result to the FSelectInstance using the $assign_result() method of the instance in order to store the best selected feature subset and its estimated performance vector at the end of the optimization.

Super class

bbotk::Optimizer -> FSelect

Methods

Public methods

FSelect$new()

FSelect$optimize()

FSelect$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelect$new(param_set, properties, packages = character(0))

Arguments

param_set

paradox::ParamSet

properties

character()

packages

character()

param_classes

character()

Method optimize()

Performs the feature selection on a FSelectInstance until termination.

Usage

FSelect$optimize(inst)

Arguments

inst

FSelectInstance

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = term("evals", n_evals = 10)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("random")
fs$optimize(instance)
instance$result

#> $feat
#> [1] "Petal.Width" "Sepal.Width"
#> 
#> $perf
#> classif.ce 
#>       0.06 
#> 

instance$archive$data

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:        FALSE       FALSE        FALSE        TRUE       0.54
#>  2:        FALSE        TRUE        FALSE        TRUE       0.06
#>  3:         TRUE       FALSE        FALSE       FALSE       0.10
#>  4:        FALSE        TRUE        FALSE        TRUE       0.06
#>  5:         TRUE        TRUE         TRUE       FALSE       0.10
#>  6:        FALSE        TRUE        FALSE        TRUE       0.06
#>  7:        FALSE       FALSE         TRUE        TRUE       0.28
#>  8:         TRUE       FALSE        FALSE        TRUE       0.10
#>  9:         TRUE       FALSE         TRUE       FALSE       0.10
#> 10:        FALSE       FALSE         TRUE        TRUE       0.28
#>      resample_result  opt_x  timestamp batch_nr
#>  1: <ResampleResult> <list> 1589088384        1
#>  2: <ResampleResult> <list> 1589088384        1
#>  3: <ResampleResult> <list> 1589088384        1
#>  4: <ResampleResult> <list> 1589088384        1
#>  5: <ResampleResult> <list> 1589088384        1
#>  6: <ResampleResult> <list> 1589088384        1
#>  7: <ResampleResult> <list> 1589088384        1
#>  8: <ResampleResult> <list> 1589088384        1
#>  9: <ResampleResult> <list> 1589088384        1
#> 10: <ResampleResult> <list> 1589088384        1