FSelectEvolutionary Class

Subclass for evolutionary feature selection. Calls ecr::ecr() from package ecr.

Dictionary

This FSelect can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("evolutionary")
fs("evolutionary")

Parameters

mu

integer(1)

lambda

integer(1)

p.recomb

double(1)

p.mut

double(1)

survival.strategy

character(1)

n.elite

integer(1)

initial.solutions

list of integer()

parent.selector

ecr selector

survival.selector

ecr selector

For the meaning of the control parameter, see ecr::ecr(). Note that mu and lambda must be set by the user. The terminators parameter is replaced by the Terminator subclasses.

Super classes

bbotk::Optimizer -> mlr3fselect::FSelect -> FSelectEvolutionary

Methods

Public methods

FSelectEvolutionary$new()

FSelectEvolutionary$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$print()

mlr3fselect::FSelect$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectEvolutionary$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectEvolutionary$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = term("evals", n_evals = 10)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("evolutionary", mu = 10, lambda = 5)
fs$optimize(instance)
instance$result

#> $feat
#> [1] "Petal.Width"
#> 
#> $perf
#> classif.ce 
#>       0.04 
#> 

instance$archive$data

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:        FALSE       FALSE         TRUE       FALSE       0.38
#>  2:         TRUE       FALSE        FALSE       FALSE       0.06
#>  3:         TRUE        TRUE         TRUE       FALSE       0.06
#>  4:         TRUE       FALSE        FALSE       FALSE       0.06
#>  5:         TRUE       FALSE        FALSE        TRUE       0.06
#>  6:        FALSE        TRUE        FALSE       FALSE       0.04
#>  7:         TRUE       FALSE        FALSE        TRUE       0.06
#>  8:         TRUE       FALSE        FALSE        TRUE       0.06
#>  9:         TRUE        TRUE         TRUE       FALSE       0.06
#> 10:         TRUE       FALSE        FALSE       FALSE       0.06
#>      resample_result  opt_x  timestamp batch_nr
#>  1: <ResampleResult> <list> 1590650647        1
#>  2: <ResampleResult> <list> 1590650647        2
#>  3: <ResampleResult> <list> 1590650647        3
#>  4: <ResampleResult> <list> 1590650647        4
#>  5: <ResampleResult> <list> 1590650647        5
#>  6: <ResampleResult> <list> 1590650647        6
#>  7: <ResampleResult> <list> 1590650647        7
#>  8: <ResampleResult> <list> 1590650647        8
#>  9: <ResampleResult> <list> 1590650647        9
#> 10: <ResampleResult> <list> 1590650647       10