Feature Selection via Design Points

FSelectorDesignPoints class that implements feature selection w.r.t. fixed feature sets. We simply search over a set of feature subsets fully specified by the user. The feature sets are evaluated in order as given.

In order to support general termination criteria and parallelization, we evaluate feature sets in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("design_points")
fs("design_points")

Parameters

batch_size

integer(1)
Maximum number of configurations to try in a batch.

design

data.table::data.table
Design points to try in search, one per row.

Super classes

mlr3fselect::FSelector -> mlr3fselect::FSelectorFromOptimizer -> FSelectorDesignPoints

Methods

Public methods

FSelectorDesignPoints$new()

FSelectorDesignPoints$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$print()

mlr3fselect::FSelectorFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorDesignPoints$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorDesignPoints$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(data.table)

terminator = trm("evals", n_evals = 10)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

design = data.table(Petal.Length = c(TRUE, FALSE),
  Petal.Width = c(TRUE, FALSE),
  Sepal.Length = c(FALSE, TRUE),
  Sepal.Width = c(FALSE, TRUE))

fselector = fs("design_points", design = design)
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)
#>    Petal.Length Petal.Width Sepal.Length Sepal.Width                 features
#> 1:         TRUE        TRUE        FALSE       FALSE Petal.Length,Petal.Width
#>    classif.ce
#> 1:       0.08

# Returns best scoring evaluation
instance$result
#>    Petal.Length Petal.Width Sepal.Length Sepal.Width                 features
#> 1:         TRUE        TRUE        FALSE       FALSE Petal.Length,Petal.Width
#>    classif.ce
#> 1:       0.08

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }
#>    Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#> 1:         TRUE        TRUE        FALSE       FALSE       0.08
#> 2:        FALSE       FALSE         TRUE        TRUE       0.38
#>    runtime_learners           timestamp batch_nr      resample_result
#> 1:            0.079 2022-02-24 04:17:45        1 <ResampleResult[22]>
#> 2:            0.092 2022-02-24 04:17:45        2 <ResampleResult[22]>