Feature Selection via Exhaustive Search

FSelectorExhaustiveSearch class that implements an Exhaustive Search.

In order to support general termination criteria and parallelization, feature sets are evaluated in batches. The size of the feature sets is increased by 1 in each batch.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("exhaustive_search")
fs("exhaustive_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

Super class

mlr3fselect::FSelector -> FSelectorExhaustiveSearch

Methods

Public methods

FSelectorExhaustiveSearch$new()

FSelectorExhaustiveSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorExhaustiveSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorExhaustiveSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = trm("evals", n_evals = 5)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

fselector = fs("exhaustive_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:         TRUE       FALSE        FALSE       FALSE       0.06
#>  2:        FALSE        TRUE        FALSE       FALSE       0.06
#>  3:        FALSE       FALSE         TRUE       FALSE       0.26
#>  4:        FALSE       FALSE        FALSE        TRUE       0.48
#>  5:         TRUE        TRUE        FALSE       FALSE       0.06
#>  6:         TRUE       FALSE         TRUE       FALSE       0.06
#>  7:         TRUE       FALSE        FALSE        TRUE       0.06
#>  8:        FALSE        TRUE         TRUE       FALSE       0.06
#>  9:        FALSE        TRUE        FALSE        TRUE       0.06
#> 10:        FALSE       FALSE         TRUE        TRUE       0.26
#>                                    uhash           timestamp batch_nr
#>  1: e2ec4b0c-1538-4be0-b0c7-484d88664657 2021-05-06 04:33:22        1
#>  2: 0d864cd8-132e-4cbe-ba83-691fb858c15d 2021-05-06 04:33:22        1
#>  3: 4cfb1e29-8233-47d9-94df-49103ca56d64 2021-05-06 04:33:22        1
#>  4: e392bb11-edaf-4b5b-82bb-aec5760506e0 2021-05-06 04:33:22        1
#>  5: b6c279b2-9624-4251-a83f-06178744aecb 2021-05-06 04:33:23        2
#>  6: caad511c-6511-42d9-8ff4-55eda3e5a0be 2021-05-06 04:33:23        2
#>  7: b75b8650-c85a-4eaa-bca5-dc627c09ce29 2021-05-06 04:33:23        2
#>  8: 486a3cd1-a37f-494e-b8f8-a5c7a841191b 2021-05-06 04:33:23        2
#>  9: 5b301a21-de9c-4df5-bc72-cbcace2692fc 2021-05-06 04:33:23        2
#> 10: ee443c01-067d-4551-8ece-2edfdac7f51c 2021-05-06 04:33:23        2