Feature Selection via Random Search

FSelectorRandomSearch class that implements a simple Random Search.

In order to support general termination criteria and parallelization, we evaluate feature sets in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = trm("evals", n_evals = 5)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

fselector = fs("random_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width                 features
#> 1:         TRUE        TRUE        FALSE       FALSE Petal.Length,Petal.Width
#>     x_domain classif.ce
#> 1: <list[4]>       0.04

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width                 features
#> 1:         TRUE        TRUE        FALSE       FALSE Petal.Length,Petal.Width
#>     x_domain classif.ce
#> 1: <list[4]>       0.04

# Allows access of data.table of full path of all evaluations
instance$archive$data()# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:        FALSE       FALSE        FALSE        TRUE       0.46
#>  2:        FALSE       FALSE        FALSE        TRUE       0.46
#>  3:         TRUE       FALSE        FALSE       FALSE       0.08
#>  4:         TRUE        TRUE        FALSE       FALSE       0.04
#>  5:         TRUE       FALSE        FALSE        TRUE       0.08
#>  6:        FALSE        TRUE         TRUE       FALSE       0.04
#>  7:         TRUE        TRUE         TRUE        TRUE       0.04
#>  8:         TRUE        TRUE         TRUE        TRUE       0.04
#>  9:        FALSE        TRUE        FALSE        TRUE       0.04
#> 10:         TRUE        TRUE         TRUE        TRUE       0.04
#>                                    uhash  x_domain           timestamp batch_nr
#>  1: 8730c2cf-311d-44cb-89bd-f6965465e040 <list[4]> 2020-12-12 10:06:25        1
#>  2: 1ed1505a-d569-43c3-8a9e-b950f26b9c95 <list[4]> 2020-12-12 10:06:25        1
#>  3: a68a5388-8a2d-4d9c-945a-0862c52d87e7 <list[4]> 2020-12-12 10:06:25        1
#>  4: f36940b7-d9b9-400c-8880-37f85d193796 <list[4]> 2020-12-12 10:06:25        1
#>  5: bd064923-b077-433a-ac8a-7c5c72f26374 <list[4]> 2020-12-12 10:06:25        1
#>  6: 2115b63d-17dc-49c9-9a40-b7e0a2047232 <list[4]> 2020-12-12 10:06:25        1
#>  7: f6cc2062-876d-4ac8-8da4-3eebeca744b5 <list[4]> 2020-12-12 10:06:25        1
#>  8: b7ac29e4-5972-4abc-bfe6-d3f8c8ebc73c <list[4]> 2020-12-12 10:06:25        1
#>  9: cb2fb0d7-eb0c-4d91-b085-9c17b448acff <list[4]> 2020-12-12 10:06:25        1
#> 10: 2ab2a89b-ac8e-4892-99ea-bcec6c4e85d1 <list[4]> 2020-12-12 10:06:25        1