Feature Selection via Random Search

FSelectorRandomSearch class that implements a simple Random Search.

In order to support general termination criteria and parallelization, we evaluate feature sets in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

terminator = trm("evals", n_evals = 5)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

fselector = fs("random_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width                 features
#> 1:         TRUE        TRUE        FALSE       FALSE Petal.Length,Petal.Width
#>    classif.ce
#> 1:       0.04

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width                 features
#> 1:         TRUE        TRUE        FALSE       FALSE Petal.Length,Petal.Width
#>    classif.ce
#> 1:       0.04

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:        FALSE       FALSE        FALSE        TRUE       0.46
#>  2:        FALSE       FALSE        FALSE        TRUE       0.46
#>  3:         TRUE       FALSE        FALSE       FALSE       0.08
#>  4:         TRUE        TRUE        FALSE       FALSE       0.04
#>  5:         TRUE       FALSE        FALSE        TRUE       0.08
#>  6:        FALSE        TRUE         TRUE       FALSE       0.04
#>  7:         TRUE        TRUE         TRUE        TRUE       0.04
#>  8:         TRUE        TRUE         TRUE        TRUE       0.04
#>  9:        FALSE        TRUE        FALSE        TRUE       0.04
#> 10:         TRUE        TRUE         TRUE        TRUE       0.04
#>                                    uhash           timestamp batch_nr
#>  1: 26497491-1728-480e-a8ff-bcfb533fbfe0 2021-03-24 04:24:51        1
#>  2: 10521b7c-708c-4280-aee4-f82ab8b1b8da 2021-03-24 04:24:51        1
#>  3: 2f4260b1-fe44-4f31-a962-846045fc0bbd 2021-03-24 04:24:51        1
#>  4: 2ffe1829-df1a-47ba-a5c3-2ee7447faf47 2021-03-24 04:24:51        1
#>  5: 58ae6b13-a08c-4bd9-9e44-f0a611d36404 2021-03-24 04:24:51        1
#>  6: 404dc964-f222-460f-a123-2d4f277c21b6 2021-03-24 04:24:51        1
#>  7: 8da69d4d-5158-4110-ad56-d897691ea718 2021-03-24 04:24:51        1
#>  8: 992be96d-e636-4bb2-9f6e-084eeda021da 2021-03-24 04:24:51        1
#>  9: d7f8433a-ac58-44c6-8fb7-dcb3b10b1103 2021-03-24 04:24:51        1
#> 10: 163a83a6-f40a-4937-be04-a148afbb9633 2021-03-24 04:24:51        1