Feature Selection via Sequential Search with Shadow Variables

FSelectorShadowVariableSearch class that implements a sequential search with shadow variables.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi: 10.1155/2017/1421409 . Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi: 10.1198/016214506000000843 .

Note

FSelectorShadowVariableSearch terminates itself and should be used with TerminatorNone.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("none")
)

fselector = fs("shadow_variable_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>  1:         TRUE       FALSE        FALSE       FALSE                  FALSE
#>  2:        FALSE        TRUE        FALSE       FALSE                  FALSE
#>  3:        FALSE       FALSE         TRUE       FALSE                  FALSE
#>  4:        FALSE       FALSE        FALSE        TRUE                  FALSE
#>  5:        FALSE       FALSE        FALSE       FALSE                   TRUE
#>  6:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  7:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  8:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  9:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 10:         TRUE       FALSE         TRUE       FALSE                  FALSE
#> 11:         TRUE       FALSE        FALSE        TRUE                  FALSE
#> 12:         TRUE       FALSE        FALSE       FALSE                   TRUE
#> 13:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 14:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 15:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 16:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 17:         TRUE        TRUE        FALSE        TRUE                  FALSE
#> 18:         TRUE        TRUE        FALSE       FALSE                   TRUE
#> 19:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 20:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 21:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 22:         TRUE        TRUE         TRUE        TRUE                  FALSE
#> 23:         TRUE        TRUE         TRUE       FALSE                   TRUE
#> 24:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 25:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 26:         TRUE        TRUE         TRUE       FALSE                  FALSE
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>  1:                 FALSE                  FALSE                 FALSE
#>  2:                 FALSE                  FALSE                 FALSE
#>  3:                 FALSE                  FALSE                 FALSE
#>  4:                 FALSE                  FALSE                 FALSE
#>  5:                 FALSE                  FALSE                 FALSE
#>  6:                  TRUE                  FALSE                 FALSE
#>  7:                 FALSE                   TRUE                 FALSE
#>  8:                 FALSE                  FALSE                  TRUE
#>  9:                 FALSE                  FALSE                 FALSE
#> 10:                 FALSE                  FALSE                 FALSE
#> 11:                 FALSE                  FALSE                 FALSE
#> 12:                 FALSE                  FALSE                 FALSE
#> 13:                  TRUE                  FALSE                 FALSE
#> 14:                 FALSE                   TRUE                 FALSE
#> 15:                 FALSE                  FALSE                  TRUE
#> 16:                 FALSE                  FALSE                 FALSE
#> 17:                 FALSE                  FALSE                 FALSE
#> 18:                 FALSE                  FALSE                 FALSE
#> 19:                  TRUE                  FALSE                 FALSE
#> 20:                 FALSE                   TRUE                 FALSE
#> 21:                 FALSE                  FALSE                  TRUE
#> 22:                 FALSE                  FALSE                 FALSE
#> 23:                 FALSE                  FALSE                 FALSE
#> 24:                  TRUE                  FALSE                 FALSE
#> 25:                 FALSE                   TRUE                 FALSE
#> 26:                 FALSE                  FALSE                  TRUE
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>     classif.ce                                uhash           timestamp
#>  1:       0.06 54285561-82aa-4a1d-89b3-ab0b8c06f8f7 2021-06-06 05:23:25
#>  2:       0.06 3c3b5e80-936f-4167-826f-11bacecdaca8 2021-06-06 05:23:25
#>  3:       0.34 c3e4bddb-5f41-473e-a164-4a73ba1dc8e3 2021-06-06 05:23:25
#>  4:       0.50 2d7d2251-859a-4cc6-9f2c-0a18180ec4e6 2021-06-06 05:23:25
#>  5:       0.70 5fe2353c-e86e-4b96-ba97-329dc2df9f9d 2021-06-06 05:23:25
#>  6:       0.62 da85bbb0-f211-45aa-b759-215aa76ccc19 2021-06-06 05:23:25
#>  7:       0.76 c60f36ed-8354-4347-ac80-a942213a34d7 2021-06-06 05:23:25
#>  8:       0.64 b65a5409-7d3e-4fd8-ae89-51796581d01f 2021-06-06 05:23:25
#>  9:       0.06 501c4994-cbcc-4f03-985a-646814b3d6aa 2021-06-06 05:23:26
#> 10:       0.06 c67ded56-68b1-4bbb-a06b-0bb250dc0858 2021-06-06 05:23:26
#> 11:       0.06 41a20525-7c11-4067-9baf-ddd75b6aa95a 2021-06-06 05:23:26
#> 12:       0.06 29b2d340-0e0c-42c4-bbb3-e73ae07ba7fb 2021-06-06 05:23:26
#> 13:       0.06 44957d75-86d7-4868-8965-a5a6b3fb0aa6 2021-06-06 05:23:26
#> 14:       0.06 3878b6de-2618-45d7-a0c3-65deb0541163 2021-06-06 05:23:26
#> 15:       0.06 76d5a78f-a144-4c19-87a3-48efdd868ed8 2021-06-06 05:23:26
#> 16:       0.06 8edbea71-21cf-4bd0-a2f7-cf891758fe39 2021-06-06 05:23:28
#> 17:       0.06 cb4139f8-f26b-4cce-9403-371c59da4d83 2021-06-06 05:23:28
#> 18:       0.06 6aa10384-e7e8-4113-ba69-6af58dc00de4 2021-06-06 05:23:28
#> 19:       0.06 aaede2e8-aff4-41bb-8ca7-9fb7be18c405 2021-06-06 05:23:28
#> 20:       0.06 9f8268a2-50f8-4c02-a400-44d86a66f90c 2021-06-06 05:23:28
#> 21:       0.06 be914635-9dfb-4fc8-a09d-390197fc03fc 2021-06-06 05:23:28
#> 22:       0.06 71437a81-ee4e-41d4-b6ec-97ed515a79a0 2021-06-06 05:23:29
#> 23:       0.06 eb10da61-80e9-4409-ab8a-3e443c8a473e 2021-06-06 05:23:29
#> 24:       0.06 668dc8e9-7a76-4a27-8f5c-19fd8e786e76 2021-06-06 05:23:29
#> 25:       0.06 3dbe162d-355f-4964-abd9-569bf1b4c7de 2021-06-06 05:23:29
#> 26:       0.06 445a8216-242a-4183-a905-4cddc7cd9214 2021-06-06 05:23:29
#>     classif.ce                                uhash           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        2
#> 10:        2
#> 11:        2
#> 12:        2
#> 13:        2
#> 14:        2
#> 15:        2
#> 16:        3
#> 17:        3
#> 18:        3
#> 19:        3
#> 20:        3
#> 21:        3
#> 22:        4
#> 23:        4
#> 24:        4
#> 25:        4
#> 26:        4
#>     batch_nr