Feature Selection via Sequential Search with Shadow Variables

FSelectorShadowVariableSearch class that implements a sequential search with shadow variables.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi: 10.1155/2017/1421409 . Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi: 10.1198/016214506000000843 .

Note

FSelectorShadowVariableSearch terminates itself and should be used with TerminatorNone.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("none")
)

fselector = fs("shadow_variable_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>  1:         TRUE       FALSE        FALSE       FALSE                  FALSE
#>  2:        FALSE        TRUE        FALSE       FALSE                  FALSE
#>  3:        FALSE       FALSE         TRUE       FALSE                  FALSE
#>  4:        FALSE       FALSE        FALSE        TRUE                  FALSE
#>  5:        FALSE       FALSE        FALSE       FALSE                   TRUE
#>  6:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  7:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  8:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  9:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 10:         TRUE       FALSE         TRUE       FALSE                  FALSE
#> 11:         TRUE       FALSE        FALSE        TRUE                  FALSE
#> 12:         TRUE       FALSE        FALSE       FALSE                   TRUE
#> 13:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 14:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 15:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 16:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 17:         TRUE        TRUE        FALSE        TRUE                  FALSE
#> 18:         TRUE        TRUE        FALSE       FALSE                   TRUE
#> 19:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 20:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 21:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 22:         TRUE        TRUE         TRUE        TRUE                  FALSE
#> 23:         TRUE        TRUE         TRUE       FALSE                   TRUE
#> 24:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 25:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 26:         TRUE        TRUE         TRUE       FALSE                  FALSE
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>  1:                 FALSE                  FALSE                 FALSE
#>  2:                 FALSE                  FALSE                 FALSE
#>  3:                 FALSE                  FALSE                 FALSE
#>  4:                 FALSE                  FALSE                 FALSE
#>  5:                 FALSE                  FALSE                 FALSE
#>  6:                  TRUE                  FALSE                 FALSE
#>  7:                 FALSE                   TRUE                 FALSE
#>  8:                 FALSE                  FALSE                  TRUE
#>  9:                 FALSE                  FALSE                 FALSE
#> 10:                 FALSE                  FALSE                 FALSE
#> 11:                 FALSE                  FALSE                 FALSE
#> 12:                 FALSE                  FALSE                 FALSE
#> 13:                  TRUE                  FALSE                 FALSE
#> 14:                 FALSE                   TRUE                 FALSE
#> 15:                 FALSE                  FALSE                  TRUE
#> 16:                 FALSE                  FALSE                 FALSE
#> 17:                 FALSE                  FALSE                 FALSE
#> 18:                 FALSE                  FALSE                 FALSE
#> 19:                  TRUE                  FALSE                 FALSE
#> 20:                 FALSE                   TRUE                 FALSE
#> 21:                 FALSE                  FALSE                  TRUE
#> 22:                 FALSE                  FALSE                 FALSE
#> 23:                 FALSE                  FALSE                 FALSE
#> 24:                  TRUE                  FALSE                 FALSE
#> 25:                 FALSE                   TRUE                 FALSE
#> 26:                 FALSE                  FALSE                  TRUE
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>     classif.ce                                uhash           timestamp
#>  1:       0.06 ed427850-6ec3-44e8-81ba-464eeefbb5a6 2021-06-02 06:22:01
#>  2:       0.06 d1e8bd6d-2634-49d5-a222-44112ab0ac8a 2021-06-02 06:22:01
#>  3:       0.34 7ca83a1e-f010-4901-824e-185dd7afc5bd 2021-06-02 06:22:01
#>  4:       0.50 17ee9124-ba10-466d-8a17-e74df3a42461 2021-06-02 06:22:01
#>  5:       0.70 d9aa1284-4a7b-438e-bdd6-b1b6c38f8528 2021-06-02 06:22:01
#>  6:       0.62 d6e2ca1d-72eb-4dac-8bcf-10b67dcbd686 2021-06-02 06:22:01
#>  7:       0.76 cde751a3-f4cb-4025-b8e7-883b364fa1f9 2021-06-02 06:22:01
#>  8:       0.64 5252940f-fb61-40ab-a7f6-8c9d4650c1e1 2021-06-02 06:22:01
#>  9:       0.06 930308ca-8a26-43bd-9faa-3b8b40a23397 2021-06-02 06:22:02
#> 10:       0.06 31ad6182-4ec5-4cf4-af38-f1f8f6470e1a 2021-06-02 06:22:02
#> 11:       0.06 e6a60572-0047-4cad-b4ee-42310892b984 2021-06-02 06:22:02
#> 12:       0.06 7913a1d2-1c92-4977-9e49-b43ba39573e0 2021-06-02 06:22:02
#> 13:       0.06 5c5ba4f5-5a96-44f7-8032-18ab643ae8e5 2021-06-02 06:22:02
#> 14:       0.06 03ee914d-30fa-4b89-b817-a47841d7af8c 2021-06-02 06:22:02
#> 15:       0.06 5f185b2c-477f-4a2c-90b5-da88d7d90747 2021-06-02 06:22:02
#> 16:       0.06 a4aeaaa8-d518-4017-b431-baa9791ade81 2021-06-02 06:22:04
#> 17:       0.06 74a8fd90-0aca-45ff-a47b-b81aecde03bf 2021-06-02 06:22:04
#> 18:       0.06 25024e12-c0ef-49f1-8be8-df2c5349ce52 2021-06-02 06:22:04
#> 19:       0.06 9572c378-e0a8-4cd9-94bc-5f5d970ab3bb 2021-06-02 06:22:04
#> 20:       0.06 84e53fa8-9ccd-40a4-8c45-dbc5f52f9448 2021-06-02 06:22:04
#> 21:       0.06 da1d7648-aa0e-4b4e-beeb-bbbb5d7e70cd 2021-06-02 06:22:04
#> 22:       0.06 da5af2ad-fbd2-4f19-a2a5-81a473f24e5e 2021-06-02 06:22:05
#> 23:       0.06 0a17611d-aa78-4684-84c1-e03216ce5425 2021-06-02 06:22:05
#> 24:       0.06 463b8650-93c3-4a8b-8a0e-1680e32469d1 2021-06-02 06:22:05
#> 25:       0.06 950c0ac8-bb42-4c61-b38b-b592f7608b65 2021-06-02 06:22:05
#> 26:       0.06 f18b6b4d-aaa5-43f8-aec6-13de273be239 2021-06-02 06:22:05
#>     classif.ce                                uhash           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        2
#> 10:        2
#> 11:        2
#> 12:        2
#> 13:        2
#> 14:        2
#> 15:        2
#> 16:        3
#> 17:        3
#> 18:        3
#> 19:        3
#> 20:        3
#> 21:        3
#> 22:        4
#> 23:        4
#> 24:        4
#> 25:        4
#> 26:        4
#>     batch_nr