Feature Selection via Sequential Search with Shadow Variables

FSelectorShadowVariableSearch class that implements a sequential search with shadow variables.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi: 10.1155/2017/1421409 . Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi: 10.1198/016214506000000843 .

Note

FSelectorShadowVariableSearch terminates itself and should be used with TerminatorNone.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("none")
)

fselector = fs("shadow_variable_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>  1:         TRUE       FALSE        FALSE       FALSE                  FALSE
#>  2:        FALSE        TRUE        FALSE       FALSE                  FALSE
#>  3:        FALSE       FALSE         TRUE       FALSE                  FALSE
#>  4:        FALSE       FALSE        FALSE        TRUE                  FALSE
#>  5:        FALSE       FALSE        FALSE       FALSE                   TRUE
#>  6:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  7:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  8:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  9:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 10:         TRUE       FALSE         TRUE       FALSE                  FALSE
#> 11:         TRUE       FALSE        FALSE        TRUE                  FALSE
#> 12:         TRUE       FALSE        FALSE       FALSE                   TRUE
#> 13:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 14:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 15:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 16:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 17:         TRUE        TRUE        FALSE        TRUE                  FALSE
#> 18:         TRUE        TRUE        FALSE       FALSE                   TRUE
#> 19:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 20:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 21:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 22:         TRUE        TRUE         TRUE        TRUE                  FALSE
#> 23:         TRUE        TRUE         TRUE       FALSE                   TRUE
#> 24:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 25:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 26:         TRUE        TRUE         TRUE       FALSE                  FALSE
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>  1:                 FALSE                  FALSE                 FALSE
#>  2:                 FALSE                  FALSE                 FALSE
#>  3:                 FALSE                  FALSE                 FALSE
#>  4:                 FALSE                  FALSE                 FALSE
#>  5:                 FALSE                  FALSE                 FALSE
#>  6:                  TRUE                  FALSE                 FALSE
#>  7:                 FALSE                   TRUE                 FALSE
#>  8:                 FALSE                  FALSE                  TRUE
#>  9:                 FALSE                  FALSE                 FALSE
#> 10:                 FALSE                  FALSE                 FALSE
#> 11:                 FALSE                  FALSE                 FALSE
#> 12:                 FALSE                  FALSE                 FALSE
#> 13:                  TRUE                  FALSE                 FALSE
#> 14:                 FALSE                   TRUE                 FALSE
#> 15:                 FALSE                  FALSE                  TRUE
#> 16:                 FALSE                  FALSE                 FALSE
#> 17:                 FALSE                  FALSE                 FALSE
#> 18:                 FALSE                  FALSE                 FALSE
#> 19:                  TRUE                  FALSE                 FALSE
#> 20:                 FALSE                   TRUE                 FALSE
#> 21:                 FALSE                  FALSE                  TRUE
#> 22:                 FALSE                  FALSE                 FALSE
#> 23:                 FALSE                  FALSE                 FALSE
#> 24:                  TRUE                  FALSE                 FALSE
#> 25:                 FALSE                   TRUE                 FALSE
#> 26:                 FALSE                  FALSE                  TRUE
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>     classif.ce                                uhash           timestamp
#>  1:       0.06 ba41bb17-7a1c-479b-806d-364a9fc18ef1 2021-06-01 19:43:51
#>  2:       0.06 8bcc5b6b-1fb3-4421-b562-e05794478ad7 2021-06-01 19:43:51
#>  3:       0.34 e3fa4d43-8510-4054-bc40-716f83a7f9d1 2021-06-01 19:43:51
#>  4:       0.50 196bc327-79ff-4a1b-aa70-3eb4d2ee018c 2021-06-01 19:43:51
#>  5:       0.70 49dae6ae-e696-49a7-b5ed-4606ff55b49c 2021-06-01 19:43:51
#>  6:       0.62 30b3a1a6-9aaa-4c59-9a66-9c7b6c1285fb 2021-06-01 19:43:51
#>  7:       0.76 c03420a6-4f98-4910-bf86-453b71dcab6c 2021-06-01 19:43:51
#>  8:       0.64 75c7bb3d-8e43-475c-9b95-ad38b34730c3 2021-06-01 19:43:51
#>  9:       0.06 2af8afc8-f1c0-46ae-9d2b-0907f617495b 2021-06-01 19:43:52
#> 10:       0.06 2460037e-f06d-4ca4-8a0e-3829ace1b565 2021-06-01 19:43:52
#> 11:       0.06 b53ee4ff-f1df-442a-b115-b5dfe9cd1c75 2021-06-01 19:43:52
#> 12:       0.06 6e86e19d-d987-4cfe-bd08-39455c566698 2021-06-01 19:43:52
#> 13:       0.06 68fa0f59-85e6-43c0-a387-a8cd20aa4223 2021-06-01 19:43:52
#> 14:       0.06 3e5b1bbc-7ff1-428d-a962-1a09039062f5 2021-06-01 19:43:52
#> 15:       0.06 f116fd9f-e28c-47cc-a958-c53267971bd1 2021-06-01 19:43:52
#> 16:       0.06 1af14654-5e10-4faa-a889-9cc6a24f24dc 2021-06-01 19:43:54
#> 17:       0.06 8c03aa2d-caf7-4090-a76e-ec3b0bd7364d 2021-06-01 19:43:54
#> 18:       0.06 f0061ac1-6844-441d-bfda-2b4c2d4e8e99 2021-06-01 19:43:54
#> 19:       0.06 2d55390b-ff2a-48de-85d7-d6dfd0234e67 2021-06-01 19:43:54
#> 20:       0.06 97126a30-0dcd-4de2-8c3f-8332db0e5842 2021-06-01 19:43:54
#> 21:       0.06 7f0619da-0ff1-4745-9a56-c3742a6660fd 2021-06-01 19:43:54
#> 22:       0.06 379cc4f8-6854-452b-8c64-27c55a5ebfb5 2021-06-01 19:43:55
#> 23:       0.06 8e9c010e-bc6c-4504-9f42-acd43412912b 2021-06-01 19:43:55
#> 24:       0.06 a643c792-8733-414f-a2a1-9d9076018f0f 2021-06-01 19:43:55
#> 25:       0.06 d1d21c49-371a-400d-a88f-54d6dacd8bcf 2021-06-01 19:43:55
#> 26:       0.06 1f9ed7ef-7914-402d-b169-9945a8571e2b 2021-06-01 19:43:55
#>     classif.ce                                uhash           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        2
#> 10:        2
#> 11:        2
#> 12:        2
#> 13:        2
#> 14:        2
#> 15:        2
#> 16:        3
#> 17:        3
#> 18:        3
#> 19:        3
#> 20:        3
#> 21:        3
#> 22:        4
#> 23:        4
#> 24:        4
#> 25:        4
#> 26:        4
#>     batch_nr