Feature Selection via Sequential Search with Shadow Variables

FSelectorShadowVariableSearch class that implements a sequential search with shadow variables.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi: 10.1155/2017/1421409 . Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi: 10.1198/016214506000000843 .

Note

FSelectorShadowVariableSearch terminates itself and should be used with TerminatorNone.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("none")
)

fselector = fs("shadow_variable_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Returns best scoring evaluation
instance$result

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.06

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }

#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>  1:         TRUE       FALSE        FALSE       FALSE                  FALSE
#>  2:        FALSE        TRUE        FALSE       FALSE                  FALSE
#>  3:        FALSE       FALSE         TRUE       FALSE                  FALSE
#>  4:        FALSE       FALSE        FALSE        TRUE                  FALSE
#>  5:        FALSE       FALSE        FALSE       FALSE                   TRUE
#>  6:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  7:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  8:        FALSE       FALSE        FALSE       FALSE                  FALSE
#>  9:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 10:         TRUE       FALSE         TRUE       FALSE                  FALSE
#> 11:         TRUE       FALSE        FALSE        TRUE                  FALSE
#> 12:         TRUE       FALSE        FALSE       FALSE                   TRUE
#> 13:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 14:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 15:         TRUE       FALSE        FALSE       FALSE                  FALSE
#> 16:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 17:         TRUE        TRUE        FALSE        TRUE                  FALSE
#> 18:         TRUE        TRUE        FALSE       FALSE                   TRUE
#> 19:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 20:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 21:         TRUE        TRUE        FALSE       FALSE                  FALSE
#> 22:         TRUE        TRUE         TRUE        TRUE                  FALSE
#> 23:         TRUE        TRUE         TRUE       FALSE                   TRUE
#> 24:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 25:         TRUE        TRUE         TRUE       FALSE                  FALSE
#> 26:         TRUE        TRUE         TRUE       FALSE                  FALSE
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width permuted__Petal.Length
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>  1:                 FALSE                  FALSE                 FALSE
#>  2:                 FALSE                  FALSE                 FALSE
#>  3:                 FALSE                  FALSE                 FALSE
#>  4:                 FALSE                  FALSE                 FALSE
#>  5:                 FALSE                  FALSE                 FALSE
#>  6:                  TRUE                  FALSE                 FALSE
#>  7:                 FALSE                   TRUE                 FALSE
#>  8:                 FALSE                  FALSE                  TRUE
#>  9:                 FALSE                  FALSE                 FALSE
#> 10:                 FALSE                  FALSE                 FALSE
#> 11:                 FALSE                  FALSE                 FALSE
#> 12:                 FALSE                  FALSE                 FALSE
#> 13:                  TRUE                  FALSE                 FALSE
#> 14:                 FALSE                   TRUE                 FALSE
#> 15:                 FALSE                  FALSE                  TRUE
#> 16:                 FALSE                  FALSE                 FALSE
#> 17:                 FALSE                  FALSE                 FALSE
#> 18:                 FALSE                  FALSE                 FALSE
#> 19:                  TRUE                  FALSE                 FALSE
#> 20:                 FALSE                   TRUE                 FALSE
#> 21:                 FALSE                  FALSE                  TRUE
#> 22:                 FALSE                  FALSE                 FALSE
#> 23:                 FALSE                  FALSE                 FALSE
#> 24:                  TRUE                  FALSE                 FALSE
#> 25:                 FALSE                   TRUE                 FALSE
#> 26:                 FALSE                  FALSE                  TRUE
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>     classif.ce                                uhash           timestamp
#>  1:       0.06 cdd62a7b-1e82-41f5-906b-a578a67cd072 2021-06-04 05:48:40
#>  2:       0.06 75bdab1a-f6ed-4515-afba-2fc31122fd28 2021-06-04 05:48:40
#>  3:       0.34 48835a2a-ca57-495c-bb6e-d9368cffdee4 2021-06-04 05:48:40
#>  4:       0.50 f11e7a21-fdf8-40cd-80c6-2cf6fd16df65 2021-06-04 05:48:40
#>  5:       0.70 235921f4-d40d-40fd-8f00-bd3dfbf52caa 2021-06-04 05:48:40
#>  6:       0.62 536e257a-9f3f-4cf5-b705-4dd9d6e6cc55 2021-06-04 05:48:40
#>  7:       0.76 2d4ce202-2c1d-46a3-bb52-288cecfa94c3 2021-06-04 05:48:40
#>  8:       0.64 fd3d29ce-c119-4ba4-bf9b-3bdff90155a0 2021-06-04 05:48:40
#>  9:       0.06 51aa0459-81bd-44ab-b3ba-adf0086675a2 2021-06-04 05:48:41
#> 10:       0.06 d2767e35-2905-4d9c-9d39-2dccf2232568 2021-06-04 05:48:41
#> 11:       0.06 59904743-3859-469e-99db-0276af735aa0 2021-06-04 05:48:41
#> 12:       0.06 35f23602-af42-49df-ba38-a959c5bf43c1 2021-06-04 05:48:41
#> 13:       0.06 4af33191-62b4-4587-9f55-3729572728fa 2021-06-04 05:48:41
#> 14:       0.06 3f7b7f57-76d3-4cb8-9748-6a5565e35560 2021-06-04 05:48:41
#> 15:       0.06 75c2b2d7-c10a-4437-b428-324a3fae71e2 2021-06-04 05:48:41
#> 16:       0.06 eebbd524-8e1d-43ee-a69c-efe62de6a411 2021-06-04 05:48:42
#> 17:       0.06 330491e5-dc56-4698-a026-d9c30d59ee3c 2021-06-04 05:48:42
#> 18:       0.06 e7b5c64e-fea3-4d12-9d71-efdf6881252a 2021-06-04 05:48:42
#> 19:       0.06 2570770b-0f6b-431f-910c-5890d2d8033d 2021-06-04 05:48:42
#> 20:       0.06 47ae248a-1a39-431d-b6f5-a359aab250ac 2021-06-04 05:48:42
#> 21:       0.06 92ff10be-2f24-4534-9f12-3429e46e9ff3 2021-06-04 05:48:42
#> 22:       0.06 433b3a5b-900c-48a0-83d4-81fab87708d6 2021-06-04 05:48:43
#> 23:       0.06 795f9c36-d180-4478-ac0c-052f43dac917 2021-06-04 05:48:43
#> 24:       0.06 7ebd42fc-0dfe-47f4-adc0-b3eb8adc973a 2021-06-04 05:48:43
#> 25:       0.06 26a1007d-a49f-40a6-a116-4671320851fd 2021-06-04 05:48:43
#> 26:       0.06 b9dbd0c1-a83b-470e-ae8f-f247c061bbb2 2021-06-04 05:48:43
#>     classif.ce                                uhash           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        2
#> 10:        2
#> 11:        2
#> 12:        2
#> 13:        2
#> 14:        2
#> 15:        2
#> 16:        3
#> 17:        3
#> 18:        3
#> 19:        3
#> 20:        3
#> 21:        3
#> 22:        4
#> 23:        4
#> 24:        4
#> 25:        4
#> 26:        4
#>     batch_nr