Feature Selection via Sequential Search with Shadow Variables

FSelectorShadowVariableSearch class that implements a sequential search with shadow variables.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi: 10.1155/2017/1421409 .

Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi: 10.1198/016214506000000843 .

Note

FSelectorShadowVariableSearch terminates itself and should be used with TerminatorNone.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("none")
)

fselector = fs("shadow_variable_search")
# \donttest{
# Modifies the instance by reference
fselector$optimize(instance)
#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.08

# Returns best scoring evaluation
instance$result
#>    Petal.Length Petal.Width Sepal.Length Sepal.Width     features classif.ce
#> 1:         TRUE       FALSE        FALSE       FALSE Petal.Length       0.08

# Allows access of data.table of full path of all evaluations
as.data.table(instance$archive)# }
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>  1:         TRUE       FALSE        FALSE       FALSE       0.08
#>  2:        FALSE        TRUE        FALSE       FALSE       0.08
#>  3:        FALSE       FALSE         TRUE       FALSE       0.36
#>  4:        FALSE       FALSE        FALSE        TRUE       0.54
#>  5:        FALSE       FALSE        FALSE       FALSE       0.78
#>  6:        FALSE       FALSE        FALSE       FALSE       0.72
#>  7:        FALSE       FALSE        FALSE       FALSE       0.70
#>  8:        FALSE       FALSE        FALSE       FALSE       0.78
#>  9:         TRUE        TRUE        FALSE       FALSE       0.08
#> 10:         TRUE       FALSE         TRUE       FALSE       0.08
#> 11:         TRUE       FALSE        FALSE        TRUE       0.08
#> 12:         TRUE       FALSE        FALSE       FALSE       0.08
#> 13:         TRUE       FALSE        FALSE       FALSE       0.08
#> 14:         TRUE       FALSE        FALSE       FALSE       0.08
#> 15:         TRUE       FALSE        FALSE       FALSE       0.08
#> 16:         TRUE        TRUE         TRUE       FALSE       0.08
#> 17:         TRUE        TRUE        FALSE        TRUE       0.08
#> 18:         TRUE        TRUE        FALSE       FALSE       0.08
#> 19:         TRUE        TRUE        FALSE       FALSE       0.08
#> 20:         TRUE        TRUE        FALSE       FALSE       0.08
#> 21:         TRUE        TRUE        FALSE       FALSE       0.08
#> 22:         TRUE        TRUE         TRUE        TRUE       0.08
#> 23:         TRUE        TRUE         TRUE       FALSE       0.08
#> 24:         TRUE        TRUE         TRUE       FALSE       0.08
#> 25:         TRUE        TRUE         TRUE       FALSE       0.08
#> 26:         TRUE        TRUE         TRUE       FALSE       0.08
#>     Petal.Length Petal.Width Sepal.Length Sepal.Width classif.ce
#>     runtime_learners           timestamp batch_nr permuted__Petal.Length
#>  1:            0.083 2021-11-23 04:17:28        1                  FALSE
#>  2:            0.089 2021-11-23 04:17:28        1                  FALSE
#>  3:            0.093 2021-11-23 04:17:28        1                  FALSE
#>  4:            0.081 2021-11-23 04:17:28        1                  FALSE
#>  5:            0.079 2021-11-23 04:17:28        1                   TRUE
#>  6:            0.091 2021-11-23 04:17:28        1                  FALSE
#>  7:            0.093 2021-11-23 04:17:28        1                  FALSE
#>  8:            0.091 2021-11-23 04:17:28        1                  FALSE
#>  9:            0.100 2021-11-23 04:17:29        2                  FALSE
#> 10:            0.083 2021-11-23 04:17:29        2                  FALSE
#> 11:            0.080 2021-11-23 04:17:29        2                  FALSE
#> 12:            0.089 2021-11-23 04:17:29        2                   TRUE
#> 13:            0.098 2021-11-23 04:17:29        2                  FALSE
#> 14:            0.084 2021-11-23 04:17:29        2                  FALSE
#> 15:            0.095 2021-11-23 04:17:29        2                  FALSE
#> 16:            0.092 2021-11-23 04:17:31        3                  FALSE
#> 17:            0.092 2021-11-23 04:17:31        3                  FALSE
#> 18:            0.100 2021-11-23 04:17:31        3                   TRUE
#> 19:            0.086 2021-11-23 04:17:31        3                  FALSE
#> 20:            0.097 2021-11-23 04:17:31        3                  FALSE
#> 21:            0.088 2021-11-23 04:17:31        3                  FALSE
#> 22:            0.100 2021-11-23 04:17:32        4                  FALSE
#> 23:            0.104 2021-11-23 04:17:32        4                   TRUE
#> 24:            0.093 2021-11-23 04:17:32        4                  FALSE
#> 25:            0.088 2021-11-23 04:17:32        4                  FALSE
#> 26:            0.104 2021-11-23 04:17:32        4                  FALSE
#>     runtime_learners           timestamp batch_nr permuted__Petal.Length
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>  1:                 FALSE                  FALSE                 FALSE
#>  2:                 FALSE                  FALSE                 FALSE
#>  3:                 FALSE                  FALSE                 FALSE
#>  4:                 FALSE                  FALSE                 FALSE
#>  5:                 FALSE                  FALSE                 FALSE
#>  6:                  TRUE                  FALSE                 FALSE
#>  7:                 FALSE                   TRUE                 FALSE
#>  8:                 FALSE                  FALSE                  TRUE
#>  9:                 FALSE                  FALSE                 FALSE
#> 10:                 FALSE                  FALSE                 FALSE
#> 11:                 FALSE                  FALSE                 FALSE
#> 12:                 FALSE                  FALSE                 FALSE
#> 13:                  TRUE                  FALSE                 FALSE
#> 14:                 FALSE                   TRUE                 FALSE
#> 15:                 FALSE                  FALSE                  TRUE
#> 16:                 FALSE                  FALSE                 FALSE
#> 17:                 FALSE                  FALSE                 FALSE
#> 18:                 FALSE                  FALSE                 FALSE
#> 19:                  TRUE                  FALSE                 FALSE
#> 20:                 FALSE                   TRUE                 FALSE
#> 21:                 FALSE                  FALSE                  TRUE
#> 22:                 FALSE                  FALSE                 FALSE
#> 23:                 FALSE                  FALSE                 FALSE
#> 24:                  TRUE                  FALSE                 FALSE
#> 25:                 FALSE                   TRUE                 FALSE
#> 26:                 FALSE                  FALSE                  TRUE
#>     permuted__Petal.Width permuted__Sepal.Length permuted__Sepal.Width
#>          resample_result
#>  1: <ResampleResult[22]>
#>  2: <ResampleResult[22]>
#>  3: <ResampleResult[22]>
#>  4: <ResampleResult[22]>
#>  5: <ResampleResult[22]>
#>  6: <ResampleResult[22]>
#>  7: <ResampleResult[22]>
#>  8: <ResampleResult[22]>
#>  9: <ResampleResult[22]>
#> 10: <ResampleResult[22]>
#> 11: <ResampleResult[22]>
#> 12: <ResampleResult[22]>
#> 13: <ResampleResult[22]>
#> 14: <ResampleResult[22]>
#> 15: <ResampleResult[22]>
#> 16: <ResampleResult[22]>
#> 17: <ResampleResult[22]>
#> 18: <ResampleResult[22]>
#> 19: <ResampleResult[22]>
#> 20: <ResampleResult[22]>
#> 21: <ResampleResult[22]>
#> 22: <ResampleResult[22]>
#> 23: <ResampleResult[22]>
#> 24: <ResampleResult[22]>
#> 25: <ResampleResult[22]>
#> 26: <ResampleResult[22]>
#>          resample_result