Class for Feature Selection Objective

Stores the objective function that estimates the performance of feature subsets. This class is usually constructed internally by by the FSelectInstanceSingleCrit / FSelectInstanceMultiCrit.

Super class

bbotk::Objective -> ObjectiveFSelect

Public fields

task

(mlr3::Task).

learner

(mlr3::Learner).

resampling

(mlr3::Resampling).

measures

(list of mlr3::Measure).

store_models

(logical(1)).

store_benchmark_result

(logical(1)).

archive

(ArchiveFSelect).

callbacks

(List of CallbackFSelects).

Methods

Public methods

ObjectiveFSelect$new()

ObjectiveFSelect$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$eval_dt()

bbotk::Objective$eval_many()

bbotk::Objective$format()

bbotk::Objective$print()

Method new()

Creates a new instance of this R6 class.

Usage

ObjectiveFSelect$new(
  task,
  learner,
  resampling,
  measures,
  check_values = TRUE,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  archive = NULL,
  callbacks = list()
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner)
Learner to optimize the feature subset for.

resampling

(mlr3::Resampling)
Resampling that is used to evaluated the performance of the feature subsets. Uninstantiated resamplings are instantiated during construction so that all feature subsets are evaluated on the same data splits. Already instantiated resamplings are kept unchanged.

measures

(list of mlr3::Measure)
Measures to optimize. If NULL, mlr3's default measure is used.

check_values

(logical(1))
Check the parameters before the evaluation and the results for validity?

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1)). Store models in benchmark result?

archive

(ArchiveFSelect)
Reference to the archive of FSelectInstanceSingleCrit | FSelectInstanceMultiCrit. If NULL (default), benchmark result and models cannot be stored.

callbacks

(list of CallbackFSelect)
List of callbacks.

Method clone()

The objects of this class are cloneable with this method.

Usage

ObjectiveFSelect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.