Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>     age glucose insulin mass pedigree pregnant pressure triceps
#> 1: TRUE    TRUE    TRUE TRUE    FALSE     TRUE    FALSE   FALSE
#>                             features classif.ce
#> 1: age,glucose,insulin,mass,pregnant   0.203125

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>   2:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3085938
#>   3: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3554688
#>   4:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2617188
#>   5: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3007812
#>   6: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3554688
#>   7:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3085938
#>   8:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2187500
#>   9:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2265625
#>  10:  TRUE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2890625
#>  11: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.2578125
#>  12: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2109375
#>  13: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  14: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2500000
#>  15:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  16: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3828125
#>  17:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>  18: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2343750
#>  19:  TRUE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2500000
#>  20:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3242188
#>  21:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>  22: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2578125
#>  23:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.2695312
#>  24:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2109375
#>  25:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3242188
#>  26:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE    FALSE   FALSE  0.2812500
#>  27:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2500000
#>  28:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2929688
#>  29:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2890625
#>  30:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3085938
#>  31:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.2343750
#>  32: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3554688
#>  33: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  34: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2265625
#>  35:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2617188
#>  36:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE    TRUE  0.2968750
#>  37:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2500000
#>  38:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  39: FALSE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3125000
#>  40:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2109375
#>  41: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2343750
#>  42: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2734375
#>  43:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.2656250
#>  44:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  45: FALSE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE    TRUE  0.3125000
#>  46:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2617188
#>  47:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  48:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2500000
#>  49: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE    TRUE  0.3515625
#>  50:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  51:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  52: FALSE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.3046875
#>  53:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  54:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2265625
#>  55:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2500000
#>  56:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2031250
#>  57:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2265625
#>  58: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2304688
#>  59:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  60:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE    FALSE    TRUE  0.2578125
#>  61:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2187500
#>  62:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>  63: FALSE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.3515625
#>  64:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2226562
#>  65: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2382812
#>  66: FALSE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE    TRUE  0.3125000
#>  67:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2500000
#>  68:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE     TRUE   FALSE  0.3125000
#>  69: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2109375
#>  70: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2382812
#>  71:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  72: FALSE   FALSE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.2929688
#>  73:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3007812
#>  74: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2578125
#>  75:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2851562
#>  76:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  77:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  78:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3242188
#>  79: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3164062
#>  80: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3671875
#>  81: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2460938
#>  82:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2109375
#>  83: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2617188
#>  84: FALSE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3046875
#>  85:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  86: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3164062
#>  87:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.2812500
#>  88:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE     TRUE    TRUE  0.3281250
#>  89: FALSE   FALSE    TRUE FALSE     TRUE    FALSE     TRUE   FALSE  0.3750000
#>  90: FALSE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.2734375
#>  91:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2617188
#>  92:  TRUE   FALSE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.3398438
#>  93:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE     TRUE    TRUE  0.3164062
#>  94:  TRUE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2382812
#>  95: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3320312
#>  96:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2968750
#>  97:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2265625
#>  98:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  99:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2265625
#> 100: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2578125
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.105 2022-05-02 04:36:52        1 <ResampleResult[22]>
#>   2:            0.096 2022-05-02 04:36:52        2 <ResampleResult[22]>
#>   3:            0.092 2022-05-02 04:36:53        3 <ResampleResult[22]>
#>   4:            0.086 2022-05-02 04:36:53        4 <ResampleResult[22]>
#>   5:            0.084 2022-05-02 04:36:53        5 <ResampleResult[22]>
#>   6:            0.086 2022-05-02 04:36:54        6 <ResampleResult[22]>
#>   7:            0.084 2022-05-02 04:36:54        7 <ResampleResult[22]>
#>   8:            0.088 2022-05-02 04:36:54        8 <ResampleResult[22]>
#>   9:            0.090 2022-05-02 04:36:55        9 <ResampleResult[22]>
#>  10:            0.089 2022-05-02 04:36:55       10 <ResampleResult[22]>
#>  11:            0.090 2022-05-02 04:36:55       11 <ResampleResult[22]>
#>  12:            0.087 2022-05-02 04:36:55       12 <ResampleResult[22]>
#>  13:            0.091 2022-05-02 04:36:56       13 <ResampleResult[22]>
#>  14:            0.092 2022-05-02 04:36:56       14 <ResampleResult[22]>
#>  15:            0.097 2022-05-02 04:36:56       15 <ResampleResult[22]>
#>  16:            0.098 2022-05-02 04:36:57       16 <ResampleResult[22]>
#>  17:            0.117 2022-05-02 04:36:57       17 <ResampleResult[22]>
#>  18:            0.128 2022-05-02 04:36:57       18 <ResampleResult[22]>
#>  19:            0.110 2022-05-02 04:36:58       19 <ResampleResult[22]>
#>  20:            0.119 2022-05-02 04:36:58       20 <ResampleResult[22]>
#>  21:            0.110 2022-05-02 04:36:58       21 <ResampleResult[22]>
#>  22:            0.114 2022-05-02 04:36:59       22 <ResampleResult[22]>
#>  23:            0.116 2022-05-02 04:36:59       23 <ResampleResult[22]>
#>  24:            0.110 2022-05-02 04:36:59       24 <ResampleResult[22]>
#>  25:            0.107 2022-05-02 04:37:00       25 <ResampleResult[22]>
#>  26:            0.108 2022-05-02 04:37:00       26 <ResampleResult[22]>
#>  27:            0.109 2022-05-02 04:37:00       27 <ResampleResult[22]>
#>  28:            0.104 2022-05-02 04:37:00       28 <ResampleResult[22]>
#>  29:            0.097 2022-05-02 04:37:01       29 <ResampleResult[22]>
#>  30:            0.095 2022-05-02 04:37:01       30 <ResampleResult[22]>
#>  31:            0.088 2022-05-02 04:37:02       31 <ResampleResult[22]>
#>  32:            0.079 2022-05-02 04:37:02       32 <ResampleResult[22]>
#>  33:            0.085 2022-05-02 04:37:02       33 <ResampleResult[22]>
#>  34:            0.085 2022-05-02 04:37:02       34 <ResampleResult[22]>
#>  35:            0.092 2022-05-02 04:37:03       35 <ResampleResult[22]>
#>  36:            0.081 2022-05-02 04:37:03       36 <ResampleResult[22]>
#>  37:            0.077 2022-05-02 04:37:03       37 <ResampleResult[22]>
#>  38:            0.079 2022-05-02 04:37:03       38 <ResampleResult[22]>
#>  39:            0.084 2022-05-02 04:37:04       39 <ResampleResult[22]>
#>  40:            0.093 2022-05-02 04:37:04       40 <ResampleResult[22]>
#>  41:            0.094 2022-05-02 04:37:04       41 <ResampleResult[22]>
#>  42:            0.089 2022-05-02 04:37:04       42 <ResampleResult[22]>
#>  43:            0.091 2022-05-02 04:37:05       43 <ResampleResult[22]>
#>  44:            0.087 2022-05-02 04:37:05       44 <ResampleResult[22]>
#>  45:            0.073 2022-05-02 04:37:05       45 <ResampleResult[22]>
#>  46:            0.087 2022-05-02 04:37:05       46 <ResampleResult[22]>
#>  47:            0.090 2022-05-02 04:37:06       47 <ResampleResult[22]>
#>  48:            0.087 2022-05-02 04:37:06       48 <ResampleResult[22]>
#>  49:            0.088 2022-05-02 04:37:06       49 <ResampleResult[22]>
#>  50:            0.089 2022-05-02 04:37:06       50 <ResampleResult[22]>
#>  51:            0.097 2022-05-02 04:37:07       51 <ResampleResult[22]>
#>  52:            0.093 2022-05-02 04:37:07       52 <ResampleResult[22]>
#>  53:            0.080 2022-05-02 04:37:07       53 <ResampleResult[22]>
#>  54:            0.087 2022-05-02 04:37:07       54 <ResampleResult[22]>
#>  55:            0.087 2022-05-02 04:37:08       55 <ResampleResult[22]>
#>  56:            0.084 2022-05-02 04:37:08       56 <ResampleResult[22]>
#>  57:            0.109 2022-05-02 04:37:08       57 <ResampleResult[22]>
#>  58:            0.076 2022-05-02 04:37:09       58 <ResampleResult[22]>
#>  59:            0.095 2022-05-02 04:37:09       59 <ResampleResult[22]>
#>  60:            0.093 2022-05-02 04:37:09       60 <ResampleResult[22]>
#>  61:            0.079 2022-05-02 04:37:09       61 <ResampleResult[22]>
#>  62:            0.095 2022-05-02 04:37:10       62 <ResampleResult[22]>
#>  63:            0.098 2022-05-02 04:37:10       63 <ResampleResult[22]>
#>  64:            0.091 2022-05-02 04:37:10       64 <ResampleResult[22]>
#>  65:            0.089 2022-05-02 04:37:10       65 <ResampleResult[22]>
#>  66:            0.085 2022-05-02 04:37:11       66 <ResampleResult[22]>
#>  67:            0.079 2022-05-02 04:37:11       67 <ResampleResult[22]>
#>  68:            0.089 2022-05-02 04:37:11       68 <ResampleResult[22]>
#>  69:            0.091 2022-05-02 04:37:12       69 <ResampleResult[22]>
#>  70:            0.099 2022-05-02 04:37:12       70 <ResampleResult[22]>
#>  71:            0.097 2022-05-02 04:37:12       71 <ResampleResult[22]>
#>  72:            0.090 2022-05-02 04:37:12       72 <ResampleResult[22]>
#>  73:            0.086 2022-05-02 04:37:13       73 <ResampleResult[22]>
#>  74:            0.086 2022-05-02 04:37:13       74 <ResampleResult[22]>
#>  75:            0.086 2022-05-02 04:37:13       75 <ResampleResult[22]>
#>  76:            0.085 2022-05-02 04:37:13       76 <ResampleResult[22]>
#>  77:            0.097 2022-05-02 04:37:14       77 <ResampleResult[22]>
#>  78:            0.078 2022-05-02 04:37:14       78 <ResampleResult[22]>
#>  79:            0.094 2022-05-02 04:37:14       79 <ResampleResult[22]>
#>  80:            0.146 2022-05-02 04:37:15       80 <ResampleResult[22]>
#>  81:            0.104 2022-05-02 04:37:15       81 <ResampleResult[22]>
#>  82:            0.090 2022-05-02 04:37:15       82 <ResampleResult[22]>
#>  83:            0.086 2022-05-02 04:37:15       83 <ResampleResult[22]>
#>  84:            0.083 2022-05-02 04:37:16       84 <ResampleResult[22]>
#>  85:            0.095 2022-05-02 04:37:16       85 <ResampleResult[22]>
#>  86:            0.104 2022-05-02 04:37:16       86 <ResampleResult[22]>
#>  87:            0.097 2022-05-02 04:37:17       87 <ResampleResult[22]>
#>  88:            0.082 2022-05-02 04:37:17       88 <ResampleResult[22]>
#>  89:            0.084 2022-05-02 04:37:17       89 <ResampleResult[22]>
#>  90:            0.087 2022-05-02 04:37:17       90 <ResampleResult[22]>
#>  91:            0.087 2022-05-02 04:37:18       91 <ResampleResult[22]>
#>  92:            0.097 2022-05-02 04:37:18       92 <ResampleResult[22]>
#>  93:            0.092 2022-05-02 04:37:18       93 <ResampleResult[22]>
#>  94:            0.099 2022-05-02 04:37:18       94 <ResampleResult[22]>
#>  95:            0.097 2022-05-02 04:37:19       95 <ResampleResult[22]>
#>  96:            0.078 2022-05-02 04:37:19       96 <ResampleResult[22]>
#>  97:            0.087 2022-05-02 04:37:19       97 <ResampleResult[22]>
#>  98:            0.092 2022-05-02 04:37:19       98 <ResampleResult[22]>
#>  99:            0.094 2022-05-02 04:37:20       99 <ResampleResult[22]>
#> 100:            0.085 2022-05-02 04:37:20      100 <ResampleResult[22]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }