Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>      age glucose insulin  mass pedigree pregnant pressure triceps
#> 1: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE
#>            features classif.ce
#> 1: glucose,pedigree  0.2226562

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2890625
#>   2:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2890625
#>   3: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2343750
#>   4:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2460938
#>   5: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>   6: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3789062
#>   7:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>   8: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2617188
#>   9: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3945312
#>  10: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3007812
#>  11: FALSE   FALSE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.3242188
#>  12:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  13:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.3437500
#>  14:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3554688
#>  15:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2460938
#>  16: FALSE   FALSE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.3515625
#>  17:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE    FALSE    TRUE  0.2773438
#>  18: FALSE   FALSE    TRUE FALSE     TRUE    FALSE     TRUE   FALSE  0.3085938
#>  19:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>  20: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3085938
#>  21: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE   FALSE  0.3203125
#>  22: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2265625
#>  23:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2500000
#>  24: FALSE    TRUE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.2890625
#>  25: FALSE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3437500
#>  26:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE   FALSE  0.2539062
#>  27:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2382812
#>  28:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2578125
#>  29: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2539062
#>  30:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2421875
#>  31: FALSE    TRUE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3164062
#>  32: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2539062
#>  33:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.2578125
#>  34: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>  35:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  36: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE    TRUE  0.3710938
#>  37: FALSE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.3046875
#>  38: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3828125
#>  39:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3789062
#>  40:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  41:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  42: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3281250
#>  43:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2460938
#>  44: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2617188
#>  45: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2539062
#>  46:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.3945312
#>  47: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2343750
#>  48: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>  49: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.3320312
#>  50:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2500000
#>  51:  TRUE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.3320312
#>  52:  TRUE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.3046875
#>  53:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  54:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2500000
#>  55: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>  56:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3320312
#>  57:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  58:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2460938
#>  59:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2500000
#>  60: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3007812
#>  61: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2617188
#>  62:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2382812
#>  63:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2304688
#>  64: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2617188
#>  65: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2617188
#>  66: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE   FALSE  0.3203125
#>  67:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  68: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3007812
#>  69:  TRUE   FALSE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.3750000
#>  70:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.3125000
#>  71:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>  72: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.2539062
#>  73:  TRUE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2539062
#>  74: FALSE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2773438
#>  75: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2226562
#>  76:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  77:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE     TRUE   FALSE  0.3515625
#>  78:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  79:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2851562
#>  80: FALSE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE   FALSE  0.3203125
#>  81:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3085938
#>  82: FALSE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2773438
#>  83: FALSE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3242188
#>  84:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  85:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  86: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.3789062
#>  87:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE    TRUE  0.2343750
#>  88:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2460938
#>  89: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3945312
#>  90:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2421875
#>  91:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.3164062
#>  92:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2500000
#>  93: FALSE   FALSE   FALSE FALSE     TRUE    FALSE     TRUE   FALSE  0.3242188
#>  94: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE    TRUE  0.3710938
#>  95:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2851562
#>  96:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2500000
#>  97: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3085938
#>  98: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2617188
#>  99:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#> 100:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.070 2022-04-16 04:15:16        1 <ResampleResult[22]>
#>   2:            0.070 2022-04-16 04:15:16        2 <ResampleResult[22]>
#>   3:            0.072 2022-04-16 04:15:16        3 <ResampleResult[22]>
#>   4:            0.073 2022-04-16 04:15:17        4 <ResampleResult[22]>
#>   5:            0.069 2022-04-16 04:15:17        5 <ResampleResult[22]>
#>   6:            0.068 2022-04-16 04:15:17        6 <ResampleResult[22]>
#>   7:            0.070 2022-04-16 04:15:17        7 <ResampleResult[22]>
#>   8:            0.070 2022-04-16 04:15:18        8 <ResampleResult[22]>
#>   9:            0.070 2022-04-16 04:15:18        9 <ResampleResult[22]>
#>  10:            0.067 2022-04-16 04:15:18       10 <ResampleResult[22]>
#>  11:            0.070 2022-04-16 04:15:18       11 <ResampleResult[22]>
#>  12:            0.079 2022-04-16 04:15:19       12 <ResampleResult[22]>
#>  13:            0.072 2022-04-16 04:15:19       13 <ResampleResult[22]>
#>  14:            0.068 2022-04-16 04:15:19       14 <ResampleResult[22]>
#>  15:            0.075 2022-04-16 04:15:19       15 <ResampleResult[22]>
#>  16:            0.081 2022-04-16 04:15:20       16 <ResampleResult[22]>
#>  17:            0.086 2022-04-16 04:15:20       17 <ResampleResult[22]>
#>  18:            0.096 2022-04-16 04:15:20       18 <ResampleResult[22]>
#>  19:            0.106 2022-04-16 04:15:20       19 <ResampleResult[22]>
#>  20:            0.101 2022-04-16 04:15:21       20 <ResampleResult[22]>
#>  21:            0.096 2022-04-16 04:15:21       21 <ResampleResult[22]>
#>  22:            0.092 2022-04-16 04:15:21       22 <ResampleResult[22]>
#>  23:            0.088 2022-04-16 04:15:21       23 <ResampleResult[22]>
#>  24:            0.082 2022-04-16 04:15:22       24 <ResampleResult[22]>
#>  25:            0.071 2022-04-16 04:15:22       25 <ResampleResult[22]>
#>  26:            0.072 2022-04-16 04:15:22       26 <ResampleResult[22]>
#>  27:            0.075 2022-04-16 04:15:22       27 <ResampleResult[22]>
#>  28:            0.074 2022-04-16 04:15:23       28 <ResampleResult[22]>
#>  29:            0.077 2022-04-16 04:15:23       29 <ResampleResult[22]>
#>  30:            0.072 2022-04-16 04:15:23       30 <ResampleResult[22]>
#>  31:            0.071 2022-04-16 04:15:23       31 <ResampleResult[22]>
#>  32:            0.073 2022-04-16 04:15:24       32 <ResampleResult[22]>
#>  33:            0.072 2022-04-16 04:15:24       33 <ResampleResult[22]>
#>  34:            0.073 2022-04-16 04:15:24       34 <ResampleResult[22]>
#>  35:            0.076 2022-04-16 04:15:24       35 <ResampleResult[22]>
#>  36:            0.287 2022-04-16 04:15:25       36 <ResampleResult[22]>
#>  37:            0.065 2022-04-16 04:15:25       37 <ResampleResult[22]>
#>  38:            0.064 2022-04-16 04:15:25       38 <ResampleResult[22]>
#>  39:            0.064 2022-04-16 04:15:25       39 <ResampleResult[22]>
#>  40:            0.073 2022-04-16 04:15:26       40 <ResampleResult[22]>
#>  41:            0.075 2022-04-16 04:15:26       41 <ResampleResult[22]>
#>  42:            0.065 2022-04-16 04:15:26       42 <ResampleResult[22]>
#>  43:            0.066 2022-04-16 04:15:26       43 <ResampleResult[22]>
#>  44:            0.065 2022-04-16 04:15:26       44 <ResampleResult[22]>
#>  45:            0.065 2022-04-16 04:15:27       45 <ResampleResult[22]>
#>  46:            0.073 2022-04-16 04:15:27       46 <ResampleResult[22]>
#>  47:            0.074 2022-04-16 04:15:27       47 <ResampleResult[22]>
#>  48:            0.065 2022-04-16 04:15:27       48 <ResampleResult[22]>
#>  49:            0.066 2022-04-16 04:15:27       49 <ResampleResult[22]>
#>  50:            0.066 2022-04-16 04:15:28       50 <ResampleResult[22]>
#>  51:            0.065 2022-04-16 04:15:28       51 <ResampleResult[22]>
#>  52:            0.064 2022-04-16 04:15:28       52 <ResampleResult[22]>
#>  53:            0.076 2022-04-16 04:15:28       53 <ResampleResult[22]>
#>  54:            0.075 2022-04-16 04:15:28       54 <ResampleResult[22]>
#>  55:            0.074 2022-04-16 04:15:29       55 <ResampleResult[22]>
#>  56:            0.067 2022-04-16 04:15:29       56 <ResampleResult[22]>
#>  57:            0.067 2022-04-16 04:15:29       57 <ResampleResult[22]>
#>  58:            0.067 2022-04-16 04:15:29       58 <ResampleResult[22]>
#>  59:            0.067 2022-04-16 04:15:29       59 <ResampleResult[22]>
#>  60:            0.062 2022-04-16 04:15:30       60 <ResampleResult[22]>
#>  61:            0.067 2022-04-16 04:15:30       61 <ResampleResult[22]>
#>  62:            0.086 2022-04-16 04:15:30       62 <ResampleResult[22]>
#>  63:            0.080 2022-04-16 04:15:30       63 <ResampleResult[22]>
#>  64:            0.066 2022-04-16 04:15:30       64 <ResampleResult[22]>
#>  65:            0.066 2022-04-16 04:15:31       65 <ResampleResult[22]>
#>  66:            0.066 2022-04-16 04:15:31       66 <ResampleResult[22]>
#>  67:            0.069 2022-04-16 04:15:31       67 <ResampleResult[22]>
#>  68:            0.081 2022-04-16 04:15:31       68 <ResampleResult[22]>
#>  69:            0.079 2022-04-16 04:15:32       69 <ResampleResult[22]>
#>  70:            0.068 2022-04-16 04:15:32       70 <ResampleResult[22]>
#>  71:            0.066 2022-04-16 04:15:32       71 <ResampleResult[22]>
#>  72:            0.065 2022-04-16 04:15:32       72 <ResampleResult[22]>
#>  73:            0.065 2022-04-16 04:15:32       73 <ResampleResult[22]>
#>  74:            0.074 2022-04-16 04:15:33       74 <ResampleResult[22]>
#>  75:            0.083 2022-04-16 04:15:33       75 <ResampleResult[22]>
#>  76:            0.085 2022-04-16 04:15:33       76 <ResampleResult[22]>
#>  77:            0.065 2022-04-16 04:15:33       77 <ResampleResult[22]>
#>  78:            0.067 2022-04-16 04:15:33       78 <ResampleResult[22]>
#>  79:            0.066 2022-04-16 04:15:34       79 <ResampleResult[22]>
#>  80:            0.064 2022-04-16 04:15:34       80 <ResampleResult[22]>
#>  81:            0.064 2022-04-16 04:15:34       81 <ResampleResult[22]>
#>  82:            0.064 2022-04-16 04:15:34       82 <ResampleResult[22]>
#>  83:            0.068 2022-04-16 04:15:34       83 <ResampleResult[22]>
#>  84:            0.075 2022-04-16 04:15:35       84 <ResampleResult[22]>
#>  85:            0.125 2022-04-16 04:15:35       85 <ResampleResult[22]>
#>  86:            0.079 2022-04-16 04:15:35       86 <ResampleResult[22]>
#>  87:            0.069 2022-04-16 04:15:35       87 <ResampleResult[22]>
#>  88:            0.067 2022-04-16 04:15:36       88 <ResampleResult[22]>
#>  89:            0.064 2022-04-16 04:15:36       89 <ResampleResult[22]>
#>  90:            0.076 2022-04-16 04:15:36       90 <ResampleResult[22]>
#>  91:            0.086 2022-04-16 04:15:36       91 <ResampleResult[22]>
#>  92:            0.068 2022-04-16 04:15:36       92 <ResampleResult[22]>
#>  93:            0.068 2022-04-16 04:15:37       93 <ResampleResult[22]>
#>  94:            0.068 2022-04-16 04:15:37       94 <ResampleResult[22]>
#>  95:            0.067 2022-04-16 04:15:37       95 <ResampleResult[22]>
#>  96:            0.073 2022-04-16 04:15:37       96 <ResampleResult[22]>
#>  97:            0.090 2022-04-16 04:15:38       97 <ResampleResult[22]>
#>  98:            0.086 2022-04-16 04:15:38       98 <ResampleResult[22]>
#>  99:            0.077 2022-04-16 04:15:38       99 <ResampleResult[22]>
#> 100:            0.069 2022-04-16 04:15:38      100 <ResampleResult[22]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }