Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>      age glucose insulin  mass pedigree pregnant pressure triceps
#> 1: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE
#>                    features classif.ce
#> 1: glucose,pedigree,triceps  0.2695312

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1:  TRUE   FALSE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.4023438
#>   2:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>   3:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.3281250
#>   4:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>   5:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2773438
#>   6:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3242188
#>   7: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3007812
#>   8:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.3593750
#>   9: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.2695312
#>  10:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.2968750
#>  11: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2773438
#>  12:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3945312
#>  13: FALSE    TRUE    TRUE FALSE    FALSE    FALSE     TRUE    TRUE  0.2929688
#>  14: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3632812
#>  15:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  16: FALSE   FALSE    TRUE FALSE     TRUE    FALSE    FALSE   FALSE  0.3632812
#>  17: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.3398438
#>  18:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2773438
#>  19:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3632812
#>  20:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  21: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.4375000
#>  22:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE   FALSE  0.3203125
#>  23: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.4375000
#>  24: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.4375000
#>  25:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.3359375
#>  26: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3906250
#>  27: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3906250
#>  28: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.3164062
#>  29:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3906250
#>  30: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  31: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2929688
#>  32:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  33:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3945312
#>  34: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  35: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2695312
#>  36:  TRUE    TRUE   FALSE FALSE     TRUE    FALSE     TRUE    TRUE  0.3046875
#>  37:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  38: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2734375
#>  39:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  40: FALSE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.2968750
#>  41:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  42:  TRUE   FALSE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.3789062
#>  43: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3007812
#>  44: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3203125
#>  45:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2773438
#>  46:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  47: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.3828125
#>  48:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  49:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2773438
#>  50:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2773438
#>  51:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3593750
#>  52:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3632812
#>  53: FALSE   FALSE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.2968750
#>  54:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3476562
#>  55:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.3593750
#>  56:  TRUE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.3203125
#>  57: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2929688
#>  58: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  59: FALSE   FALSE    TRUE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2890625
#>  60:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.3046875
#>  61: FALSE    TRUE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3046875
#>  62: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.3085938
#>  63:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#>  64: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3828125
#>  65: FALSE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.3242188
#>  66: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2929688
#>  67:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2773438
#>  68:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3203125
#>  69:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  70: FALSE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.3320312
#>  71: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3007812
#>  72:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3164062
#>  73:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#>  74:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE    TRUE  0.3242188
#>  75:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3320312
#>  76:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2773438
#>  77: FALSE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3085938
#>  78: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2968750
#>  79:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2773438
#>  80:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3281250
#>  81:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  82:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2773438
#>  83: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE    TRUE  0.4375000
#>  84: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3203125
#>  85:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2773438
#>  86:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE    TRUE  0.2773438
#>  87: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE    TRUE  0.3945312
#>  88:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  89: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2812500
#>  90:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  91: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3007812
#>  92: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2773438
#>  93: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2812500
#>  94:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.3046875
#>  95: FALSE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE    TRUE  0.4023438
#>  96:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE   FALSE  0.3203125
#>  97:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3398438
#>  98: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  99: FALSE   FALSE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.3437500
#> 100:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3945312
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.066 2022-08-10 04:13:39        1 <ResampleResult[21]>
#>   2:            0.066 2022-08-10 04:13:40        2 <ResampleResult[21]>
#>   3:            0.063 2022-08-10 04:13:40        3 <ResampleResult[21]>
#>   4:            0.064 2022-08-10 04:13:40        4 <ResampleResult[21]>
#>   5:            0.063 2022-08-10 04:13:40        5 <ResampleResult[21]>
#>   6:            0.063 2022-08-10 04:13:40        6 <ResampleResult[21]>
#>   7:            0.061 2022-08-10 04:13:41        7 <ResampleResult[21]>
#>   8:            0.070 2022-08-10 04:13:41        8 <ResampleResult[21]>
#>   9:            0.074 2022-08-10 04:13:41        9 <ResampleResult[21]>
#>  10:            0.080 2022-08-10 04:13:41       10 <ResampleResult[21]>
#>  11:            0.085 2022-08-10 04:13:42       11 <ResampleResult[21]>
#>  12:            0.076 2022-08-10 04:13:42       12 <ResampleResult[21]>
#>  13:            0.068 2022-08-10 04:13:42       13 <ResampleResult[21]>
#>  14:            0.062 2022-08-10 04:13:42       14 <ResampleResult[21]>
#>  15:            0.060 2022-08-10 04:13:42       15 <ResampleResult[21]>
#>  16:            0.069 2022-08-10 04:13:43       16 <ResampleResult[21]>
#>  17:            0.071 2022-08-10 04:13:43       17 <ResampleResult[21]>
#>  18:            0.085 2022-08-10 04:13:43       18 <ResampleResult[21]>
#>  19:            0.070 2022-08-10 04:13:43       19 <ResampleResult[21]>
#>  20:            0.059 2022-08-10 04:13:44       20 <ResampleResult[21]>
#>  21:            0.059 2022-08-10 04:13:44       21 <ResampleResult[21]>
#>  22:            0.064 2022-08-10 04:13:44       22 <ResampleResult[21]>
#>  23:            0.062 2022-08-10 04:13:44       23 <ResampleResult[21]>
#>  24:            0.064 2022-08-10 04:13:44       24 <ResampleResult[21]>
#>  25:            0.060 2022-08-10 04:13:44       25 <ResampleResult[21]>
#>  26:            0.058 2022-08-10 04:13:45       26 <ResampleResult[21]>
#>  27:            0.058 2022-08-10 04:13:45       27 <ResampleResult[21]>
#>  28:            0.058 2022-08-10 04:13:45       28 <ResampleResult[21]>
#>  29:            0.058 2022-08-10 04:13:45       29 <ResampleResult[21]>
#>  30:            0.059 2022-08-10 04:13:45       30 <ResampleResult[21]>
#>  31:            0.058 2022-08-10 04:13:45       31 <ResampleResult[21]>
#>  32:            0.064 2022-08-10 04:13:46       32 <ResampleResult[21]>
#>  33:            0.058 2022-08-10 04:13:46       33 <ResampleResult[21]>
#>  34:            0.063 2022-08-10 04:13:46       34 <ResampleResult[21]>
#>  35:            0.065 2022-08-10 04:13:46       35 <ResampleResult[21]>
#>  36:            0.059 2022-08-10 04:13:46       36 <ResampleResult[21]>
#>  37:            0.061 2022-08-10 04:13:47       37 <ResampleResult[21]>
#>  38:            0.060 2022-08-10 04:13:47       38 <ResampleResult[21]>
#>  39:            0.061 2022-08-10 04:13:47       39 <ResampleResult[21]>
#>  40:            0.060 2022-08-10 04:13:47       40 <ResampleResult[21]>
#>  41:            0.061 2022-08-10 04:13:47       41 <ResampleResult[21]>
#>  42:            0.060 2022-08-10 04:13:47       42 <ResampleResult[21]>
#>  43:            0.059 2022-08-10 04:13:48       43 <ResampleResult[21]>
#>  44:            0.058 2022-08-10 04:13:48       44 <ResampleResult[21]>
#>  45:            0.059 2022-08-10 04:13:48       45 <ResampleResult[21]>
#>  46:            0.065 2022-08-10 04:13:48       46 <ResampleResult[21]>
#>  47:            0.071 2022-08-10 04:13:48       47 <ResampleResult[21]>
#>  48:            0.069 2022-08-10 04:13:49       48 <ResampleResult[21]>
#>  49:            0.061 2022-08-10 04:13:49       49 <ResampleResult[21]>
#>  50:            0.061 2022-08-10 04:13:49       50 <ResampleResult[21]>
#>  51:            0.060 2022-08-10 04:13:49       51 <ResampleResult[21]>
#>  52:            0.060 2022-08-10 04:13:49       52 <ResampleResult[21]>
#>  53:            0.059 2022-08-10 04:13:49       53 <ResampleResult[21]>
#>  54:            0.058 2022-08-10 04:13:50       54 <ResampleResult[21]>
#>  55:            0.059 2022-08-10 04:13:50       55 <ResampleResult[21]>
#>  56:            0.063 2022-08-10 04:13:50       56 <ResampleResult[21]>
#>  57:            0.072 2022-08-10 04:13:50       57 <ResampleResult[21]>
#>  58:            0.070 2022-08-10 04:13:50       58 <ResampleResult[21]>
#>  59:            0.071 2022-08-10 04:13:51       59 <ResampleResult[21]>
#>  60:            0.073 2022-08-10 04:13:51       60 <ResampleResult[21]>
#>  61:            0.072 2022-08-10 04:13:51       61 <ResampleResult[21]>
#>  62:            0.070 2022-08-10 04:13:51       62 <ResampleResult[21]>
#>  63:            0.068 2022-08-10 04:13:51       63 <ResampleResult[21]>
#>  64:            0.070 2022-08-10 04:13:52       64 <ResampleResult[21]>
#>  65:            0.094 2022-08-10 04:13:52       65 <ResampleResult[21]>
#>  66:            0.060 2022-08-10 04:13:52       66 <ResampleResult[21]>
#>  67:            0.062 2022-08-10 04:13:52       67 <ResampleResult[21]>
#>  68:            0.061 2022-08-10 04:13:52       68 <ResampleResult[21]>
#>  69:            0.062 2022-08-10 04:13:52       69 <ResampleResult[21]>
#>  70:            0.062 2022-08-10 04:13:53       70 <ResampleResult[21]>
#>  71:            0.069 2022-08-10 04:13:53       71 <ResampleResult[21]>
#>  72:            0.075 2022-08-10 04:13:53       72 <ResampleResult[21]>
#>  73:            0.071 2022-08-10 04:13:53       73 <ResampleResult[21]>
#>  74:            0.061 2022-08-10 04:13:53       74 <ResampleResult[21]>
#>  75:            0.061 2022-08-10 04:13:54       75 <ResampleResult[21]>
#>  76:            0.060 2022-08-10 04:13:54       76 <ResampleResult[21]>
#>  77:            0.062 2022-08-10 04:13:54       77 <ResampleResult[21]>
#>  78:            0.060 2022-08-10 04:13:54       78 <ResampleResult[21]>
#>  79:            0.060 2022-08-10 04:13:54       79 <ResampleResult[21]>
#>  80:            0.060 2022-08-10 04:13:55       80 <ResampleResult[21]>
#>  81:            0.060 2022-08-10 04:13:55       81 <ResampleResult[21]>
#>  82:            0.060 2022-08-10 04:13:55       82 <ResampleResult[21]>
#>  83:            0.061 2022-08-10 04:13:55       83 <ResampleResult[21]>
#>  84:            0.062 2022-08-10 04:13:55       84 <ResampleResult[21]>
#>  85:            0.065 2022-08-10 04:13:56       85 <ResampleResult[21]>
#>  86:            0.066 2022-08-10 04:13:56       86 <ResampleResult[21]>
#>  87:            0.066 2022-08-10 04:13:56       87 <ResampleResult[21]>
#>  88:            0.075 2022-08-10 04:13:56       88 <ResampleResult[21]>
#>  89:            0.076 2022-08-10 04:13:56       89 <ResampleResult[21]>
#>  90:            0.070 2022-08-10 04:13:57       90 <ResampleResult[21]>
#>  91:            0.061 2022-08-10 04:13:57       91 <ResampleResult[21]>
#>  92:            0.062 2022-08-10 04:13:57       92 <ResampleResult[21]>
#>  93:            0.060 2022-08-10 04:13:57       93 <ResampleResult[21]>
#>  94:            0.061 2022-08-10 04:13:57       94 <ResampleResult[21]>
#>  95:            0.061 2022-08-10 04:13:58       95 <ResampleResult[21]>
#>  96:            0.066 2022-08-10 04:13:58       96 <ResampleResult[21]>
#>  97:            0.070 2022-08-10 04:13:58       97 <ResampleResult[21]>
#>  98:            0.080 2022-08-10 04:13:58       98 <ResampleResult[21]>
#>  99:            0.078 2022-08-10 04:13:58       99 <ResampleResult[21]>
#> 100:            0.073 2022-08-10 04:13:59      100 <ResampleResult[21]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }