Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>     age glucose insulin  mass pedigree pregnant pressure triceps
#> 1: TRUE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE
#>                                 features classif.ce
#> 1: age,glucose,insulin,pregnant,pressure  0.2148438

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2656250
#>   2: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>   3:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>   4: FALSE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>   5: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2226562
#>   6:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2617188
#>   7:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2421875
#>   8: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2304688
#>   9: FALSE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.2773438
#>  10: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2382812
#>  11: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3046875
#>  12:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  13: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2304688
#>  14:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2773438
#>  15:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3164062
#>  16: FALSE   FALSE    TRUE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2812500
#>  17:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2421875
#>  18:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.2421875
#>  19:  TRUE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.3476562
#>  20: FALSE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2500000
#>  21: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2226562
#>  22:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2304688
#>  23:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3007812
#>  24: FALSE   FALSE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.3164062
#>  25: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.3398438
#>  26:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2382812
#>  27:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.2148438
#>  28:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.2148438
#>  29:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3007812
#>  30: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2226562
#>  31: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2773438
#>  32:  TRUE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2226562
#>  33:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2500000
#>  34: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2656250
#>  35:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2773438
#>  36:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2382812
#>  37:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  38: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2304688
#>  39: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2343750
#>  40:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3085938
#>  41: FALSE   FALSE   FALSE FALSE     TRUE    FALSE     TRUE   FALSE  0.4023438
#>  42:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  43:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2773438
#>  44:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2539062
#>  45: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2382812
#>  46: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.3710938
#>  47: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2304688
#>  48:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3164062
#>  49: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.4687500
#>  50:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  51: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.3828125
#>  52: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.3398438
#>  53:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2851562
#>  54:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2304688
#>  55: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2812500
#>  56:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3007812
#>  57: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2304688
#>  58:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2539062
#>  59:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  60: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2968750
#>  61: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3476562
#>  62:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2421875
#>  63: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2304688
#>  64: FALSE   FALSE   FALSE FALSE     TRUE    FALSE     TRUE    TRUE  0.3750000
#>  65: FALSE    TRUE    TRUE FALSE    FALSE    FALSE     TRUE    TRUE  0.2265625
#>  66: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.2656250
#>  67:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  68:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  69: FALSE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.3125000
#>  70:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2500000
#>  71:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  72:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  73:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2968750
#>  74:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2382812
#>  75:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2773438
#>  76:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3164062
#>  77: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3710938
#>  78: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.4687500
#>  79:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2539062
#>  80: FALSE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.2851562
#>  81:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3164062
#>  82:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE    TRUE  0.3242188
#>  83: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3515625
#>  84: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3281250
#>  85:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE    TRUE  0.2421875
#>  86:  TRUE   FALSE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.3632812
#>  87: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2421875
#>  88:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2539062
#>  89:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2265625
#>  90: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3164062
#>  91: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2500000
#>  92: FALSE    TRUE   FALSE FALSE     TRUE    FALSE     TRUE   FALSE  0.2460938
#>  93:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE    TRUE  0.3359375
#>  94: FALSE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.2968750
#>  95:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2304688
#>  96:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  97:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2539062
#>  98:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  99:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2539062
#> 100: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2890625
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.068 2022-03-26 04:15:39        1 <ResampleResult[22]>
#>   2:            0.069 2022-03-26 04:15:39        2 <ResampleResult[22]>
#>   3:            0.070 2022-03-26 04:15:39        3 <ResampleResult[22]>
#>   4:            0.071 2022-03-26 04:15:39        4 <ResampleResult[22]>
#>   5:            0.067 2022-03-26 04:15:40        5 <ResampleResult[22]>
#>   6:            0.070 2022-03-26 04:15:40        6 <ResampleResult[22]>
#>   7:            0.069 2022-03-26 04:15:40        7 <ResampleResult[22]>
#>   8:            0.069 2022-03-26 04:15:40        8 <ResampleResult[22]>
#>   9:            0.070 2022-03-26 04:15:41        9 <ResampleResult[22]>
#>  10:            0.069 2022-03-26 04:15:41       10 <ResampleResult[22]>
#>  11:            0.066 2022-03-26 04:15:41       11 <ResampleResult[22]>
#>  12:            0.069 2022-03-26 04:15:41       12 <ResampleResult[22]>
#>  13:            0.070 2022-03-26 04:15:42       13 <ResampleResult[22]>
#>  14:            0.070 2022-03-26 04:15:42       14 <ResampleResult[22]>
#>  15:            0.070 2022-03-26 04:15:42       15 <ResampleResult[22]>
#>  16:            0.070 2022-03-26 04:15:42       16 <ResampleResult[22]>
#>  17:            0.073 2022-03-26 04:15:43       17 <ResampleResult[22]>
#>  18:            0.079 2022-03-26 04:15:43       18 <ResampleResult[22]>
#>  19:            0.080 2022-03-26 04:15:43       19 <ResampleResult[22]>
#>  20:            0.085 2022-03-26 04:15:43       20 <ResampleResult[22]>
#>  21:            0.093 2022-03-26 04:15:43       21 <ResampleResult[22]>
#>  22:            0.096 2022-03-26 04:15:44       22 <ResampleResult[22]>
#>  23:            0.094 2022-03-26 04:15:44       23 <ResampleResult[22]>
#>  24:            0.103 2022-03-26 04:15:44       24 <ResampleResult[22]>
#>  25:            0.101 2022-03-26 04:15:44       25 <ResampleResult[22]>
#>  26:            0.101 2022-03-26 04:15:45       26 <ResampleResult[22]>
#>  27:            0.102 2022-03-26 04:15:45       27 <ResampleResult[22]>
#>  28:            0.104 2022-03-26 04:15:45       28 <ResampleResult[22]>
#>  29:            0.103 2022-03-26 04:15:45       29 <ResampleResult[22]>
#>  30:            0.094 2022-03-26 04:15:46       30 <ResampleResult[22]>
#>  31:            0.094 2022-03-26 04:15:46       31 <ResampleResult[22]>
#>  32:            0.089 2022-03-26 04:15:46       32 <ResampleResult[22]>
#>  33:            0.086 2022-03-26 04:15:46       33 <ResampleResult[22]>
#>  34:            0.083 2022-03-26 04:15:47       34 <ResampleResult[22]>
#>  35:            0.081 2022-03-26 04:15:47       35 <ResampleResult[22]>
#>  36:            0.065 2022-03-26 04:15:47       36 <ResampleResult[22]>
#>  37:            0.065 2022-03-26 04:15:47       37 <ResampleResult[22]>
#>  38:            0.072 2022-03-26 04:15:48       38 <ResampleResult[22]>
#>  39:            0.066 2022-03-26 04:15:48       39 <ResampleResult[22]>
#>  40:            0.066 2022-03-26 04:15:48       40 <ResampleResult[22]>
#>  41:            0.064 2022-03-26 04:15:48       41 <ResampleResult[22]>
#>  42:            0.064 2022-03-26 04:15:48       42 <ResampleResult[22]>
#>  43:            0.063 2022-03-26 04:15:49       43 <ResampleResult[22]>
#>  44:            0.074 2022-03-26 04:15:49       44 <ResampleResult[22]>
#>  45:            0.064 2022-03-26 04:15:49       45 <ResampleResult[22]>
#>  46:            0.067 2022-03-26 04:15:49       46 <ResampleResult[22]>
#>  47:            0.067 2022-03-26 04:15:49       47 <ResampleResult[22]>
#>  48:            0.066 2022-03-26 04:15:50       48 <ResampleResult[22]>
#>  49:            0.065 2022-03-26 04:15:50       49 <ResampleResult[22]>
#>  50:            0.064 2022-03-26 04:15:50       50 <ResampleResult[22]>
#>  51:            0.073 2022-03-26 04:15:50       51 <ResampleResult[22]>
#>  52:            0.074 2022-03-26 04:15:50       52 <ResampleResult[22]>
#>  53:            0.066 2022-03-26 04:15:51       53 <ResampleResult[22]>
#>  54:            0.067 2022-03-26 04:15:51       54 <ResampleResult[22]>
#>  55:            0.066 2022-03-26 04:15:51       55 <ResampleResult[22]>
#>  56:            0.066 2022-03-26 04:15:51       56 <ResampleResult[22]>
#>  57:            0.065 2022-03-26 04:15:51       57 <ResampleResult[22]>
#>  58:            0.066 2022-03-26 04:15:52       58 <ResampleResult[22]>
#>  59:            0.066 2022-03-26 04:15:52       59 <ResampleResult[22]>
#>  60:            0.071 2022-03-26 04:15:52       60 <ResampleResult[22]>
#>  61:            0.076 2022-03-26 04:15:52       61 <ResampleResult[22]>
#>  62:            0.068 2022-03-26 04:15:52       62 <ResampleResult[22]>
#>  63:            0.067 2022-03-26 04:15:53       63 <ResampleResult[22]>
#>  64:            0.064 2022-03-26 04:15:53       64 <ResampleResult[22]>
#>  65:            0.065 2022-03-26 04:15:53       65 <ResampleResult[22]>
#>  66:            0.075 2022-03-26 04:15:53       66 <ResampleResult[22]>
#>  67:            0.087 2022-03-26 04:15:54       67 <ResampleResult[22]>
#>  68:            0.068 2022-03-26 04:15:54       68 <ResampleResult[22]>
#>  69:            0.066 2022-03-26 04:15:54       69 <ResampleResult[22]>
#>  70:            0.066 2022-03-26 04:15:54       70 <ResampleResult[22]>
#>  71:            0.067 2022-03-26 04:15:54       71 <ResampleResult[22]>
#>  72:            0.067 2022-03-26 04:15:55       72 <ResampleResult[22]>
#>  73:            0.075 2022-03-26 04:15:55       73 <ResampleResult[22]>
#>  74:            0.083 2022-03-26 04:15:55       74 <ResampleResult[22]>
#>  75:            0.079 2022-03-26 04:15:55       75 <ResampleResult[22]>
#>  76:            0.066 2022-03-26 04:15:55       76 <ResampleResult[22]>
#>  77:            0.066 2022-03-26 04:15:56       77 <ResampleResult[22]>
#>  78:            0.064 2022-03-26 04:15:56       78 <ResampleResult[22]>
#>  79:            0.065 2022-03-26 04:15:56       79 <ResampleResult[22]>
#>  80:            0.066 2022-03-26 04:15:56       80 <ResampleResult[22]>
#>  81:            0.065 2022-03-26 04:15:56       81 <ResampleResult[22]>
#>  82:            0.073 2022-03-26 04:15:57       82 <ResampleResult[22]>
#>  83:            0.082 2022-03-26 04:15:57       83 <ResampleResult[22]>
#>  84:            0.124 2022-03-26 04:15:57       84 <ResampleResult[22]>
#>  85:            0.081 2022-03-26 04:15:57       85 <ResampleResult[22]>
#>  86:            0.071 2022-03-26 04:15:58       86 <ResampleResult[22]>
#>  87:            0.068 2022-03-26 04:15:58       87 <ResampleResult[22]>
#>  88:            0.067 2022-03-26 04:15:58       88 <ResampleResult[22]>
#>  89:            0.071 2022-03-26 04:15:58       89 <ResampleResult[22]>
#>  90:            0.085 2022-03-26 04:15:58       90 <ResampleResult[22]>
#>  91:            0.083 2022-03-26 04:15:59       91 <ResampleResult[22]>
#>  92:            0.069 2022-03-26 04:15:59       92 <ResampleResult[22]>
#>  93:            0.068 2022-03-26 04:15:59       93 <ResampleResult[22]>
#>  94:            0.067 2022-03-26 04:15:59       94 <ResampleResult[22]>
#>  95:            0.067 2022-03-26 04:16:00       95 <ResampleResult[22]>
#>  96:            0.069 2022-03-26 04:16:00       96 <ResampleResult[22]>
#>  97:            0.087 2022-03-26 04:16:00       97 <ResampleResult[22]>
#>  98:            0.086 2022-03-26 04:16:00       98 <ResampleResult[22]>
#>  99:            0.083 2022-03-26 04:16:00       99 <ResampleResult[22]>
#> 100:            0.069 2022-03-26 04:16:01      100 <ResampleResult[22]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }