Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>     age glucose insulin mass pedigree pregnant pressure triceps
#> 1: TRUE    TRUE    TRUE TRUE     TRUE     TRUE     TRUE    TRUE
#>                                          features classif.ce
#> 1: age,glucose,insulin,mass,pedigree,pregnant,...  0.2265625

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2968750
#>   2:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2421875
#>   3:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3007812
#>   4:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3554688
#>   5: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2500000
#>   6:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3242188
#>   7: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2656250
#>   8: FALSE   FALSE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.3007812
#>   9:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3007812
#>  10: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3867188
#>  11:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  12: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2460938
#>  13:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>  14:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2929688
#>  15:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE    FALSE   FALSE  0.2421875
#>  16:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  17:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  18:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2304688
#>  19: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2500000
#>  20:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.3085938
#>  21: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3242188
#>  22:  TRUE   FALSE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.2968750
#>  23: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3671875
#>  24:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2304688
#>  25:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2656250
#>  26:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE   FALSE  0.2421875
#>  27:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2304688
#>  28:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2265625
#>  29: FALSE    TRUE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.2734375
#>  30:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>  31:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2265625
#>  32:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  33: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.4062500
#>  34: FALSE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2695312
#>  35:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2890625
#>  36:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2265625
#>  37: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2929688
#>  38:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2968750
#>  39: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2656250
#>  40:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  41:  TRUE   FALSE   FALSE FALSE     TRUE     TRUE    FALSE    TRUE  0.3046875
#>  42: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.4062500
#>  43: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2578125
#>  44: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2617188
#>  45: FALSE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2617188
#>  46: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3671875
#>  47:  TRUE    TRUE    TRUE FALSE    FALSE    FALSE     TRUE   FALSE  0.2460938
#>  48:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2968750
#>  49: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3906250
#>  50:  TRUE    TRUE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.2460938
#>  51: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.3007812
#>  52:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2265625
#>  53:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2968750
#>  54:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  55:  TRUE    TRUE    TRUE FALSE    FALSE    FALSE     TRUE   FALSE  0.2460938
#>  56:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2460938
#>  57:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2656250
#>  58:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2890625
#>  59: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2500000
#>  60:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>  61:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2968750
#>  62:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE    FALSE   FALSE  0.2421875
#>  63: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3476562
#>  64:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2656250
#>  65: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2695312
#>  66: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3085938
#>  67:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  68:  TRUE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE    TRUE  0.2460938
#>  69:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3085938
#>  70: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2460938
#>  71: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3437500
#>  72: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2929688
#>  73:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  74: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2539062
#>  75: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2734375
#>  76:  TRUE   FALSE   FALSE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2812500
#>  77:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3281250
#>  78: FALSE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.2265625
#>  79:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2265625
#>  80:  TRUE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2421875
#>  81:  TRUE    TRUE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.2460938
#>  82: FALSE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.2578125
#>  83:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  84: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.2578125
#>  85:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2968750
#>  86:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  87:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2304688
#>  88:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2265625
#>  89:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2265625
#>  90:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2304688
#>  91: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3437500
#>  92:  TRUE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2890625
#>  93:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2421875
#>  94:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2265625
#>  95:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE     TRUE   FALSE  0.2773438
#>  96:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>  97: FALSE   FALSE   FALSE  TRUE     TRUE    FALSE     TRUE   FALSE  0.3085938
#>  98: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.2578125
#>  99: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.3437500
#> 100:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2304688
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.116 2022-03-18 04:18:47        1 <ResampleResult[22]>
#>   2:            0.092 2022-03-18 04:18:47        2 <ResampleResult[22]>
#>   3:            0.094 2022-03-18 04:18:47        3 <ResampleResult[22]>
#>   4:            0.088 2022-03-18 04:18:47        4 <ResampleResult[22]>
#>   5:            0.089 2022-03-18 04:18:48        5 <ResampleResult[22]>
#>   6:            0.087 2022-03-18 04:18:48        6 <ResampleResult[22]>
#>   7:            0.093 2022-03-18 04:18:48        7 <ResampleResult[22]>
#>   8:            0.082 2022-03-18 04:18:49        8 <ResampleResult[22]>
#>   9:            0.086 2022-03-18 04:18:49        9 <ResampleResult[22]>
#>  10:            0.080 2022-03-18 04:18:49       10 <ResampleResult[22]>
#>  11:            0.093 2022-03-18 04:18:49       11 <ResampleResult[22]>
#>  12:            0.101 2022-03-18 04:18:50       12 <ResampleResult[22]>
#>  13:            0.101 2022-03-18 04:18:50       13 <ResampleResult[22]>
#>  14:            0.100 2022-03-18 04:18:50       14 <ResampleResult[22]>
#>  15:            0.099 2022-03-18 04:18:51       15 <ResampleResult[22]>
#>  16:            0.104 2022-03-18 04:18:51       16 <ResampleResult[22]>
#>  17:            0.107 2022-03-18 04:18:51       17 <ResampleResult[22]>
#>  18:            0.107 2022-03-18 04:18:51       18 <ResampleResult[22]>
#>  19:            0.105 2022-03-18 04:18:52       19 <ResampleResult[22]>
#>  20:            0.084 2022-03-18 04:18:52       20 <ResampleResult[22]>
#>  21:            0.088 2022-03-18 04:18:52       21 <ResampleResult[22]>
#>  22:            0.086 2022-03-18 04:18:53       22 <ResampleResult[22]>
#>  23:            0.092 2022-03-18 04:18:53       23 <ResampleResult[22]>
#>  24:            0.089 2022-03-18 04:18:53       24 <ResampleResult[22]>
#>  25:            0.102 2022-03-18 04:18:53       25 <ResampleResult[22]>
#>  26:            0.108 2022-03-18 04:18:54       26 <ResampleResult[22]>
#>  27:            0.118 2022-03-18 04:18:54       27 <ResampleResult[22]>
#>  28:            0.117 2022-03-18 04:18:54       28 <ResampleResult[22]>
#>  29:            0.112 2022-03-18 04:18:55       29 <ResampleResult[22]>
#>  30:            0.111 2022-03-18 04:18:55       30 <ResampleResult[22]>
#>  31:            0.106 2022-03-18 04:18:55       31 <ResampleResult[22]>
#>  32:            0.103 2022-03-18 04:18:56       32 <ResampleResult[22]>
#>  33:            0.090 2022-03-18 04:18:56       33 <ResampleResult[22]>
#>  34:            0.094 2022-03-18 04:18:56       34 <ResampleResult[22]>
#>  35:            0.089 2022-03-18 04:18:56       35 <ResampleResult[22]>
#>  36:            0.092 2022-03-18 04:18:57       36 <ResampleResult[22]>
#>  37:            0.091 2022-03-18 04:18:57       37 <ResampleResult[22]>
#>  38:            0.094 2022-03-18 04:18:57       38 <ResampleResult[22]>
#>  39:            0.093 2022-03-18 04:18:58       39 <ResampleResult[22]>
#>  40:            0.090 2022-03-18 04:18:58       40 <ResampleResult[22]>
#>  41:            0.086 2022-03-18 04:18:58       41 <ResampleResult[22]>
#>  42:            0.086 2022-03-18 04:18:59       42 <ResampleResult[22]>
#>  43:            0.095 2022-03-18 04:18:59       43 <ResampleResult[22]>
#>  44:            0.084 2022-03-18 04:18:59       44 <ResampleResult[22]>
#>  45:            0.087 2022-03-18 04:18:59       45 <ResampleResult[22]>
#>  46:            0.081 2022-03-18 04:19:00       46 <ResampleResult[22]>
#>  47:            0.085 2022-03-18 04:19:00       47 <ResampleResult[22]>
#>  48:            0.094 2022-03-18 04:19:00       48 <ResampleResult[22]>
#>  49:            0.085 2022-03-18 04:19:00       49 <ResampleResult[22]>
#>  50:            0.086 2022-03-18 04:19:01       50 <ResampleResult[22]>
#>  51:            0.089 2022-03-18 04:19:01       51 <ResampleResult[22]>
#>  52:            0.089 2022-03-18 04:19:01       52 <ResampleResult[22]>
#>  53:            0.088 2022-03-18 04:19:01       53 <ResampleResult[22]>
#>  54:            0.098 2022-03-18 04:19:02       54 <ResampleResult[22]>
#>  55:            0.095 2022-03-18 04:19:02       55 <ResampleResult[22]>
#>  56:            0.087 2022-03-18 04:19:02       56 <ResampleResult[22]>
#>  57:            0.087 2022-03-18 04:19:03       57 <ResampleResult[22]>
#>  58:            0.089 2022-03-18 04:19:03       58 <ResampleResult[22]>
#>  59:            0.087 2022-03-18 04:19:03       59 <ResampleResult[22]>
#>  60:            0.086 2022-03-18 04:19:03       60 <ResampleResult[22]>
#>  61:            0.093 2022-03-18 04:19:04       61 <ResampleResult[22]>
#>  62:            0.085 2022-03-18 04:19:04       62 <ResampleResult[22]>
#>  63:            0.089 2022-03-18 04:19:04       63 <ResampleResult[22]>
#>  64:            0.083 2022-03-18 04:19:04       64 <ResampleResult[22]>
#>  65:            0.084 2022-03-18 04:19:05       65 <ResampleResult[22]>
#>  66:            0.085 2022-03-18 04:19:05       66 <ResampleResult[22]>
#>  67:            0.090 2022-03-18 04:19:05       67 <ResampleResult[22]>
#>  68:            0.103 2022-03-18 04:19:05       68 <ResampleResult[22]>
#>  69:            0.089 2022-03-18 04:19:06       69 <ResampleResult[22]>
#>  70:            0.091 2022-03-18 04:19:06       70 <ResampleResult[22]>
#>  71:            0.089 2022-03-18 04:19:06       71 <ResampleResult[22]>
#>  72:            0.085 2022-03-18 04:19:06       72 <ResampleResult[22]>
#>  73:            0.102 2022-03-18 04:19:07       73 <ResampleResult[22]>
#>  74:            0.100 2022-03-18 04:19:07       74 <ResampleResult[22]>
#>  75:            0.091 2022-03-18 04:19:07       75 <ResampleResult[22]>
#>  76:            0.091 2022-03-18 04:19:08       76 <ResampleResult[22]>
#>  77:            0.081 2022-03-18 04:19:08       77 <ResampleResult[22]>
#>  78:            0.088 2022-03-18 04:19:08       78 <ResampleResult[22]>
#>  79:            0.094 2022-03-18 04:19:08       79 <ResampleResult[22]>
#>  80:            0.093 2022-03-18 04:19:09       80 <ResampleResult[22]>
#>  81:            0.096 2022-03-18 04:19:09       81 <ResampleResult[22]>
#>  82:            0.092 2022-03-18 04:19:09       82 <ResampleResult[22]>
#>  83:            0.092 2022-03-18 04:19:09       83 <ResampleResult[22]>
#>  84:            0.087 2022-03-18 04:19:10       84 <ResampleResult[22]>
#>  85:            0.086 2022-03-18 04:19:10       85 <ResampleResult[22]>
#>  86:            0.079 2022-03-18 04:19:10       86 <ResampleResult[22]>
#>  87:            0.087 2022-03-18 04:19:10       87 <ResampleResult[22]>
#>  88:            0.104 2022-03-18 04:19:11       88 <ResampleResult[22]>
#>  89:            0.140 2022-03-18 04:19:11       89 <ResampleResult[22]>
#>  90:            0.110 2022-03-18 04:19:11       90 <ResampleResult[22]>
#>  91:            0.086 2022-03-18 04:19:12       91 <ResampleResult[22]>
#>  92:            0.083 2022-03-18 04:19:12       92 <ResampleResult[22]>
#>  93:            0.094 2022-03-18 04:19:12       93 <ResampleResult[22]>
#>  94:            0.104 2022-03-18 04:19:12       94 <ResampleResult[22]>
#>  95:            0.095 2022-03-18 04:19:13       95 <ResampleResult[22]>
#>  96:            0.091 2022-03-18 04:19:13       96 <ResampleResult[22]>
#>  97:            0.088 2022-03-18 04:19:13       97 <ResampleResult[22]>
#>  98:            0.084 2022-03-18 04:19:14       98 <ResampleResult[22]>
#>  99:            0.092 2022-03-18 04:19:14       99 <ResampleResult[22]>
#> 100:            0.102 2022-03-18 04:19:14      100 <ResampleResult[22]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }