Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>     age glucose insulin mass pedigree pregnant pressure triceps
#> 1: TRUE    TRUE    TRUE TRUE     TRUE    FALSE     TRUE   FALSE
#>                                      features classif.ce
#> 1: age,glucose,insulin,mass,pedigree,pressure   0.203125

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1: FALSE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.2500000
#>   2:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.3242188
#>   3:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>   4: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.2382812
#>   5:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2070312
#>   6: FALSE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3281250
#>   7:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3007812
#>   8:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2187500
#>   9:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2187500
#>  10:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3398438
#>  11:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  12: FALSE    TRUE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.2226562
#>  13:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  14:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  15: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2500000
#>  16: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3710938
#>  17: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2539062
#>  18: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3671875
#>  19:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2031250
#>  20: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.3320312
#>  21:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  22: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2578125
#>  23:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  24:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2343750
#>  25:  TRUE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2460938
#>  26:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3398438
#>  27:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2031250
#>  28:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2382812
#>  29:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2187500
#>  30: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>  31:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.2343750
#>  32: FALSE    TRUE   FALSE  TRUE    FALSE     TRUE    FALSE    TRUE  0.2500000
#>  33: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.2500000
#>  34:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  35:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3007812
#>  36: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.2539062
#>  37: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.3515625
#>  38:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2031250
#>  39: FALSE   FALSE   FALSE FALSE     TRUE    FALSE     TRUE    TRUE  0.3867188
#>  40: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2500000
#>  41:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  42:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  43:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.2304688
#>  44:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  45:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE   FALSE  0.2265625
#>  46:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  47: FALSE    TRUE   FALSE FALSE     TRUE    FALSE     TRUE    TRUE  0.2617188
#>  48:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  49:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE    FALSE   FALSE  0.2187500
#>  50:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2343750
#>  51:  TRUE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3046875
#>  52:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE    FALSE    TRUE  0.2187500
#>  53: FALSE   FALSE    TRUE FALSE     TRUE    FALSE    FALSE   FALSE  0.3085938
#>  54:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2343750
#>  55:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  56:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.2421875
#>  57:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2226562
#>  58: FALSE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE    TRUE  0.2304688
#>  59:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  60: FALSE    TRUE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.2617188
#>  61: FALSE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE   FALSE  0.2500000
#>  62: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2500000
#>  63: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2500000
#>  64: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE    FALSE    TRUE  0.3515625
#>  65:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2226562
#>  66: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2421875
#>  67:  TRUE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2226562
#>  68:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  69: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3437500
#>  70: FALSE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3242188
#>  71: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3945312
#>  72: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3710938
#>  73: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2578125
#>  74:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2031250
#>  75: FALSE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3281250
#>  76:  TRUE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE   FALSE  0.3359375
#>  77:  TRUE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2226562
#>  78: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.3125000
#>  79:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  80:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  81:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2031250
#>  82: FALSE    TRUE    TRUE FALSE    FALSE    FALSE     TRUE   FALSE  0.2500000
#>  83: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2578125
#>  84: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3320312
#>  85:  TRUE   FALSE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.3242188
#>  86: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3945312
#>  87:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2265625
#>  88: FALSE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3828125
#>  89: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2500000
#>  90:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2226562
#>  91:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2031250
#>  92: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3476562
#>  93:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3242188
#>  94:  TRUE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2929688
#>  95:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2343750
#>  96: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3710938
#>  97: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2578125
#>  98: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3867188
#>  99: FALSE    TRUE    TRUE FALSE     TRUE     TRUE    FALSE    TRUE  0.2500000
#> 100:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE     TRUE    TRUE  0.2265625
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.078 2022-11-11 04:21:39        1 <ResampleResult[21]>
#>   2:            0.079 2022-11-11 04:21:39        2 <ResampleResult[21]>
#>   3:            0.065 2022-11-11 04:21:39        3 <ResampleResult[21]>
#>   4:            0.077 2022-11-11 04:21:40        4 <ResampleResult[21]>
#>   5:            0.079 2022-11-11 04:21:40        5 <ResampleResult[21]>
#>   6:            0.071 2022-11-11 04:21:40        6 <ResampleResult[21]>
#>   7:            0.076 2022-11-11 04:21:40        7 <ResampleResult[21]>
#>   8:            0.081 2022-11-11 04:21:41        8 <ResampleResult[21]>
#>   9:            0.074 2022-11-11 04:21:41        9 <ResampleResult[21]>
#>  10:            0.076 2022-11-11 04:21:41       10 <ResampleResult[21]>
#>  11:            0.081 2022-11-11 04:21:41       11 <ResampleResult[21]>
#>  12:            0.069 2022-11-11 04:21:41       12 <ResampleResult[21]>
#>  13:            0.070 2022-11-11 04:21:42       13 <ResampleResult[21]>
#>  14:            0.068 2022-11-11 04:21:42       14 <ResampleResult[21]>
#>  15:            0.065 2022-11-11 04:21:42       15 <ResampleResult[21]>
#>  16:            0.064 2022-11-11 04:21:42       16 <ResampleResult[21]>
#>  17:            0.065 2022-11-11 04:21:42       17 <ResampleResult[21]>
#>  18:            0.102 2022-11-11 04:21:43       18 <ResampleResult[21]>
#>  19:            0.068 2022-11-11 04:21:43       19 <ResampleResult[21]>
#>  20:            0.075 2022-11-11 04:21:43       20 <ResampleResult[21]>
#>  21:            0.084 2022-11-11 04:21:43       21 <ResampleResult[21]>
#>  22:            0.072 2022-11-11 04:21:44       22 <ResampleResult[21]>
#>  23:            0.091 2022-11-11 04:21:44       23 <ResampleResult[21]>
#>  24:            0.072 2022-11-11 04:21:44       24 <ResampleResult[21]>
#>  25:            0.068 2022-11-11 04:21:44       25 <ResampleResult[21]>
#>  26:            0.074 2022-11-11 04:21:45       26 <ResampleResult[21]>
#>  27:            0.085 2022-11-11 04:21:45       27 <ResampleResult[21]>
#>  28:            0.082 2022-11-11 04:21:45       28 <ResampleResult[21]>
#>  29:            0.082 2022-11-11 04:21:45       29 <ResampleResult[21]>
#>  30:            0.071 2022-11-11 04:21:46       30 <ResampleResult[21]>
#>  31:            0.069 2022-11-11 04:21:46       31 <ResampleResult[21]>
#>  32:            0.068 2022-11-11 04:21:46       32 <ResampleResult[21]>
#>  33:            0.068 2022-11-11 04:21:46       33 <ResampleResult[21]>
#>  34:            0.068 2022-11-11 04:21:46       34 <ResampleResult[21]>
#>  35:            0.085 2022-11-11 04:21:47       35 <ResampleResult[21]>
#>  36:            0.088 2022-11-11 04:21:47       36 <ResampleResult[21]>
#>  37:            0.084 2022-11-11 04:21:47       37 <ResampleResult[21]>
#>  38:            0.090 2022-11-11 04:21:47       38 <ResampleResult[21]>
#>  39:            0.065 2022-11-11 04:21:48       39 <ResampleResult[21]>
#>  40:            0.083 2022-11-11 04:21:48       40 <ResampleResult[21]>
#>  41:            0.075 2022-11-11 04:21:48       41 <ResampleResult[21]>
#>  42:            0.071 2022-11-11 04:21:48       42 <ResampleResult[21]>
#>  43:            0.085 2022-11-11 04:21:49       43 <ResampleResult[21]>
#>  44:            0.105 2022-11-11 04:21:49       44 <ResampleResult[21]>
#>  45:            0.083 2022-11-11 04:21:49       45 <ResampleResult[21]>
#>  46:            0.072 2022-11-11 04:21:49       46 <ResampleResult[21]>
#>  47:            0.069 2022-11-11 04:21:50       47 <ResampleResult[21]>
#>  48:            0.077 2022-11-11 04:21:50       48 <ResampleResult[21]>
#>  49:            0.113 2022-11-11 04:21:50       49 <ResampleResult[21]>
#>  50:            0.083 2022-11-11 04:21:50       50 <ResampleResult[21]>
#>  51:            0.073 2022-11-11 04:21:51       51 <ResampleResult[21]>
#>  52:            0.069 2022-11-11 04:21:51       52 <ResampleResult[21]>
#>  53:            0.069 2022-11-11 04:21:51       53 <ResampleResult[21]>
#>  54:            0.077 2022-11-11 04:21:51       54 <ResampleResult[21]>
#>  55:            0.087 2022-11-11 04:21:51       55 <ResampleResult[21]>
#>  56:            0.081 2022-11-11 04:21:52       56 <ResampleResult[21]>
#>  57:            0.071 2022-11-11 04:21:52       57 <ResampleResult[21]>
#>  58:            0.078 2022-11-11 04:21:52       58 <ResampleResult[21]>
#>  59:            0.085 2022-11-11 04:21:52       59 <ResampleResult[21]>
#>  60:            0.072 2022-11-11 04:21:53       60 <ResampleResult[21]>
#>  61:            0.075 2022-11-11 04:21:53       61 <ResampleResult[21]>
#>  62:            0.089 2022-11-11 04:21:53       62 <ResampleResult[21]>
#>  63:            0.080 2022-11-11 04:21:53       63 <ResampleResult[21]>
#>  64:            0.082 2022-11-11 04:21:54       64 <ResampleResult[21]>
#>  65:            0.069 2022-11-11 04:21:54       65 <ResampleResult[21]>
#>  66:            0.070 2022-11-11 04:21:54       66 <ResampleResult[21]>
#>  67:            0.076 2022-11-11 04:21:54       67 <ResampleResult[21]>
#>  68:            0.089 2022-11-11 04:21:55       68 <ResampleResult[21]>
#>  69:            0.078 2022-11-11 04:21:55       69 <ResampleResult[21]>
#>  70:            0.097 2022-11-11 04:21:55       70 <ResampleResult[21]>
#>  71:            0.087 2022-11-11 04:21:55       71 <ResampleResult[21]>
#>  72:            0.074 2022-11-11 04:21:56       72 <ResampleResult[21]>
#>  73:            0.069 2022-11-11 04:21:56       73 <ResampleResult[21]>
#>  74:            0.077 2022-11-11 04:21:56       74 <ResampleResult[21]>
#>  75:            0.099 2022-11-11 04:21:56       75 <ResampleResult[21]>
#>  76:            0.080 2022-11-11 04:21:57       76 <ResampleResult[21]>
#>  77:            0.083 2022-11-11 04:21:57       77 <ResampleResult[21]>
#>  78:            0.072 2022-11-11 04:21:57       78 <ResampleResult[21]>
#>  79:            0.075 2022-11-11 04:21:57       79 <ResampleResult[21]>
#>  80:            0.078 2022-11-11 04:21:58       80 <ResampleResult[21]>
#>  81:            0.086 2022-11-11 04:21:58       81 <ResampleResult[21]>
#>  82:            0.099 2022-11-11 04:21:58       82 <ResampleResult[21]>
#>  83:            0.102 2022-11-11 04:21:58       83 <ResampleResult[21]>
#>  84:            0.086 2022-11-11 04:21:59       84 <ResampleResult[21]>
#>  85:            0.075 2022-11-11 04:21:59       85 <ResampleResult[21]>
#>  86:            0.073 2022-11-11 04:21:59       86 <ResampleResult[21]>
#>  87:            0.071 2022-11-11 04:21:59       87 <ResampleResult[21]>
#>  88:            0.075 2022-11-11 04:22:00       88 <ResampleResult[21]>
#>  89:            0.082 2022-11-11 04:22:00       89 <ResampleResult[21]>
#>  90:            0.131 2022-11-11 04:22:00       90 <ResampleResult[21]>
#>  91:            0.110 2022-11-11 04:22:00       91 <ResampleResult[21]>
#>  92:            0.127 2022-11-11 04:22:01       92 <ResampleResult[21]>
#>  93:            0.073 2022-11-11 04:22:01       93 <ResampleResult[21]>
#>  94:            0.071 2022-11-11 04:22:01       94 <ResampleResult[21]>
#>  95:            0.069 2022-11-11 04:22:01       95 <ResampleResult[21]>
#>  96:            0.124 2022-11-11 04:22:02       96 <ResampleResult[21]>
#>  97:            0.121 2022-11-11 04:22:02       97 <ResampleResult[21]>
#>  98:            0.070 2022-11-11 04:22:02       98 <ResampleResult[21]>
#>  99:            0.078 2022-11-11 04:22:03       99 <ResampleResult[21]>
#> 100:            0.075 2022-11-11 04:22:03      100 <ResampleResult[21]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }