Feature Selection via Random Search

Random search randomly draws feature sets.

Feature sets are evaluated in batches of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This FSelector can be instantiated via the dictionary mlr_fselectors or with the associated sugar function fs():

mlr_fselectors$get("random_search")
fs("random_search")

Parameters

max_features

integer(1)
Maximum number of features. By default, number of features in mlr3::Task.

batch_size

integer(1)
Maximum number of feature sets to try in a batch.

Super class

mlr3fselect::FSelector -> FSelectorRandomSearch

Methods

Public methods

FSelectorRandomSearch$new()

FSelectorRandomSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.

Usage

FSelectorRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

# \donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "random_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 100
)

# best performing feature subset
instance$result
#>      age glucose insulin mass pedigree pregnant pressure triceps
#> 1: FALSE    TRUE   FALSE TRUE     TRUE    FALSE    FALSE    TRUE
#>                         features classif.ce
#> 1: glucose,mass,pedigree,triceps   0.203125

# all evaluated feature subsets
as.data.table(instance$archive)
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>   1: FALSE   FALSE   FALSE FALSE    FALSE     TRUE     TRUE    TRUE  0.3789062
#>   2:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.3281250
#>   3: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2343750
#>   4: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3281250
#>   5: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3164062
#>   6: FALSE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE    TRUE  0.2343750
#>   7: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3750000
#>   8: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3750000
#>   9: FALSE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE    TRUE  0.2343750
#>  10: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3750000
#>  11: FALSE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2382812
#>  12:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE    FALSE   FALSE  0.2656250
#>  13: FALSE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE    TRUE  0.2343750
#>  14: FALSE   FALSE   FALSE FALSE     TRUE    FALSE    FALSE   FALSE  0.3281250
#>  15:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  16:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  17:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  18: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.2304688
#>  19: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3750000
#>  20: FALSE   FALSE    TRUE FALSE    FALSE    FALSE     TRUE    TRUE  0.3476562
#>  21: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE    TRUE  0.3789062
#>  22:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2656250
#>  23:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2656250
#>  24:  TRUE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3671875
#>  25:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  26: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.2304688
#>  27:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2656250
#>  28: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  29: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE    TRUE  0.3750000
#>  30:  TRUE   FALSE    TRUE FALSE    FALSE     TRUE     TRUE   FALSE  0.3593750
#>  31: FALSE   FALSE   FALSE FALSE    FALSE    FALSE    FALSE    TRUE  0.3359375
#>  32: FALSE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3750000
#>  33:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  34:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2656250
#>  35: FALSE    TRUE   FALSE FALSE     TRUE     TRUE     TRUE    TRUE  0.2382812
#>  36:  TRUE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.3554688
#>  37:  TRUE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE    TRUE  0.2656250
#>  38:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  39: FALSE    TRUE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.2382812
#>  40: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  41: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3554688
#>  42: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2031250
#>  43: FALSE    TRUE    TRUE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2304688
#>  44: FALSE   FALSE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.3203125
#>  45:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  46: FALSE   FALSE    TRUE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3593750
#>  47:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE   FALSE  0.3828125
#>  48: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2304688
#>  49:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  50:  TRUE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2460938
#>  51: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.2343750
#>  52: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  53: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3398438
#>  54: FALSE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>  55:  TRUE   FALSE   FALSE FALSE    FALSE    FALSE     TRUE    TRUE  0.3828125
#>  56: FALSE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3554688
#>  57:  TRUE    TRUE   FALSE FALSE     TRUE    FALSE    FALSE    TRUE  0.2734375
#>  58:  TRUE    TRUE    TRUE FALSE     TRUE    FALSE    FALSE   FALSE  0.2460938
#>  59: FALSE   FALSE   FALSE FALSE    FALSE     TRUE    FALSE    TRUE  0.3359375
#>  60: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE     TRUE   FALSE  0.3203125
#>  61:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  62: FALSE    TRUE   FALSE FALSE     TRUE     TRUE    FALSE   FALSE  0.2500000
#>  63:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  64:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  65: FALSE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3164062
#>  66: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE    FALSE   FALSE  0.3789062
#>  67:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  68:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  69: FALSE    TRUE   FALSE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2031250
#>  70:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  71: FALSE    TRUE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.2304688
#>  72: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3320312
#>  73: FALSE   FALSE    TRUE FALSE    FALSE     TRUE    FALSE   FALSE  0.3242188
#>  74:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>  75: FALSE   FALSE   FALSE  TRUE    FALSE     TRUE     TRUE   FALSE  0.3789062
#>  76:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  77:  TRUE    TRUE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2656250
#>  78:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  79:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  80: FALSE   FALSE    TRUE  TRUE     TRUE    FALSE    FALSE    TRUE  0.2890625
#>  81:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  82:  TRUE   FALSE    TRUE  TRUE     TRUE    FALSE     TRUE   FALSE  0.3554688
#>  83:  TRUE   FALSE    TRUE  TRUE    FALSE    FALSE    FALSE    TRUE  0.3476562
#>  84:  TRUE    TRUE   FALSE FALSE    FALSE     TRUE     TRUE    TRUE  0.2734375
#>  85:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE     TRUE   FALSE  0.3437500
#>  86: FALSE    TRUE   FALSE FALSE    FALSE    FALSE    FALSE   FALSE  0.2578125
#>  87: FALSE    TRUE   FALSE FALSE    FALSE     TRUE    FALSE   FALSE  0.2500000
#>  88:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  89:  TRUE    TRUE    TRUE  TRUE    FALSE    FALSE     TRUE    TRUE  0.2851562
#>  90:  TRUE   FALSE   FALSE  TRUE     TRUE     TRUE     TRUE   FALSE  0.3554688
#>  91:  TRUE   FALSE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.3281250
#>  92: FALSE   FALSE   FALSE FALSE     TRUE     TRUE    FALSE    TRUE  0.3242188
#>  93:  TRUE    TRUE   FALSE  TRUE    FALSE     TRUE     TRUE    TRUE  0.2617188
#>  94: FALSE   FALSE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.3593750
#>  95: FALSE    TRUE   FALSE  TRUE    FALSE    FALSE    FALSE   FALSE  0.2421875
#>  96:  TRUE   FALSE    TRUE  TRUE    FALSE     TRUE     TRUE    TRUE  0.3476562
#>  97:  TRUE    TRUE    TRUE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2656250
#>  98:  TRUE   FALSE    TRUE FALSE    FALSE    FALSE    FALSE   FALSE  0.3437500
#>  99: FALSE    TRUE   FALSE  TRUE     TRUE     TRUE     TRUE    TRUE  0.2343750
#> 100:  TRUE    TRUE    TRUE FALSE     TRUE     TRUE     TRUE    TRUE  0.2460938
#>        age glucose insulin  mass pedigree pregnant pressure triceps classif.ce
#>      runtime_learners           timestamp batch_nr      resample_result
#>   1:            0.103 2022-03-16 04:16:14        1 <ResampleResult[22]>
#>   2:            0.098 2022-03-16 04:16:14        2 <ResampleResult[22]>
#>   3:            0.088 2022-03-16 04:16:14        3 <ResampleResult[22]>
#>   4:            0.072 2022-03-16 04:16:15        4 <ResampleResult[22]>
#>   5:            0.072 2022-03-16 04:16:15        5 <ResampleResult[22]>
#>   6:            0.071 2022-03-16 04:16:15        6 <ResampleResult[22]>
#>   7:            0.073 2022-03-16 04:16:15        7 <ResampleResult[22]>
#>   8:            0.071 2022-03-16 04:16:16        8 <ResampleResult[22]>
#>   9:            0.075 2022-03-16 04:16:16        9 <ResampleResult[22]>
#>  10:            0.073 2022-03-16 04:16:16       10 <ResampleResult[22]>
#>  11:            0.077 2022-03-16 04:16:16       11 <ResampleResult[22]>
#>  12:            0.082 2022-03-16 04:16:17       12 <ResampleResult[22]>
#>  13:            0.085 2022-03-16 04:16:17       13 <ResampleResult[22]>
#>  14:            0.083 2022-03-16 04:16:17       14 <ResampleResult[22]>
#>  15:            0.086 2022-03-16 04:16:17       15 <ResampleResult[22]>
#>  16:            0.089 2022-03-16 04:16:18       16 <ResampleResult[22]>
#>  17:            0.101 2022-03-16 04:16:18       17 <ResampleResult[22]>
#>  18:            0.099 2022-03-16 04:16:18       18 <ResampleResult[22]>
#>  19:            0.088 2022-03-16 04:16:18       19 <ResampleResult[22]>
#>  20:            0.072 2022-03-16 04:16:19       20 <ResampleResult[22]>
#>  21:            0.074 2022-03-16 04:16:19       21 <ResampleResult[22]>
#>  22:            0.074 2022-03-16 04:16:19       22 <ResampleResult[22]>
#>  23:            0.073 2022-03-16 04:16:19       23 <ResampleResult[22]>
#>  24:            0.078 2022-03-16 04:16:20       24 <ResampleResult[22]>
#>  25:            0.086 2022-03-16 04:16:20       25 <ResampleResult[22]>
#>  26:            0.092 2022-03-16 04:16:20       26 <ResampleResult[22]>
#>  27:            0.099 2022-03-16 04:16:20       27 <ResampleResult[22]>
#>  28:            0.103 2022-03-16 04:16:21       28 <ResampleResult[22]>
#>  29:            0.096 2022-03-16 04:16:21       29 <ResampleResult[22]>
#>  30:            0.091 2022-03-16 04:16:21       30 <ResampleResult[22]>
#>  31:            0.086 2022-03-16 04:16:21       31 <ResampleResult[22]>
#>  32:            0.290 2022-03-16 04:16:22       32 <ResampleResult[22]>
#>  33:            0.071 2022-03-16 04:16:22       33 <ResampleResult[22]>
#>  34:            0.070 2022-03-16 04:16:22       34 <ResampleResult[22]>
#>  35:            0.067 2022-03-16 04:16:22       35 <ResampleResult[22]>
#>  36:            0.077 2022-03-16 04:16:23       36 <ResampleResult[22]>
#>  37:            0.077 2022-03-16 04:16:23       37 <ResampleResult[22]>
#>  38:            0.069 2022-03-16 04:16:23       38 <ResampleResult[22]>
#>  39:            0.070 2022-03-16 04:16:23       39 <ResampleResult[22]>
#>  40:            0.066 2022-03-16 04:16:24       40 <ResampleResult[22]>
#>  41:            0.075 2022-03-16 04:16:24       41 <ResampleResult[22]>
#>  42:            0.093 2022-03-16 04:16:24       42 <ResampleResult[22]>
#>  43:            0.069 2022-03-16 04:16:24       43 <ResampleResult[22]>
#>  44:            0.068 2022-03-16 04:16:24       44 <ResampleResult[22]>
#>  45:            0.068 2022-03-16 04:16:25       45 <ResampleResult[22]>
#>  46:            0.070 2022-03-16 04:16:25       46 <ResampleResult[22]>
#>  47:            0.082 2022-03-16 04:16:25       47 <ResampleResult[22]>
#>  48:            0.069 2022-03-16 04:16:25       48 <ResampleResult[22]>
#>  49:            0.071 2022-03-16 04:16:25       49 <ResampleResult[22]>
#>  50:            0.066 2022-03-16 04:16:26       50 <ResampleResult[22]>
#>  51:            0.068 2022-03-16 04:16:26       51 <ResampleResult[22]>
#>  52:            0.081 2022-03-16 04:16:26       52 <ResampleResult[22]>
#>  53:            0.080 2022-03-16 04:16:26       53 <ResampleResult[22]>
#>  54:            0.067 2022-03-16 04:16:26       54 <ResampleResult[22]>
#>  55:            0.068 2022-03-16 04:16:27       55 <ResampleResult[22]>
#>  56:            0.068 2022-03-16 04:16:27       56 <ResampleResult[22]>
#>  57:            0.067 2022-03-16 04:16:27       57 <ResampleResult[22]>
#>  58:            0.072 2022-03-16 04:16:27       58 <ResampleResult[22]>
#>  59:            0.078 2022-03-16 04:16:27       59 <ResampleResult[22]>
#>  60:            0.081 2022-03-16 04:16:28       60 <ResampleResult[22]>
#>  61:            0.070 2022-03-16 04:16:28       61 <ResampleResult[22]>
#>  62:            0.069 2022-03-16 04:16:28       62 <ResampleResult[22]>
#>  63:            0.069 2022-03-16 04:16:28       63 <ResampleResult[22]>
#>  64:            0.068 2022-03-16 04:16:29       64 <ResampleResult[22]>
#>  65:            0.066 2022-03-16 04:16:29       65 <ResampleResult[22]>
#>  66:            0.066 2022-03-16 04:16:29       66 <ResampleResult[22]>
#>  67:            0.079 2022-03-16 04:16:29       67 <ResampleResult[22]>
#>  68:            0.085 2022-03-16 04:16:29       68 <ResampleResult[22]>
#>  69:            0.069 2022-03-16 04:16:30       69 <ResampleResult[22]>
#>  70:            0.071 2022-03-16 04:16:30       70 <ResampleResult[22]>
#>  71:            0.069 2022-03-16 04:16:30       71 <ResampleResult[22]>
#>  72:            0.076 2022-03-16 04:16:30       72 <ResampleResult[22]>
#>  73:            0.088 2022-03-16 04:16:31       73 <ResampleResult[22]>
#>  74:            0.069 2022-03-16 04:16:31       74 <ResampleResult[22]>
#>  75:            0.070 2022-03-16 04:16:31       75 <ResampleResult[22]>
#>  76:            0.070 2022-03-16 04:16:31       76 <ResampleResult[22]>
#>  77:            0.069 2022-03-16 04:16:31       77 <ResampleResult[22]>
#>  78:            0.089 2022-03-16 04:16:32       78 <ResampleResult[22]>
#>  79:            0.090 2022-03-16 04:16:32       79 <ResampleResult[22]>
#>  80:            0.068 2022-03-16 04:16:32       80 <ResampleResult[22]>
#>  81:            0.071 2022-03-16 04:16:32       81 <ResampleResult[22]>
#>  82:            0.070 2022-03-16 04:16:32       82 <ResampleResult[22]>
#>  83:            0.070 2022-03-16 04:16:33       83 <ResampleResult[22]>
#>  84:            0.071 2022-03-16 04:16:33       84 <ResampleResult[22]>
#>  85:            0.085 2022-03-16 04:16:33       85 <ResampleResult[22]>
#>  86:            0.087 2022-03-16 04:16:33       86 <ResampleResult[22]>
#>  87:            0.085 2022-03-16 04:16:34       87 <ResampleResult[22]>
#>  88:            0.074 2022-03-16 04:16:34       88 <ResampleResult[22]>
#>  89:            0.072 2022-03-16 04:16:34       89 <ResampleResult[22]>
#>  90:            0.071 2022-03-16 04:16:34       90 <ResampleResult[22]>
#>  91:            0.071 2022-03-16 04:16:35       91 <ResampleResult[22]>
#>  92:            0.069 2022-03-16 04:16:35       92 <ResampleResult[22]>
#>  93:            0.085 2022-03-16 04:16:35       93 <ResampleResult[22]>
#>  94:            0.090 2022-03-16 04:16:35       94 <ResampleResult[22]>
#>  95:            0.070 2022-03-16 04:16:35       95 <ResampleResult[22]>
#>  96:            0.071 2022-03-16 04:16:36       96 <ResampleResult[22]>
#>  97:            0.070 2022-03-16 04:16:36       97 <ResampleResult[22]>
#>  98:            0.081 2022-03-16 04:16:36       98 <ResampleResult[22]>
#>  99:            0.098 2022-03-16 04:16:36       99 <ResampleResult[22]>
#> 100:            0.083 2022-03-16 04:16:37      100 <ResampleResult[22]>
#>      runtime_learners           timestamp batch_nr      resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }