Feature Selection with Shadow Variable Search

Feature selection using the Shadow Variable Search Algorithm. Shadow variable search creates for each feature a permutated copy and stops when one of them is selected.

Source

Thomas J, Hepp T, Mayr A, Bischl B (2017). “Probing for Sparse and Fast Variable Selection with Model-Based Boosting.” Computational and Mathematical Methods in Medicine, 2017, 1--8. doi:10.1155/2017/1421409 .

Wu Y, Boos DD, Stefanski LA (2007). “Controlling Variable Selection by the Addition of Pseudovariables.” Journal of the American Statistical Association, 102(477), 235--243. doi:10.1198/016214506000000843 .

Details

The feature selection terminates itself when the first shadow variable is selected. It is not necessary to set a termination criterion.

Resources

The gallery features a collection of case studies and demos about optimization.

Run a feature selection with Shadow Variable Search.

Dictionary

This FSelector can be instantiated with the associated sugar function fs():

fs("shadow_variable_search")

See also

Super class

mlr3fselect::FSelector -> FSelectorShadowVariableSearch

Methods

Public methods

FSelectorShadowVariableSearch$new()

FSelectorShadowVariableSearch$optimization_path()

FSelectorShadowVariableSearch$clone()

Inherited methods

mlr3fselect::FSelector$format()

mlr3fselect::FSelector$help()

mlr3fselect::FSelector$optimize()

mlr3fselect::FSelector$print()

Method new()

Creates a new instance of this R6 class.`

Usage

FSelectorShadowVariableSearch$new()

Method optimization_path()

Returns the optimization path.

Usage

FSelectorShadowVariableSearch$optimization_path(inst)

Arguments

inst

(FSelectInstanceSingleCrit)
Instance optimized with FSelectorShadowVariableSearch.

Returns

data.table::data.table

Method clone()

The objects of this class are cloneable with this method.

Usage

FSelectorShadowVariableSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Feature Selection
# \donttest{

# retrieve task and load learner
task = tsk("penguins")
learner = lrn("classif.rpart")

# run feature selection on the Palmer Penguins data set
instance = fselect(
  fselector = fs("shadow_variable_search"),
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
)

# best performing feature subset
instance$result
#>    bill_depth bill_length body_mass flipper_length island   sex  year
#> 1:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#>                            features classif.ce
#> 1: bill_depth,bill_length,body_mass 0.03478261

# all evaluated feature subsets
as.data.table(instance$archive)
#>     bill_depth bill_length body_mass flipper_length island   sex  year
#>  1:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#>  2:      FALSE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#>  3:      FALSE       FALSE      TRUE          FALSE  FALSE FALSE FALSE
#>  4:      FALSE       FALSE     FALSE           TRUE  FALSE FALSE FALSE
#>  5:      FALSE       FALSE     FALSE          FALSE   TRUE FALSE FALSE
#>  6:      FALSE       FALSE     FALSE          FALSE  FALSE  TRUE FALSE
#>  7:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE  TRUE
#>  8:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#>  9:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 10:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 11:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 12:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 13:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 14:      FALSE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 15:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 16:       TRUE       FALSE      TRUE          FALSE  FALSE FALSE FALSE
#> 17:       TRUE       FALSE     FALSE           TRUE  FALSE FALSE FALSE
#> 18:       TRUE       FALSE     FALSE          FALSE   TRUE FALSE FALSE
#> 19:       TRUE       FALSE     FALSE          FALSE  FALSE  TRUE FALSE
#> 20:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE  TRUE
#> 21:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 22:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 23:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 24:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 25:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 26:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 27:       TRUE       FALSE     FALSE          FALSE  FALSE FALSE FALSE
#> 28:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 29:       TRUE        TRUE     FALSE           TRUE  FALSE FALSE FALSE
#> 30:       TRUE        TRUE     FALSE          FALSE   TRUE FALSE FALSE
#> 31:       TRUE        TRUE     FALSE          FALSE  FALSE  TRUE FALSE
#> 32:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE  TRUE
#> 33:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 34:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 35:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 36:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 37:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 38:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 39:       TRUE        TRUE     FALSE          FALSE  FALSE FALSE FALSE
#> 40:       TRUE        TRUE      TRUE           TRUE  FALSE FALSE FALSE
#> 41:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 42:       TRUE        TRUE      TRUE          FALSE  FALSE  TRUE FALSE
#> 43:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE  TRUE
#> 44:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 45:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 46:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 47:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 48:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 49:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 50:       TRUE        TRUE      TRUE          FALSE  FALSE FALSE FALSE
#> 51:       TRUE        TRUE      TRUE           TRUE   TRUE FALSE FALSE
#> 52:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 53:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE  TRUE
#> 54:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 55:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 56:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 57:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 58:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 59:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 60:       TRUE        TRUE      TRUE          FALSE   TRUE FALSE FALSE
#> 61:       TRUE        TRUE      TRUE           TRUE   TRUE  TRUE FALSE
#> 62:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE  TRUE
#> 63:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 64:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 65:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 66:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 67:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 68:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#> 69:       TRUE        TRUE      TRUE          FALSE   TRUE  TRUE FALSE
#>     bill_depth bill_length body_mass flipper_length island   sex  year
#>     classif.ce runtime_learners           timestamp batch_nr
#>  1: 0.22608696            0.033 2023-03-02 10:47:26        1
#>  2: 0.23478261            0.027 2023-03-02 10:47:26        1
#>  3: 0.26086957            0.024 2023-03-02 10:47:26        1
#>  4: 0.26086957            0.027 2023-03-02 10:47:26        1
#>  5: 0.28695652            0.026 2023-03-02 10:47:26        1
#>  6: 0.55652174            0.025 2023-03-02 10:47:26        1
#>  7: 0.55652174            0.026 2023-03-02 10:47:26        1
#>  8: 0.53913043            0.059 2023-03-02 10:47:26        1
#>  9: 0.70434783            0.028 2023-03-02 10:47:26        1
#> 10: 0.60869565            0.025 2023-03-02 10:47:26        1
#> 11: 0.53913043            0.022 2023-03-02 10:47:26        1
#> 12: 0.57391304            0.025 2023-03-02 10:47:26        1
#> 13: 0.55652174            0.023 2023-03-02 10:47:26        1
#> 14: 0.55652174            0.021 2023-03-02 10:47:26        1
#> 15: 0.06086957            0.025 2023-03-02 10:47:27        2
#> 16: 0.21739130            0.030 2023-03-02 10:47:27        2
#> 17: 0.23478261            0.030 2023-03-02 10:47:27        2
#> 18: 0.20000000            0.026 2023-03-02 10:47:27        2
#> 19: 0.22608696            0.027 2023-03-02 10:47:27        2
#> 20: 0.22608696            0.028 2023-03-02 10:47:27        2
#> 21: 0.22608696            0.026 2023-03-02 10:47:27        2
#> 22: 0.22608696            0.026 2023-03-02 10:47:27        2
#> 23: 0.22608696            0.025 2023-03-02 10:47:27        2
#> 24: 0.22608696            0.032 2023-03-02 10:47:27        2
#> 25: 0.22608696            0.030 2023-03-02 10:47:27        2
#> 26: 0.22608696            0.030 2023-03-02 10:47:27        2
#> 27: 0.23478261            0.026 2023-03-02 10:47:27        2
#> 28: 0.03478261            0.028 2023-03-02 10:47:28        3
#> 29: 0.07826087            0.028 2023-03-02 10:47:28        3
#> 30: 0.03478261            0.027 2023-03-02 10:47:28        3
#> 31: 0.06086957            0.031 2023-03-02 10:47:28        3
#> 32: 0.06086957            0.063 2023-03-02 10:47:28        3
#> 33: 0.06086957            0.032 2023-03-02 10:47:28        3
#> 34: 0.06086957            0.030 2023-03-02 10:47:28        3
#> 35: 0.06956522            0.027 2023-03-02 10:47:28        3
#> 36: 0.06086957            0.026 2023-03-02 10:47:28        3
#> 37: 0.06086957            0.029 2023-03-02 10:47:28        3
#> 38: 0.06086957            0.030 2023-03-02 10:47:28        3
#> 39: 0.06956522            0.027 2023-03-02 10:47:28        3
#> 40: 0.07826087            0.025 2023-03-02 10:47:29        4
#> 41: 0.03478261            0.057 2023-03-02 10:47:29        4
#> 42: 0.03478261            0.035 2023-03-02 10:47:29        4
#> 43: 0.03478261            0.034 2023-03-02 10:47:29        4
#> 44: 0.03478261            0.026 2023-03-02 10:47:29        4
#> 45: 0.03478261            0.028 2023-03-02 10:47:29        4
#> 46: 0.04347826            0.026 2023-03-02 10:47:29        4
#> 47: 0.03478261            0.025 2023-03-02 10:47:29        4
#> 48: 0.03478261            0.025 2023-03-02 10:47:29        4
#> 49: 0.03478261            0.029 2023-03-02 10:47:29        4
#> 50: 0.04347826            0.033 2023-03-02 10:47:29        4
#> 51: 0.06956522            0.034 2023-03-02 10:47:30        5
#> 52: 0.03478261            0.034 2023-03-02 10:47:30        5
#> 53: 0.03478261            0.026 2023-03-02 10:47:30        5
#> 54: 0.03478261            0.031 2023-03-02 10:47:30        5
#> 55: 0.03478261            0.028 2023-03-02 10:47:30        5
#> 56: 0.03478261            0.029 2023-03-02 10:47:30        5
#> 57: 0.03478261            0.029 2023-03-02 10:47:30        5
#> 58: 0.03478261            0.067 2023-03-02 10:47:30        5
#> 59: 0.03478261            0.040 2023-03-02 10:47:30        5
#> 60: 0.03478261            0.033 2023-03-02 10:47:30        5
#> 61: 0.06956522            0.033 2023-03-02 10:47:31        6
#> 62: 0.03478261            0.031 2023-03-02 10:47:31        6
#> 63: 0.03478261            0.026 2023-03-02 10:47:31        6
#> 64: 0.03478261            0.027 2023-03-02 10:47:31        6
#> 65: 0.03478261            0.028 2023-03-02 10:47:31        6
#> 66: 0.03478261            0.027 2023-03-02 10:47:31        6
#> 67: 0.03478261            0.058 2023-03-02 10:47:31        6
#> 68: 0.03478261            0.033 2023-03-02 10:47:31        6
#> 69: 0.03478261            0.027 2023-03-02 10:47:31        6
#>     classif.ce runtime_learners           timestamp batch_nr
#>     permuted__bill_depth permuted__bill_length permuted__body_mass
#>  1:                FALSE                 FALSE               FALSE
#>  2:                FALSE                 FALSE               FALSE
#>  3:                FALSE                 FALSE               FALSE
#>  4:                FALSE                 FALSE               FALSE
#>  5:                FALSE                 FALSE               FALSE
#>  6:                FALSE                 FALSE               FALSE
#>  7:                FALSE                 FALSE               FALSE
#>  8:                 TRUE                 FALSE               FALSE
#>  9:                FALSE                  TRUE               FALSE
#> 10:                FALSE                 FALSE                TRUE
#> 11:                FALSE                 FALSE               FALSE
#> 12:                FALSE                 FALSE               FALSE
#> 13:                FALSE                 FALSE               FALSE
#> 14:                FALSE                 FALSE               FALSE
#> 15:                FALSE                 FALSE               FALSE
#> 16:                FALSE                 FALSE               FALSE
#> 17:                FALSE                 FALSE               FALSE
#> 18:                FALSE                 FALSE               FALSE
#> 19:                FALSE                 FALSE               FALSE
#> 20:                FALSE                 FALSE               FALSE
#> 21:                 TRUE                 FALSE               FALSE
#> 22:                FALSE                  TRUE               FALSE
#> 23:                FALSE                 FALSE                TRUE
#> 24:                FALSE                 FALSE               FALSE
#> 25:                FALSE                 FALSE               FALSE
#> 26:                FALSE                 FALSE               FALSE
#> 27:                FALSE                 FALSE               FALSE
#> 28:                FALSE                 FALSE               FALSE
#> 29:                FALSE                 FALSE               FALSE
#> 30:                FALSE                 FALSE               FALSE
#> 31:                FALSE                 FALSE               FALSE
#> 32:                FALSE                 FALSE               FALSE
#> 33:                 TRUE                 FALSE               FALSE
#> 34:                FALSE                  TRUE               FALSE
#> 35:                FALSE                 FALSE                TRUE
#> 36:                FALSE                 FALSE               FALSE
#> 37:                FALSE                 FALSE               FALSE
#> 38:                FALSE                 FALSE               FALSE
#> 39:                FALSE                 FALSE               FALSE
#> 40:                FALSE                 FALSE               FALSE
#> 41:                FALSE                 FALSE               FALSE
#> 42:                FALSE                 FALSE               FALSE
#> 43:                FALSE                 FALSE               FALSE
#> 44:                 TRUE                 FALSE               FALSE
#> 45:                FALSE                  TRUE               FALSE
#> 46:                FALSE                 FALSE                TRUE
#> 47:                FALSE                 FALSE               FALSE
#> 48:                FALSE                 FALSE               FALSE
#> 49:                FALSE                 FALSE               FALSE
#> 50:                FALSE                 FALSE               FALSE
#> 51:                FALSE                 FALSE               FALSE
#> 52:                FALSE                 FALSE               FALSE
#> 53:                FALSE                 FALSE               FALSE
#> 54:                 TRUE                 FALSE               FALSE
#> 55:                FALSE                  TRUE               FALSE
#> 56:                FALSE                 FALSE                TRUE
#> 57:                FALSE                 FALSE               FALSE
#> 58:                FALSE                 FALSE               FALSE
#> 59:                FALSE                 FALSE               FALSE
#> 60:                FALSE                 FALSE               FALSE
#> 61:                FALSE                 FALSE               FALSE
#> 62:                FALSE                 FALSE               FALSE
#> 63:                 TRUE                 FALSE               FALSE
#> 64:                FALSE                  TRUE               FALSE
#> 65:                FALSE                 FALSE                TRUE
#> 66:                FALSE                 FALSE               FALSE
#> 67:                FALSE                 FALSE               FALSE
#> 68:                FALSE                 FALSE               FALSE
#> 69:                FALSE                 FALSE               FALSE
#>     permuted__bill_depth permuted__bill_length permuted__body_mass
#>     permuted__flipper_length permuted__island permuted__sex permuted__year
#>  1:                    FALSE            FALSE         FALSE          FALSE
#>  2:                    FALSE            FALSE         FALSE          FALSE
#>  3:                    FALSE            FALSE         FALSE          FALSE
#>  4:                    FALSE            FALSE         FALSE          FALSE
#>  5:                    FALSE            FALSE         FALSE          FALSE
#>  6:                    FALSE            FALSE         FALSE          FALSE
#>  7:                    FALSE            FALSE         FALSE          FALSE
#>  8:                    FALSE            FALSE         FALSE          FALSE
#>  9:                    FALSE            FALSE         FALSE          FALSE
#> 10:                    FALSE            FALSE         FALSE          FALSE
#> 11:                     TRUE            FALSE         FALSE          FALSE
#> 12:                    FALSE             TRUE         FALSE          FALSE
#> 13:                    FALSE            FALSE          TRUE          FALSE
#> 14:                    FALSE            FALSE         FALSE           TRUE
#> 15:                    FALSE            FALSE         FALSE          FALSE
#> 16:                    FALSE            FALSE         FALSE          FALSE
#> 17:                    FALSE            FALSE         FALSE          FALSE
#> 18:                    FALSE            FALSE         FALSE          FALSE
#> 19:                    FALSE            FALSE         FALSE          FALSE
#> 20:                    FALSE            FALSE         FALSE          FALSE
#> 21:                    FALSE            FALSE         FALSE          FALSE
#> 22:                    FALSE            FALSE         FALSE          FALSE
#> 23:                    FALSE            FALSE         FALSE          FALSE
#> 24:                     TRUE            FALSE         FALSE          FALSE
#> 25:                    FALSE             TRUE         FALSE          FALSE
#> 26:                    FALSE            FALSE          TRUE          FALSE
#> 27:                    FALSE            FALSE         FALSE           TRUE
#> 28:                    FALSE            FALSE         FALSE          FALSE
#> 29:                    FALSE            FALSE         FALSE          FALSE
#> 30:                    FALSE            FALSE         FALSE          FALSE
#> 31:                    FALSE            FALSE         FALSE          FALSE
#> 32:                    FALSE            FALSE         FALSE          FALSE
#> 33:                    FALSE            FALSE         FALSE          FALSE
#> 34:                    FALSE            FALSE         FALSE          FALSE
#> 35:                    FALSE            FALSE         FALSE          FALSE
#> 36:                     TRUE            FALSE         FALSE          FALSE
#> 37:                    FALSE             TRUE         FALSE          FALSE
#> 38:                    FALSE            FALSE          TRUE          FALSE
#> 39:                    FALSE            FALSE         FALSE           TRUE
#> 40:                    FALSE            FALSE         FALSE          FALSE
#> 41:                    FALSE            FALSE         FALSE          FALSE
#> 42:                    FALSE            FALSE         FALSE          FALSE
#> 43:                    FALSE            FALSE         FALSE          FALSE
#> 44:                    FALSE            FALSE         FALSE          FALSE
#> 45:                    FALSE            FALSE         FALSE          FALSE
#> 46:                    FALSE            FALSE         FALSE          FALSE
#> 47:                     TRUE            FALSE         FALSE          FALSE
#> 48:                    FALSE             TRUE         FALSE          FALSE
#> 49:                    FALSE            FALSE          TRUE          FALSE
#> 50:                    FALSE            FALSE         FALSE           TRUE
#> 51:                    FALSE            FALSE         FALSE          FALSE
#> 52:                    FALSE            FALSE         FALSE          FALSE
#> 53:                    FALSE            FALSE         FALSE          FALSE
#> 54:                    FALSE            FALSE         FALSE          FALSE
#> 55:                    FALSE            FALSE         FALSE          FALSE
#> 56:                    FALSE            FALSE         FALSE          FALSE
#> 57:                     TRUE            FALSE         FALSE          FALSE
#> 58:                    FALSE             TRUE         FALSE          FALSE
#> 59:                    FALSE            FALSE          TRUE          FALSE
#> 60:                    FALSE            FALSE         FALSE           TRUE
#> 61:                    FALSE            FALSE         FALSE          FALSE
#> 62:                    FALSE            FALSE         FALSE          FALSE
#> 63:                    FALSE            FALSE         FALSE          FALSE
#> 64:                    FALSE            FALSE         FALSE          FALSE
#> 65:                    FALSE            FALSE         FALSE          FALSE
#> 66:                     TRUE            FALSE         FALSE          FALSE
#> 67:                    FALSE             TRUE         FALSE          FALSE
#> 68:                    FALSE            FALSE          TRUE          FALSE
#> 69:                    FALSE            FALSE         FALSE           TRUE
#>     permuted__flipper_length permuted__island permuted__sex permuted__year
#>     warnings errors                                                   features
#>  1:        0      0                                                 bill_depth
#>  2:        0      0                                                bill_length
#>  3:        0      0                                                  body_mass
#>  4:        0      0                                             flipper_length
#>  5:        0      0                                                     island
#>  6:        0      0                                                        sex
#>  7:        0      0                                                       year
#>  8:        0      0                                                           
#>  9:        0      0                                                           
#> 10:        0      0                                                           
#> 11:        0      0                                                           
#> 12:        0      0                                                           
#> 13:        0      0                                                           
#> 14:        0      0                                                           
#> 15:        0      0                                     bill_depth,bill_length
#> 16:        0      0                                       bill_depth,body_mass
#> 17:        0      0                                  bill_depth,flipper_length
#> 18:        0      0                                          bill_depth,island
#> 19:        0      0                                             bill_depth,sex
#> 20:        0      0                                            bill_depth,year
#> 21:        0      0                                                 bill_depth
#> 22:        0      0                                                 bill_depth
#> 23:        0      0                                                 bill_depth
#> 24:        0      0                                                 bill_depth
#> 25:        0      0                                                 bill_depth
#> 26:        0      0                                                 bill_depth
#> 27:        0      0                                                 bill_depth
#> 28:        0      0                           bill_depth,bill_length,body_mass
#> 29:        0      0                      bill_depth,bill_length,flipper_length
#> 30:        0      0                              bill_depth,bill_length,island
#> 31:        0      0                                 bill_depth,bill_length,sex
#> 32:        0      0                                bill_depth,bill_length,year
#> 33:        0      0                                     bill_depth,bill_length
#> 34:        0      0                                     bill_depth,bill_length
#> 35:        0      0                                     bill_depth,bill_length
#> 36:        0      0                                     bill_depth,bill_length
#> 37:        0      0                                     bill_depth,bill_length
#> 38:        0      0                                     bill_depth,bill_length
#> 39:        0      0                                     bill_depth,bill_length
#> 40:        0      0            bill_depth,bill_length,body_mass,flipper_length
#> 41:        0      0                    bill_depth,bill_length,body_mass,island
#> 42:        0      0                       bill_depth,bill_length,body_mass,sex
#> 43:        0      0                      bill_depth,bill_length,body_mass,year
#> 44:        0      0                           bill_depth,bill_length,body_mass
#> 45:        0      0                           bill_depth,bill_length,body_mass
#> 46:        0      0                           bill_depth,bill_length,body_mass
#> 47:        0      0                           bill_depth,bill_length,body_mass
#> 48:        0      0                           bill_depth,bill_length,body_mass
#> 49:        0      0                           bill_depth,bill_length,body_mass
#> 50:        0      0                           bill_depth,bill_length,body_mass
#> 51:        0      0     bill_depth,bill_length,body_mass,flipper_length,island
#> 52:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 53:        0      0               bill_depth,bill_length,body_mass,island,year
#> 54:        0      0                    bill_depth,bill_length,body_mass,island
#> 55:        0      0                    bill_depth,bill_length,body_mass,island
#> 56:        0      0                    bill_depth,bill_length,body_mass,island
#> 57:        0      0                    bill_depth,bill_length,body_mass,island
#> 58:        0      0                    bill_depth,bill_length,body_mass,island
#> 59:        0      0                    bill_depth,bill_length,body_mass,island
#> 60:        0      0                    bill_depth,bill_length,body_mass,island
#> 61:        0      0 bill_depth,bill_length,body_mass,flipper_length,island,sex
#> 62:        0      0           bill_depth,bill_length,body_mass,island,sex,year
#> 63:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 64:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 65:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 66:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 67:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 68:        0      0                bill_depth,bill_length,body_mass,island,sex
#> 69:        0      0                bill_depth,bill_length,body_mass,island,sex
#>     warnings errors                                                   features
#>          resample_result
#>  1: <ResampleResult[21]>
#>  2: <ResampleResult[21]>
#>  3: <ResampleResult[21]>
#>  4: <ResampleResult[21]>
#>  5: <ResampleResult[21]>
#>  6: <ResampleResult[21]>
#>  7: <ResampleResult[21]>
#>  8: <ResampleResult[21]>
#>  9: <ResampleResult[21]>
#> 10: <ResampleResult[21]>
#> 11: <ResampleResult[21]>
#> 12: <ResampleResult[21]>
#> 13: <ResampleResult[21]>
#> 14: <ResampleResult[21]>
#> 15: <ResampleResult[21]>
#> 16: <ResampleResult[21]>
#> 17: <ResampleResult[21]>
#> 18: <ResampleResult[21]>
#> 19: <ResampleResult[21]>
#> 20: <ResampleResult[21]>
#> 21: <ResampleResult[21]>
#> 22: <ResampleResult[21]>
#> 23: <ResampleResult[21]>
#> 24: <ResampleResult[21]>
#> 25: <ResampleResult[21]>
#> 26: <ResampleResult[21]>
#> 27: <ResampleResult[21]>
#> 28: <ResampleResult[21]>
#> 29: <ResampleResult[21]>
#> 30: <ResampleResult[21]>
#> 31: <ResampleResult[21]>
#> 32: <ResampleResult[21]>
#> 33: <ResampleResult[21]>
#> 34: <ResampleResult[21]>
#> 35: <ResampleResult[21]>
#> 36: <ResampleResult[21]>
#> 37: <ResampleResult[21]>
#> 38: <ResampleResult[21]>
#> 39: <ResampleResult[21]>
#> 40: <ResampleResult[21]>
#> 41: <ResampleResult[21]>
#> 42: <ResampleResult[21]>
#> 43: <ResampleResult[21]>
#> 44: <ResampleResult[21]>
#> 45: <ResampleResult[21]>
#> 46: <ResampleResult[21]>
#> 47: <ResampleResult[21]>
#> 48: <ResampleResult[21]>
#> 49: <ResampleResult[21]>
#> 50: <ResampleResult[21]>
#> 51: <ResampleResult[21]>
#> 52: <ResampleResult[21]>
#> 53: <ResampleResult[21]>
#> 54: <ResampleResult[21]>
#> 55: <ResampleResult[21]>
#> 56: <ResampleResult[21]>
#> 57: <ResampleResult[21]>
#> 58: <ResampleResult[21]>
#> 59: <ResampleResult[21]>
#> 60: <ResampleResult[21]>
#> 61: <ResampleResult[21]>
#> 62: <ResampleResult[21]>
#> 63: <ResampleResult[21]>
#> 64: <ResampleResult[21]>
#> 65: <ResampleResult[21]>
#> 66: <ResampleResult[21]>
#> 67: <ResampleResult[21]>
#> 68: <ResampleResult[21]>
#> 69: <ResampleResult[21]>
#>          resample_result

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
# }